Generate samples of correlated quantitative data

generate_correlated() is a function to generate samples of correlated quantitative data with specified first four moments and specified slope.

generate_correlated(
  n = 100,
  mean = 0,
  var = 1,
  skew = 0,
  kurt = 0,
  slope = 1,
  r = 0.5,
  summary = TRUE,
  decimals = 3,
  replication = 1
)

Arguments

n

A numeric vector for the sample size. The default value is 100.

mean

A numeric vector for the population mean of the first variable. The default value is 0.

var

A numeric vector for the population variance of the first variable. The default value is 1.

skew

A numeric vector for the population skew of the variables. Separate values of skew can be entered. The default value is 0.

kurt

A numeric vector for the population kurtosis of the variables. Separate values of kurtosis can be entered. The default value is 0.

slope

A numeric vector for the slope of the regression line. The default value is 0.

r

A numeric vector for the correlation between the variables. The default value is 0.5.

summary

A logical vector for whether the return should include summary statistics. The default value is TRUE.

decimals

A numeric vector for the number of decimals to round the sample data to. The default value is 3.

replication

A numeric vector for the number of times to replicate the sampling. The default value is 1.

Value

If summary = TRUE, a list containing the correlation of the sample, the slope of the sample, a data frame ("sample") with n rows and 2 columns, and a matrix ("summary") of the input parameters and descriptive statistics. If summary=FALSE, a data frame with n rows and 2 columns.

Examples

generate_correlated()
#> $correlation
#> [1] 0.518
#> 
#> $slope
#> [1] 0.885
#> 
#> $summary
#>          mean   sd  skew kurtosis
#> input 1  0.00 1.00  0.00     0.00
#> k 1     -0.07 1.04 -0.09     0.10
#> input 2  0.00 2.00  0.00     0.00
#> k 2     -0.20 1.77  0.00     0.46
#> 
#> $sample
#>         X1     X2
#> 1   -2.364 -0.582
#> 2   -0.560 -0.328
#> 3   -0.987 -1.255
#> 4    0.482 -1.430
#> 5    0.369  3.855
#> 6   -0.878 -1.180
#> 7    0.544  1.175
#> 8   -2.490 -0.074
#> 9    1.474  1.138
#> 10  -2.024 -0.952
#> 11   1.168  1.463
#> 12   1.289 -0.160
#> 13  -0.734 -1.477
#> 14  -0.981 -1.782
#> 15   0.095 -0.329
#> 16   0.830  1.100
#> 17  -0.501 -0.682
#> 18   1.087  1.083
#> 19  -0.231 -0.115
#> 20  -1.371 -1.105
#> 21  -1.674  0.001
#> 22   0.482 -0.007
#> 23   1.671 -0.234
#> 24  -0.674 -2.582
#> 25  -0.456 -2.987
#> 26  -1.176 -2.611
#> 27  -0.947  0.221
#> 28   0.149  2.599
#> 29   0.401  1.901
#> 30  -0.465  0.388
#> 31   0.409  1.264
#> 32   0.164 -3.019
#> 33   0.240  2.573
#> 34  -0.063  0.179
#> 35   0.163  0.437
#> 36  -1.096 -0.312
#> 37  -0.827 -2.534
#> 38   0.390  0.591
#> 39  -0.594 -1.365
#> 40  -0.575  0.638
#> 41   0.730 -0.291
#> 42  -0.624  0.296
#> 43   0.404 -0.486
#> 44   0.676  1.093
#> 45  -0.444 -1.187
#> 46  -1.769 -0.386
#> 47   0.541  0.724
#> 48   0.265  1.992
#> 49  -0.855 -2.479
#> 50   1.613  0.371
#> 51   0.565  0.433
#> 52  -0.092  2.934
#> 53  -0.131 -0.174
#> 54  -0.526 -2.125
#> 55  -0.039 -1.182
#> 56   0.543 -0.601
#> 57   0.675  0.032
#> 58   2.141  4.460
#> 59  -2.745 -1.993
#> 60  -0.145  1.766
#> 61   0.099 -1.449
#> 62  -0.244 -1.728
#> 63  -0.099  0.035
#> 64   0.803 -0.210
#> 65  -1.011 -0.135
#> 66   1.112  1.738
#> 67  -0.455  0.137
#> 68  -0.938 -3.776
#> 69  -1.598 -2.067
#> 70  -2.071 -3.100
#> 71  -0.849 -0.866
#> 72  -0.404  0.383
#> 73   1.078  1.301
#> 74   2.193  3.272
#> 75   2.358  3.482
#> 76  -2.099  0.521
#> 77   0.187 -0.226
#> 78   0.031 -2.740
#> 79   0.008 -2.737
#> 80  -0.094 -2.318
#> 81   0.884  0.550
#> 82   0.326 -1.482
#> 83  -0.606 -3.973
#> 84  -0.308 -1.373
#> 85  -0.643 -5.672
#> 86   0.399  0.560
#> 87   1.155 -0.423
#> 88  -0.209  0.540
#> 89   1.565  0.013
#> 90  -0.511 -0.236
#> 91   0.251  1.570
#> 92   0.401 -0.858
#> 93  -0.269 -1.237
#> 94  -0.948 -0.278
#> 95  -0.635 -0.883
#> 96  -0.655  1.729
#> 97   1.125  1.806
#> 98   0.384  0.744
#> 99   2.224  3.580
#> 100  0.900 -1.372
#> 

generate_correlated(skew = c(0.4, 0.25))
#> $correlation
#> [1] 0.463
#> 
#> $slope
#> [1] 0.899
#> 
#> $summary
#>          mean   sd skew kurtosis
#> input 1  0.00 1.00 0.40     0.00
#> k 1     -0.06 0.95 0.07     0.38
#> input 2  0.00 2.00 0.25     0.00
#> k 2     -0.23 1.86 0.11    -0.62
#> 
#> $sample
#>         X1     X2
#> 1    0.845  1.086
#> 2    0.762  0.847
#> 3    0.456  0.761
#> 4   -1.100 -2.017
#> 5    0.763 -2.242
#> 6    0.777 -0.254
#> 7   -1.064  0.282
#> 8   -0.536 -1.181
#> 9    2.894  3.535
#> 10   0.059 -0.892
#> 11  -0.830 -0.024
#> 12   0.372  0.797
#> 13  -1.721 -2.704
#> 14  -1.913 -2.446
#> 15   0.472  0.605
#> 16   0.876  0.761
#> 17  -0.118 -2.626
#> 18   0.897  1.201
#> 19   1.335  0.558
#> 20  -1.267 -1.956
#> 21   0.098  0.741
#> 22   0.607  0.391
#> 23   0.392 -1.942
#> 24   0.609 -2.099
#> 25  -0.098 -1.008
#> 26  -0.663 -2.568
#> 27  -0.017 -3.738
#> 28   1.939  3.877
#> 29  -0.699 -1.796
#> 30   0.479  0.378
#> 31  -0.581  0.761
#> 32  -0.644 -1.183
#> 33  -0.251 -4.263
#> 34  -0.328 -0.893
#> 35  -1.743 -2.113
#> 36  -0.657  0.960
#> 37  -1.449  0.232
#> 38  -0.786 -1.634
#> 39   0.599 -1.075
#> 40   1.276  1.723
#> 41   0.025  2.289
#> 42   0.768  0.140
#> 43  -1.260 -1.061
#> 44  -0.316 -1.384
#> 45  -0.489 -1.602
#> 46  -0.968  1.199
#> 47  -0.831  2.178
#> 48  -0.495 -1.629
#> 49   0.150 -2.240
#> 50  -0.179 -1.833
#> 51   0.790  2.491
#> 52   0.610 -0.725
#> 53   0.271  0.058
#> 54   0.380  2.846
#> 55  -0.426 -3.522
#> 56  -0.274  1.614
#> 57  -0.459 -1.160
#> 58   0.329  0.128
#> 59  -0.594  1.790
#> 60  -0.199 -2.949
#> 61   0.259 -0.559
#> 62   1.112  0.772
#> 63  -0.805  0.070
#> 64   0.737  1.678
#> 65  -0.971 -0.198
#> 66  -0.750 -3.469
#> 67  -0.955 -2.748
#> 68  -0.762 -3.050
#> 69  -0.706 -1.232
#> 70  -0.741 -0.312
#> 71   1.030  2.820
#> 72   1.397  3.321
#> 73   1.380  1.295
#> 74  -0.790  0.876
#> 75  -0.819  0.841
#> 76   0.453  1.675
#> 77  -0.371 -0.883
#> 78   0.661  2.914
#> 79   1.975  0.510
#> 80   0.383  1.370
#> 81  -0.738  1.193
#> 82   0.670 -0.442
#> 83   0.499 -0.342
#> 84  -0.606 -3.389
#> 85   1.391 -0.679
#> 86  -0.155  0.064
#> 87  -0.648  1.415
#> 88  -0.019 -2.315
#> 89  -1.278  0.890
#> 90  -0.237  0.257
#> 91  -1.633  0.485
#> 92   1.398  1.959
#> 93   0.200  4.065
#> 94   1.184  2.804
#> 95   0.318 -1.099
#> 96  -0.070 -2.094
#> 97   0.481 -1.800
#> 98  -1.761 -0.724
#> 99  -2.976 -1.356
#> 100 -1.064 -2.611
#> 

generate_correlated(slope = 3, r = 0.8)
#> $correlation
#> [1] 0.799
#> 
#> $slope
#> [1] 3.103
#> 
#> $summary
#>         mean   sd  skew kurtosis
#> input 1 0.00 1.00  0.00     0.00
#> k 1     0.18 0.99 -0.24     0.44
#> input 2 0.00 3.75  0.00     0.00
#> k 2     0.67 3.83 -0.38    -0.23
#> 
#> $sample
#>         X1      X2
#> 1   -0.134   1.909
#> 2   -0.156  -0.999
#> 3   -0.572   1.378
#> 4   -0.183  -1.166
#> 5   -1.838  -4.479
#> 6   -0.643  -1.461
#> 7    0.439  -1.236
#> 8    0.370  -0.371
#> 9    0.444  -3.564
#> 10  -0.440  -1.943
#> 11   0.088   3.705
#> 12  -0.419  -5.079
#> 13   0.197   4.205
#> 14   1.459   3.586
#> 15  -0.576  -2.877
#> 16   0.277   1.810
#> 17   0.413   0.180
#> 18   0.423   0.915
#> 19  -1.095  -4.446
#> 20   0.157   2.454
#> 21  -0.976  -4.340
#> 22  -0.838   1.791
#> 23   0.932   6.284
#> 24   0.979   5.561
#> 25   2.475   6.311
#> 26  -0.011   0.691
#> 27  -0.469  -3.013
#> 28   0.043   3.119
#> 29  -1.722  -6.381
#> 30  -1.849  -7.947
#> 31  -0.177  -2.251
#> 32   0.990   0.238
#> 33   0.319   4.662
#> 34  -0.255  -1.745
#> 35   1.423   5.233
#> 36  -0.017   2.788
#> 37   0.525  -1.069
#> 38   2.152   6.598
#> 39   1.158   2.616
#> 40   0.177  -3.619
#> 41   0.130  -0.065
#> 42   0.172  -0.106
#> 43   0.103  -0.170
#> 44   0.514   3.012
#> 45   1.198   3.348
#> 46   0.697   2.753
#> 47   1.983   5.502
#> 48   1.074  -0.982
#> 49   2.070   7.166
#> 50  -2.794  -9.027
#> 51  -1.118  -6.709
#> 52   0.122  -0.689
#> 53  -0.392  -3.657
#> 54   1.716   2.662
#> 55   0.197   0.365
#> 56   0.077   3.990
#> 57  -1.529  -3.169
#> 58   0.117  -2.155
#> 59   1.967   5.008
#> 60  -0.160   6.911
#> 61   0.453  -0.709
#> 62   0.506   0.309
#> 63   0.214   0.359
#> 64   0.917   5.318
#> 65   1.875   6.900
#> 66  -0.308   1.805
#> 67  -0.055  -0.013
#> 68  -0.227  -2.294
#> 69  -1.539  -4.092
#> 70   0.370   1.433
#> 71  -0.843  -1.524
#> 72  -0.228   3.784
#> 73   0.487   1.696
#> 74   0.549   2.625
#> 75   0.824   3.036
#> 76   0.204  -1.099
#> 77  -0.594   0.946
#> 78   0.782   1.245
#> 79   0.260   4.127
#> 80  -2.394 -10.392
#> 81   0.680   7.096
#> 82   0.377   3.021
#> 83   0.833   3.342
#> 84   0.603   3.031
#> 85  -0.357  -1.176
#> 86  -0.391  -0.938
#> 87   0.292   5.949
#> 88   0.091  -1.805
#> 89   0.239  -0.355
#> 90   1.268   5.413
#> 91  -0.970  -4.130
#> 92  -1.149  -3.246
#> 93   0.376   1.573
#> 94   1.218   0.903
#> 95   1.611   7.227
#> 96   0.767   5.052
#> 97  -0.573  -5.352
#> 98   1.789   3.928
#> 99   1.885   4.132
#> 100 -0.720  -2.646
#> 

generate_correlated(n = 10000, skew = c(0, 1), kurt = c(0, 1.5))
#> $correlation
#> [1] 0.492
#> 
#> $slope
#> [1] 0.986
#> 
#> $summary
#>          mean   sd  skew kurtosis
#> input 1  0.00 1.00  0.00     0.00
#> k 1     -0.01 0.99  0.03     1.00
#> input 2  0.00 2.00  1.00     1.50
#> k 2      0.01 1.99 -0.01     0.92
#> 
#> $sample
#>         X1     X2
#> 1   -0.398 -2.736
#> 2    0.040  0.969
#> 3   -1.411 -0.484
#> 4   -0.054  1.843
#> 5   -0.553  2.738
#> 6   -0.032  0.357
#> 7   -0.125 -0.458
#> 8   -1.257 -0.601
#> 9    0.535  0.072
#> 10   0.470  2.205
#> 11   0.850  0.212
#> 12  -0.438 -0.497
#> 13  -0.673 -3.944
#> 14   0.908  0.084
#> 15   2.163  0.221
#> 16   0.410 -0.362
#> 17   2.362  0.499
#> 18  -0.408  2.186
#> 19   0.164  0.586
#> 20  -1.998 -1.092
#> 21  -0.540 -1.260
#> 22   0.259  0.813
#> 23  -0.522 -2.628
#> 24  -0.103 -0.554
#> 25   0.036  1.273
#> 26   0.527 -1.107
#> 27  -0.893  0.102
#> 28   0.226  2.310
#> 29   1.039  0.236
#> 30  -0.113 -0.231
#> 31  -0.313 -1.607
#> 32  -1.447 -1.179
#> 33  -0.683 -1.781
#> 34  -0.497  0.903
#> 35  -0.133 -0.503
#> 36  -0.281  2.196
#> 37   0.960  0.865
#> 38  -0.355 -2.395
#> 39  -2.391 -0.712
#> 40   0.705  1.004
#> 41  -0.076  3.920
#> 42   0.203  3.152
#> 43  -0.662  1.092
#> 44   0.435  2.051
#> 45  -0.352 -0.378
#> 46   0.056 -1.361
#> 47  -0.851  0.237
#> 48  -1.271 -3.208
#> 49  -0.787 -1.255
#> 50  -0.057 -0.864
#> 51   0.747 -0.386
#> 52  -0.067  0.269
#> 53   0.589  2.366
#> 54  -0.750 -0.082
#> 55   0.128  1.129
#> 56   1.169  2.242
#> 57   0.537 -1.336
#> 58  -0.655  0.134
#> 59   1.088 -0.434
#> 60   0.016 -2.702
#> 61  -0.062  0.826
#> 62  -0.625 -1.027
#> 63   0.415  0.871
#> 64   0.754  1.897
#> 65  -0.685 -1.446
#> 66   0.484  1.058
#> 67  -0.445 -0.317
#> 68  -1.026 -1.485
#> 69  -0.116  0.910
#> 70   0.047 -0.628
#> 71  -1.103 -3.468
#> 72   3.170  1.014
#> 73   0.000 -2.745
#> 74  -0.356  1.999
#> 75  -0.141 -1.035
#> 76  -0.579 -0.917
#> 77   0.006  1.253
#> 78  -1.494 -0.851
#> 79   1.386  1.501
#> 80   1.117  2.138
#> 81  -0.514  0.293
#> 82  -0.378 -0.443
#> 83   1.583  0.526
#> 84   0.100  1.862
#> 85  -0.826  1.764
#> 86   2.440  4.123
#> 87   2.605  4.890
#> 88   0.814  2.164
#> 89  -1.149 -1.647
#> 90   1.090 -0.802
#> 91   1.547  0.109
#> 92  -0.845  0.221
#> 93   1.530  2.235
#> 94   2.062 -1.547
#> 95   0.093  1.168
#> 96  -1.242  0.032
#> 97   0.358  2.280
#> 98   0.371 -0.303
#> 99  -0.238 -1.863
#> 100  1.680  1.303
#> 101  0.299  2.280
#> 102  1.328  0.619
#> 103  0.709  0.167
#> 104 -0.834 -1.307
#> 105  0.563  2.677
#> 106 -0.302 -0.866
#> 107 -0.055 -0.619
#> 108 -0.576  0.183
#> 109 -0.177  0.564
#> 110  0.574 -2.021
#> 111 -0.222 -4.657
#> 112 -0.557  0.987
#> 113  0.632 -0.019
#> 114  0.574 -0.536
#> 115 -0.277 -6.055
#> 116  0.293  1.281
#> 117 -0.463 -2.146
#> 118  0.949  1.027
#> 119 -2.686 -0.855
#> 120  1.163  1.021
#> 121 -1.021  1.638
#> 122  2.020  2.698
#> 123 -0.594 -2.382
#> 124 -0.708  0.546
#> 125 -1.391 -0.190
#> 126  0.962 -1.432
#> 127 -0.918 -0.440
#> 128 -0.055  1.712
#> 129 -0.097 -0.342
#> 130  0.082 -0.084
#> 131 -0.292  1.741
#> 132 -0.167  0.254
#> 133  1.320  1.025
#> 134  0.535  0.013
#> 135  0.577 -0.198
#> 136 -1.021 -0.514
#> 137  1.645  1.358
#> 138  0.459 -1.544
#> 139  0.675 -0.510
#> 140 -1.092 -1.169
#> 141 -2.565 -3.686
#> 142 -0.116 -1.390
#> 143  0.626 -1.497
#> 144  0.827  1.471
#> 145  0.963 -1.173
#> 146 -0.736 -0.011
#> 147  0.798  3.520
#> 148 -1.374  0.674
#> 149  0.517  5.175
#> 150 -1.010 -1.044
#> 151  1.303  1.294
#> 152  1.087  1.162
#> 153 -0.786 -0.258
#> 154  0.082  1.328
#> 155  0.276  0.749
#> 156  1.957  1.227
#> 157  1.566 -0.060
#> 158 -0.518 -0.502
#> 159  0.793  2.511
#> 160  1.048 -0.201
#> 161 -1.841 -0.823
#> 162 -0.166  1.289
#> 163  1.062  1.600
#> 164  1.516  0.086
#> 165 -0.630 -1.042
#> 166  0.055 -1.337
#> 167  0.163  1.042
#> 168 -0.316  1.202
#> 169 -1.767 -0.966
#> 170  1.197  1.295
#> 171  0.375 -0.091
#> 172  0.436  0.500
#> 173 -0.222 -1.096
#> 174  0.360  4.198
#> 175 -1.085 -1.955
#> 176  0.789  0.317
#> 177  0.391  2.353
#> 178  0.094 -1.190
#> 179 -0.009 -0.222
#> 180 -0.435 -5.657
#> 181  0.078 -3.515
#> 182  0.984  0.523
#> 183  0.325  0.846
#> 184  0.205  2.957
#> 185  0.112  2.462
#> 186  0.879 -0.435
#> 187 -0.218 -1.523
#> 188  0.291  0.987
#> 189 -0.072 -1.461
#> 190 -0.630  0.726
#> 191 -0.657 -0.928
#> 192 -1.119  0.248
#> 193  1.678  2.235
#> 194  0.071  2.627
#> 195 -0.192  1.422
#> 196  0.487 -2.428
#> 197  0.536  1.613
#> 198  1.409  1.033
#> 199  0.432 -1.010
#> 200  0.951  0.497
#> 201  1.081  1.471
#> 202 -0.210 -0.337
#> 203 -0.901  0.765
#> 204 -0.804  0.687
#> 205 -0.826 -0.744
#> 206 -1.174  1.973
#> 207  0.013 -0.232
#> 208  1.048  0.150
#> 209  0.483 -0.593
#> 210  2.027  0.564
#> 211 -0.510 -2.509
#> 212 -0.719 -1.048
#> 213 -1.192 -2.551
#> 214 -0.289 -0.218
#> 215 -0.296  0.718
#> 216 -0.484  0.940
#> 217 -3.099 -5.668
#> 218 -0.333  0.727
#> 219 -0.031 -0.211
#> 220  0.920 -0.011
#> 221 -0.078 -1.073
#> 222  0.898  0.148
#> 223 -0.620 -0.439
#> 224 -0.292  0.107
#> 225 -2.015 -2.795
#> 226 -2.161 -1.123
#> 227 -0.075 -3.880
#> 228  0.722  4.154
#> 229  0.115 -2.345
#> 230  2.249  1.116
#> 231 -1.314 -2.091
#> 232 -0.112 -2.296
#> 233 -0.325 -0.780
#> 234  0.312  0.506
#> 235  0.830  1.260
#> 236 -0.825 -0.107
#> 237  1.604  2.189
#> 238  0.395 -0.143
#> 239 -0.748 -0.049
#> 240 -0.450 -1.539
#> 241  0.544 -0.110
#> 242 -0.843  1.360
#> 243 -0.236  1.411
#> 244 -1.579  1.278
#> 245 -0.350  0.017
#> 246  0.689  1.048
#> 247 -1.755 -1.761
#> 248  0.816 -1.310
#> 249 -0.369 -0.297
#> 250  0.132 -0.632
#> 251  0.077  1.019
#> 252 -0.987  0.124
#> 253  0.667 -2.030
#> 254 -0.057 -0.644
#> 255  0.011  0.181
#> 256  0.188  0.627
#> 257  0.514  0.547
#> 258  0.454 -0.792
#> 259 -1.164 -3.075
#> 260 -0.226 -3.061
#> 261  1.797  3.830
#> 262 -0.040  0.577
#> 263  1.110  1.290
#> 264 -0.269  1.968
#> 265 -0.092  1.922
#> 266  0.672  0.546
#> 267  0.928  0.801
#> 268  0.293  0.398
#> 269  0.545 -0.935
#> 270  0.422  1.648
#> 271  0.092 -1.006
#> 272 -0.219 -1.997
#> 273  0.194 -0.383
#> 274 -0.931 -1.823
#> 275  0.010  1.029
#> 276 -0.650 -0.306
#> 277  1.120 -0.010
#> 278  0.319 -0.818
#> 279 -1.966 -0.341
#> 280 -0.009 -0.316
#> 281  0.069 -2.492
#> 282 -1.823 -1.881
#> 283  0.144  1.146
#> 284  1.715  0.956
#> 285 -1.114 -3.220
#> 286 -0.324 -3.687
#> 287  1.948  1.570
#> 288  0.840  2.267
#> 289  0.075 -0.749
#> 290  0.317 -2.952
#> 291  0.364  0.396
#> 292  1.839  0.928
#> 293  0.650 -1.426
#> 294  0.958 -0.194
#> 295  0.373  5.952
#> 296 -0.521  0.097
#> 297  0.578 -2.859
#> 298  0.257  0.657
#> 299 -0.984 -2.130
#> 300  0.287  1.431
#> 301 -0.944  4.090
#> 302  0.740  3.335
#> 303  1.315  1.225
#> 304  0.162  0.339
#> 305 -1.319  1.484
#> 306  1.121  2.203
#> 307 -0.298 -1.466
#> 308  1.267  4.041
#> 309  2.415  1.700
#> 310 -1.352 -1.677
#> 311  0.132 -1.203
#> 312 -0.279 -2.760
#> 313 -0.773 -0.279
#> 314 -0.184 -0.439
#> 315  1.085  2.793
#> 316 -0.041  1.162
#> 317  0.059 -0.174
#> 318 -0.747 -1.447
#> 319 -1.137 -2.238
#> 320  1.037  2.630
#> 321 -0.273 -0.647
#> 322  0.165 -2.901
#> 323 -0.530 -0.991
#> 324  2.449  0.598
#> 325  1.407  0.033
#> 326  1.170  1.765
#> 327 -0.448  1.905
#> 328 -0.439 -0.274
#> 329 -0.446  0.482
#> 330 -0.335  0.041
#> 331  2.350  1.587
#> 332 -0.722  1.173
#> 333  0.353 -0.348
#> 334  0.208  0.688
#> 335  0.485  0.116
#> 336  0.062  0.338
#> 337  0.701  2.188
#> 338  0.110 -0.178
#> 339  0.002  0.055
#> 340 -0.523 -1.387
#> 341  0.343 -0.548
#> 342 -0.266 -0.288
#> 343  1.919  0.435
#> 344  0.377 -0.659
#> 345  0.401 -0.282
#> 346  0.373 -0.520
#> 347  0.443  0.704
#> 348 -0.858  1.146
#> 349 -0.268 -0.103
#> 350  0.298  0.085
#> 351 -0.755 -1.259
#> 352  1.157  2.673
#> 353  0.107 -3.065
#> 354  0.273  4.159
#> 355  0.743  0.922
#> 356  0.802 -1.196
#> 357  1.308  0.262
#> 358 -2.093 -2.161
#> 359 -0.283 -2.482
#> 360 -0.674 -0.055
#> 361 -2.257 -2.148
#> 362 -0.734  2.429
#> 363 -1.074  1.365
#> 364 -2.599 -3.820
#> 365  0.269 -0.529
#> 366 -0.024  2.203
#> 367 -0.169  1.005
#> 368 -0.105 -1.561
#> 369  0.243 -1.787
#> 370  1.159  2.247
#> 371  0.228  0.217
#> 372 -0.537 -3.909
#> 373  0.094 -1.113
#> 374 -0.258 -0.847
#> 375  0.567  1.328
#> 376 -0.605  1.043
#> 377  0.683  1.481
#> 378  1.488  1.962
#> 379 -0.342 -0.284
#> 380  0.400  1.059
#> 381 -1.308 -2.446
#> 382  1.111 -0.257
#> 383 -0.221 -1.000
#> 384 -0.423 -0.309
#> 385 -0.165 -0.989
#> 386  2.898  2.545
#> 387  0.569  4.062
#> 388 -1.460  1.512
#> 389  0.230  2.302
#> 390 -0.319  0.031
#> 391  1.452  4.417
#> 392 -1.872 -2.697
#> 393  0.358 -0.221
#> 394 -0.477 -1.274
#> 395 -1.312 -0.947
#> 396  1.029  1.674
#> 397  0.511  1.972
#> 398  0.230  3.719
#> 399 -0.094  2.256
#> 400  1.574 -1.089
#> 401 -0.390 -1.667
#> 402 -0.380  0.301
#> 403 -0.431  0.810
#> 404 -0.152 -0.801
#> 405  0.067 -1.338
#> 406  0.595  2.982
#> 407  0.594  1.222
#> 408 -0.235  1.820
#> 409  2.937  0.711
#> 410 -2.472 -3.319
#> 411 -0.591  0.616
#> 412  0.050  1.999
#> 413 -0.295 -1.812
#> 414 -1.299 -2.599
#> 415  0.685  2.471
#> 416  1.274  2.386
#> 417 -1.188 -1.299
#> 418  0.828  0.196
#> 419  1.772  1.480
#> 420 -0.370 -0.512
#> 421 -0.532  1.999
#> 422 -0.991 -1.857
#> 423 -0.466  2.684
#> 424  0.720  4.522
#> 425 -0.067  0.033
#> 426 -0.244 -0.757
#> 427  0.222  1.774
#> 428 -0.681 -0.326
#> 429  0.269 -0.325
#> 430  1.074 -1.918
#> 431  0.922  0.719
#> 432  1.545  6.778
#> 433  0.495 -0.703
#> 434  0.152 -1.977
#> 435  0.634  0.995
#> 436  0.659  2.601
#> 437  2.155  2.391
#> 438 -0.325  0.291
#> 439 -0.347 -1.555
#> 440 -0.190 -1.084
#> 441  0.422 -1.341
#> 442  0.025  0.318
#> 443 -0.858 -1.219
#> 444 -1.912 -0.568
#> 445 -0.077  0.026
#> 446  1.917  0.239
#> 447  0.488  3.234
#> 448 -0.669 -0.272
#> 449  0.271  1.468
#> 450 -0.253 -2.456
#> 451 -0.236 -2.700
#> 452 -1.203  2.295
#> 453 -0.872  3.653
#> 454 -0.365 -0.199
#> 455  0.102  0.110
#> 456  0.845  1.502
#> 457  0.082  0.760
#> 458  0.690  0.931
#> 459 -0.383  2.297
#> 460 -0.662 -1.765
#> 461  0.387 -1.271
#> 462 -2.155 -1.475
#> 463  2.912  5.532
#> 464  0.692  1.497
#> 465  0.477  0.496
#> 466 -0.321  1.462
#> 467  0.073 -2.783
#> 468 -1.619 -3.019
#> 469 -0.255  1.045
#> 470 -0.485 -0.530
#> 471  0.068  1.015
#> 472  0.188 -0.729
#> 473 -1.281  2.268
#> 474  1.697  4.256
#> 475  0.724  2.841
#> 476  1.427  1.039
#> 477 -0.785 -0.748
#> 478 -0.828  0.281
#> 479 -0.762  0.735
#> 480  0.284  0.458
#> 481  0.108 -0.312
#> 482  0.396  0.050
#> 483 -0.661  0.634
#> 484 -0.247  1.056
#> 485 -0.188 -0.867
#> 486  0.456 -1.192
#> 487  0.470  0.879
#> 488  1.244  0.302
#> 489 -1.003  0.581
#> 490  0.170 -0.303
#> 491 -0.840  1.869
#> 492  0.977 -0.236
#> 493  0.021  1.522
#> 494 -1.853 -5.574
#> 495 -0.650 -1.066
#> 496 -1.651  1.355
#> 497 -1.055  0.030
#> 498 -0.363 -0.104
#> 499 -0.089 -0.598
#> 500 -0.049  0.482
#>  [ reached 'max' / getOption("max.print") -- omitted 9500 rows ]
#>