Generate sample data from named distributions

generate_distribution() is a wrapper function for generating sample data from named continuous distributions.

The distributional shapes include beta, Cauchy, chi-squared, exponential, F, gamma, Laplace, logistic, log normal, normal, t, triangular, uniform, and Weibull. See additional details below for distribution parameters and default values.

generate_distribution(
  distribution = c("beta", "cauchy", "chi-squared", "exponential", "f", "gamma",
    "laplace", "logistic", "log normal", "normal", "t", "triangular", "uniform",
    "weibull"),
  n = 1000,
  seed = NULL,
  alpha_shape = NULL,
  beta_shape = NULL,
  location = NULL,
  scale = NULL,
  df = NULL,
  df2 = NULL,
  rate = NULL,
  mean_log = NULL,
  sd_log = NULL,
  mean = NULL,
  sd = NULL,
  theta = NULL,
  min = NULL,
  max = NULL
)

Arguments

distribution

A character vector to identify which distributional shape to sample. Options include beta, cauchy, chi-squared, exponential, f, gamma, laplace, logistic, log normal, normal, t, triangular, uniform, and weibull.

n

A numeric vector for the sample size. The default value is 1000.

seed

An optional numeric vector to use in set.seed()

alpha_shape

A non-negative numeric vector required for the shape of the beta, gamma, and Weibull distributions.

beta_shape

A non-negative numeric vector required for the shape of the beta distribution.

location

A numeric vector for the location of the Cauchy, Laplace, and logistic distributions.

scale

A numeric vector for the scale of the Cauchy, Laplace, logistic, and Weibull distributions.

df

A numeric vector required for the degrees of freedom of the chi-squared, F, and t distributions.

df2

A numeric vector required for the degrees of freedom of the F distribution.

rate

A numeric vector for the rate of the exponential and gamma distributions.

mean_log

A numeric vector for the mean (on the log scale) of the log normal distribution.

sd_log

A numeric vector for the standard deviation (on the log scale) of the log normal distribution.

mean

A numeric vector for the mean of the normal distribution.

sd

A numeric vector for the standard deviation of the normal distribution.

theta

A numeric vector required for the theta of the triangular distribution.

min

A numeric vector for the minimum value of the triangular and uniform distributions.

max

A numeric vector for the maximum value of the triangular and uniform distributions.

Value

A list of

data, a vector of length n containing the generated values

summary, a descriptive matrix of input and sample values of the specified distribution's parameters

Distributional Shapes & Default Values

The beta distribution requires the alpha and beta shape parameters. There are no default values for this distribution. See stats::rbeta() for more information.

The Cauchy distribution uses the location and scale parameters. The default values are location = 0 and scale = 1. See stats::rcauchy() for more information.

The chi-squared distribution requires the df parameter. There is no default value for this distribution. See stats::rchisq() for more information.

The exponential distribution uses the rate parameter, which has a default value of 1. See stats::rexp() for more information.

The F distribution requires the df and df2 parameters. There are no default values for this distribution. See stats::rf() for more information.

The gamma distribution requires the alpha shape and rate parameter. There is no default value for alpha shape, but the default value for rate is 1. See stats::rgamma() for more information.

The Laplace distribution uses the location and scale parameters. The default values are location = 0 and scale = 1. See VGAM::rlaplace() for more information.

The logistic distribution uses the location and scale parameters. The default values are location = 0 and scale = 1. See stats::rlogis() for more information.

The log normal distribution uses the mean log and sd log parameters. The default values are mean log = 0 and sd log = 1. See stats::rlnorm() for more information.

The normal distribution uses the mean and sd parameters. The default values are mean = 0 and sd = 1. See stats::rnorm() for more information.

The t distribution requires the df parameter. There is no default value for this distribution. See stats::rt() for more information.

The triangular distribution requires the theta, min, and max parameters. There is no default value for theta, but the other default values are min = 0 and max = 1. See VGAM::rtriangle() for more information.

The uniform distribution uses the min and max parameters. The default values are min = 0 and max = 1. See stats::runif() for more information.

The Weibull distribution requires the alpha shape and scale parameter. There is no default value for alpha shape, but the default value for scale is 1. See stats::rweibull() for more information.

Examples

generate_distribution(distribution = "laplace")
#> $data
#>    [1] -1.2124289454 -2.3462164414 -0.4034243219 -1.4756350158 -0.1913653025
#>    [6]  0.6526954796  1.6806995111  1.2614069686 -0.1089598200  0.0497891832
#>   [11] -0.3485403288  0.7527280873  1.5872993782 -0.2525679636  1.0248187590
#>   [16] -0.1190258327  0.7083276921  0.7163674715 -0.6344118689  0.3543774225
#>   [21]  1.1563115450 -0.3844753151  0.5700024844  1.3860717495  1.1075342583
#>   [26]  0.3404986809 -2.0494390823 -0.1130898348 -0.9776763448  1.1835949515
#>   [31]  0.1656665361  2.4342645779  0.2123301052  1.2835620414 -8.1473432232
#>   [36]  0.0673062195  0.7739134832  4.8012382421  0.9113856268  1.3053272048
#>   [41]  0.1579708542  0.6472391122  0.9627187361 -0.4761796066 -1.1795209892
#>   [46]  1.3415447943 -0.2319581446 -0.2438804792 -0.2125551497 -2.7928459681
#>   [51] -3.2321106041  1.4896794344 -0.1592239593  0.0029898551  0.4001366912
#>   [56] -0.1660989616  1.4628807919  0.8356471530 -0.0608689709  2.7900761700
#>   [61]  4.0877519617  1.3809561044 -1.2966099421  0.2928731211  0.4608224496
#>   [66]  0.2876559442 -0.2693955341  2.7991929592 -0.8018104345  0.3359010666
#>   [71] -1.5584755687 -0.4822587955 -0.1594247657  0.0246099577 -0.0931093130
#>   [76]  1.4495716608 -0.6836068948 -0.5612955713 -0.3088333478 -0.0155069272
#>   [81]  0.3401315456  3.1832926211  0.4077208599  0.3751022976  0.0461001300
#>   [86] -6.3259369107  0.1866239536  0.9179465141  1.8209024212  0.8298397272
#>   [91] -1.1813625920  0.1638684495  0.0678196223 -0.0929269237 -1.4455328841
#>   [96]  0.4013459891  0.6628835177  0.0553208341 -1.8242814204 -0.3224050056
#>  [101]  0.6201467752 -1.0626257143 -0.3307350778  1.6540773568  0.2658291969
#>  [106]  3.9098002211  0.4668795550 -0.7732511375 -0.0336870446  0.4708500418
#>  [111]  0.0341499536  1.0556762479  0.4911640162 -0.2896968136  0.3782401044
#>  [116]  0.0739377828 -0.6465320627  0.4165156791 -3.2198409884 -0.2779712360
#>  [121] -0.8574837264  0.6298182027 -0.5603681620 -0.3748135966 -0.6720790592
#>  [126] -0.2384752172 -1.1648498000 -0.4071176521 -0.1948637288 -1.2015463404
#>  [131] -1.0734989586 -0.6507868969 -2.6836286806 -0.7561488797 -1.4301708227
#>  [136] -0.0554299856 -0.9126867063  1.5028704241  0.3511950925  0.2134721839
#>  [141] -0.3446054470  0.2732047921 -4.0405291187  1.3811290956 -0.1674935145
#>  [146] -3.5245588250 -0.8991049266 -0.3394806466 -0.7467796223  0.9497494762
#>  [151] -0.0630964738  1.2473749286  1.2940126997 -1.6621949150  2.1420035049
#>  [156]  0.5751126664 -0.8505534288 -0.0661983524  2.0798965106 -0.5784325014
#>  [161] -0.3893951676  1.5609514434 -0.0387534978  1.3224667403  0.3091176299
#>  [166] -0.0487002176  2.5130994409 -0.2865377843 -3.3915883348 -1.2059257228
#>  [171]  3.7083293162 -0.3427076843 -2.3284022369 -0.4493223967  0.9921234668
#>  [176] -1.2361831337  0.9438122028  0.6632238634 -1.9769620888 -1.1605278679
#>  [181]  0.0815273966  0.6342219612  1.5002533466  0.0998631814 -0.4541358992
#>  [186]  0.5345637060  3.2206846787 -1.4513768212 -0.0916067736 -2.5303548226
#>  [191] -0.4219391893  0.8020082829  0.2918765715 -0.3885768735  0.0160476620
#>  [196]  0.0018968493  2.7094110383  0.0621615262 -0.0855559861  0.1977583163
#>  [201]  0.3130567272 -0.0003658732  0.9933383196  0.3614747883  0.1231470142
#>  [206]  0.1814954440  0.0176328582  1.3885494714  1.4976568648  0.0963608545
#>  [211] -0.2946210399 -0.7175644197  1.2660044492 -0.9172056161  0.8626963375
#>  [216] -0.3824947255 -0.8869771032 -1.3542901267  0.2553027769  1.4482322155
#>  [221]  2.2145595524  0.8093691584 -0.4650025177 -1.6451955708  0.1324358008
#>  [226] -0.5058746059 -1.3117452110  4.3440915222 -0.0984092376  0.3494105889
#>  [231] -0.0492736719  0.0150230560  1.9092990035 -0.0561298681  0.2164822759
#>  [236] -0.1097245222  2.2631110089  0.6118659138 -0.6212695357  0.1341953181
#>  [241] -3.8085597898 -1.8897133644 -0.6687362074  0.0041356677  0.4746771954
#>  [246]  0.1847724079 -1.1190918125  1.5667652614 -0.2268050533 -0.7567125176
#>  [251] -0.4436303097  0.5767845608 -0.0764358717  0.3509411510 -0.9685758229
#>  [256]  1.2727484608  0.7635999912  0.2417391549 -2.2498847226  0.9054189360
#>  [261] -0.2996201125  0.4230906065  2.8573862693 -0.1139252748 -5.9926624086
#>  [266]  0.1294505280 -1.6283745971  2.2240957773  0.1213040393  1.7845175011
#>  [271] -0.8037772219  0.7759767305 -0.3916473937 -1.4421042814 -0.5251633699
#>  [276]  0.1667105301  2.9433611850 -0.1248690302  1.3573696747  0.0285328484
#>  [281] -0.2878806495  0.7863913491  0.6049228287  0.6760998986 -0.5841816560
#>  [286]  0.0448366446 -1.9484815291 -0.7371735782  0.1487312908 -0.5817079777
#>  [291]  0.9851431348  2.1078552706 -0.5214125163 -0.0392959211 -2.3671102131
#>  [296] -0.2457398510  0.4787070582 -4.0485541283 -2.1201193304  0.1957239948
#>  [301]  3.8468065329  2.6028039781  1.0848940284  2.4299328608  1.8508146718
#>  [306] -0.5635398063 -1.4217010724 -2.7382308444  0.4134139524 -0.7064391586
#>  [311] -0.7011386813 -1.7655768195 -0.0418607684 -1.2997845798  2.3090445523
#>  [316]  0.1368008788  0.4190533624  0.3825575779  0.4393719227 -0.1046548151
#>  [321] -0.2100712827 -0.8886364533 -2.3686884266 -0.2125452654 -0.4233022006
#>  [326] -0.3942173054 -1.1303703392  0.8175660152 -0.3010604674 -0.3486956060
#>  [331]  1.1184964631  0.6888692371  0.2050478792  2.3252356463  0.5815620890
#>  [336] -0.8758416578  0.7270325812  0.5292759818  0.3434412176 -0.3998563528
#>  [341]  3.7858920162 -2.8596444406 -0.0503408285 -0.8005853184 -0.3730668388
#>  [346]  0.5452221705  0.1027950954  3.7521087821  0.2964479342  0.7954871651
#>  [351]  0.5744184868  0.5089673890 -0.8383003966 -1.7340926363 -0.3407772758
#>  [356] -0.0833733927  0.7515830976 -3.1550846417  0.7932670202 -0.4262547287
#>  [361] -1.3562016853  0.8831097326 -0.1705550697 -0.1943374471 -0.0329447888
#>  [366] -0.3476590140  4.6042033112 -1.0517116566 -1.1543074073  0.0029469037
#>  [371]  0.5146344460 -2.0815654266 -1.8503964561  0.5444550606 -0.1801176744
#>  [376]  3.2485221183  1.0232453394 -0.8029148010  2.2795892666 -1.5108429371
#>  [381]  1.5826110995  0.1311264539  0.7520743928  0.1343318504 -1.0869953493
#>  [386] -1.6169700429 -0.4443032002  2.4903568863 -3.1967831457 -1.1189110932
#>  [391]  1.0799443714  0.9383305451 -0.5837997796 -0.1089401602  2.0612893964
#>  [396]  0.8608216356 -1.5494058930 -0.7249142123 -0.7402527835  2.1731912473
#>  [401] -1.5590793856 -1.8942547391  0.6753303141 -1.2542636024 -0.1286311172
#>  [406]  0.9851367831 -1.5888840544  0.5148457738  1.5208085722  0.0252547627
#>  [411]  0.5447911560 -0.8699188658 -2.4901601747  0.4960938647  0.6694267029
#>  [416] -0.4031302687 -0.0615254334  0.8207760526 -0.5482927809  1.9732781367
#>  [421] -0.3094729006 -1.8436178084 -1.0276860296  2.3519985092 -0.2765545842
#>  [426] -1.0736602424 -1.2010450277  0.6971555626 -0.7635788272  0.7867399247
#>  [431]  1.6996554916  1.3864798214 -0.6631520203  0.0312432494 -0.0348306222
#>  [436] -0.8325312135 -0.9290893787  0.0711617214 -0.6435930740  0.1935536582
#>  [441]  2.0337699065  1.3317112738  0.9891104406  1.3783992388  1.7277481798
#>  [446]  0.5576353915  0.7896175212 -0.0173141435  0.0412582557  0.2460969219
#>  [451] -0.3768992657 -1.2496565373  0.3144199039 -1.2265472606  1.1289248426
#>  [456]  1.1910236426 -1.3927296940 -1.1988597952 -1.7955177522  1.3220761466
#>  [461]  0.0177777893  1.6332943004 -0.3245371054 -2.0614644494 -0.3801554715
#>  [466] -1.4627206041  0.1735860469 -0.9319588368 -2.9946616341  0.1988326695
#>  [471] -0.0647196875  2.1050805809  0.0598125831 -0.0500072303  1.9609257070
#>  [476]  0.8925402717 -0.3446531079 -1.7151982333  1.1786276078  0.0315281278
#>  [481]  1.1690277301  0.9353424941 -0.0326210447 -0.5279716891 -1.2381755010
#>  [486] -0.2351606238  1.0233363234 -0.6219670211 -0.3935845842  0.0145678154
#>  [491] -0.2668278069 -0.0695330781 -3.0359799632 -0.1114332859  1.8134318296
#>  [496]  1.2713105651 -1.2204665946  0.1505597539 -1.4766190051  0.4641401730
#>  [501]  0.4087605018 -0.0868345148 -0.8212853877  0.0260945818  0.7851734999
#>  [506]  0.6258444639 -0.5422068896  0.9205765819 -0.0804659296  0.6165221135
#>  [511] -1.4873295027 -1.0134920544 -0.3737494372 -0.7633740872  1.2016337956
#>  [516]  0.4771405810 -1.9197200454  0.1895862306  0.1895844103  0.7075617215
#>  [521]  1.0199515756 -0.4134930594  0.0856387873  0.9357868875  0.5335782183
#>  [526]  0.8993938627  1.1625471240 -0.1212327152  0.6661889087  0.7108187007
#>  [531] -1.1901871904 -0.7743854465 -0.8397975137 -2.6244590460 -0.2698316941
#>  [536]  2.8509153959 -0.7235651700  1.6074608951  2.9540384108  1.4349384697
#>  [541]  0.2268646914 -1.2881256801  0.5166302210 -1.5623113804 -2.1573588833
#>  [546]  0.4851757629 -0.5959148929 -0.3835797682 -0.1926566496  1.1637444769
#>  [551]  1.6368150291  1.5519008707 -1.5453588162 -1.4392227989 -0.0042049241
#>  [556] -1.1457424854 -1.0974711176  2.1614897888 -2.7804400923  1.3635450985
#>  [561]  0.2180348525 -0.0412921392 -1.3539473740  0.2226201351  2.9271313894
#>  [566]  0.4209627142 -0.1867510562  3.9321293184 -2.0991800977 -1.3921505698
#>  [571]  0.1005124550  0.4178843415 -0.1844664468  0.7835554369  1.0813821499
#>  [576]  0.6559991320  0.0664093545 -1.5183286696  1.2058113011  3.2987544800
#>  [581] -1.3459282249  1.4012061905  0.9720035010  1.0185304760  0.1610037970
#>  [586] -0.1148368968  0.0672197473 -0.5633593137 -0.3716169044 -1.0577265635
#>  [591] -0.1024982637  1.1986138405 -0.2627067699  0.3692510293 -0.5119262708
#>  [596] -0.1563745694 -2.8811735465 -0.7380262792 -0.2190060215  1.8973512721
#>  [601] -0.5440792626 -0.0976927444  0.2742132660 -1.6309452365  1.8033233187
#>  [606]  0.0639023233 -1.6948767238  0.1292928736 -4.2312868574  0.0617153634
#>  [611]  0.4012873535  0.9059568501 -1.2894125675  3.0255366006 -1.2060504032
#>  [616] -1.5059787928  0.9054493412 -0.1470613204 -0.4775153780 -0.7074968033
#>  [621]  1.0214785426  1.1657502209 -1.1265901124  0.9716693435 -0.0670490973
#>  [626] -0.2741007612 -0.9627737592 -1.3368880723  2.0446442434  0.1426165100
#>  [631]  0.1838717668  2.9234283420 -0.1174659413  0.5285442358 -1.5183782749
#>  [636] -0.1456683593  0.7436224328 -1.4025182643 -0.7280827079 -0.8195119470
#>  [641] -1.2390392323  0.4056616721 -0.5256450550  2.0471410192  0.8620711360
#>  [646]  0.4363385418  2.1905877359  0.0077888528  0.0625210643  2.1274355061
#>  [651]  2.6493838650  1.1291551375  2.6227566700 -0.2359820347  0.2333334731
#>  [656]  0.4327075966 -0.4417027640 -0.2777422753  0.2684389843  0.7054709225
#>  [661]  0.8411089119 -1.6847229314  1.4295250905  1.1235532451  1.9063955246
#>  [666]  2.3591222413  0.4633183373  1.2243202897 -0.1030040056  2.4655063640
#>  [671]  0.6364677294 -2.9771468223  0.0380181950 -1.3244914819 -0.6977229693
#>  [676]  0.3489365078  0.0354741809 -0.1129423621  0.8253506463  0.2559283605
#>  [681] -0.1034606980 -1.9245205749  2.3550297704 -1.1635068589  0.5914710358
#>  [686] -1.1155816170  0.5743820381 -1.2346267245 -0.1869203120  1.5147856036
#>  [691]  0.6541025736  0.3275177336  3.2877693765  0.3630533170  0.2681479326
#>  [696]  0.5113012250 -2.7153441234  0.3857158872 -0.2369325311  1.9897888220
#>  [701] -0.0458858353  2.0621424609 -1.1524648674  0.7434525473 -0.7182378035
#>  [706] -0.4602255720  0.8162645064  0.0811492145 -0.0975963731  1.8741664495
#>  [711]  0.1320260332 -3.4758522027 -0.3337148305 -0.0280630127  0.0283353820
#>  [716] -0.9952331849 -0.3071891960 -2.0980857340  1.2802504352 -0.9789614280
#>  [721]  1.1986566277  0.0583884624  0.1839120078  2.0933592264 -0.3225932472
#>  [726]  1.6454033020 -2.0069905007  1.4435279511  0.2991992358 -0.2268125605
#>  [731] -1.8558183178  1.2828776923 -0.7179107028  2.3955977425 -0.5389197284
#>  [736]  4.3462636399  0.3304800218  0.7180930267 -0.7505043081 -1.5509478954
#>  [741]  0.7174885342  0.8216363717 -0.3725459204 -1.8384341978 -1.5168694842
#>  [746] -0.0111569709 -0.0969618111 -0.9065379875 -0.4641098884  0.4990330074
#>  [751]  0.9562727787 -0.8570884096 -0.4336646247 -0.4361375363  0.2506424869
#>  [756] -0.7428842571  0.4839101719 -0.5063492702  0.6287533269  1.4932937942
#>  [761]  0.6529402462 -1.8929047321 -0.0184258789  0.1766321769 -1.3023854214
#>  [766] -1.0661612374  0.3731395289  0.0652530357  0.6867081301 -0.3571314079
#>  [771]  0.1121048817 -0.1241526198  0.1064139881  2.7399899693  0.3522681208
#>  [776]  1.8958541170 -1.5105948312  0.7742279963 -0.1360829160  0.8617551403
#>  [781] -0.3300336067 -0.0509399927 -1.3979291470 -1.3124577142  1.8152007814
#>  [786]  1.8043294637  2.1286407239 -1.1622138332  1.5434128811 -0.6420197321
#>  [791]  0.8790603164 -1.5329633620 -0.0418597385 -1.5568993793 -0.5932006143
#>  [796]  0.5295618959  1.1693261215 -0.0271798051 -0.7519300583  0.6656508014
#>  [801] -0.0768678113 -0.0304812424  0.1682496179 -0.3102963244 -0.8630317821
#>  [806] -0.1814799398  0.6588221665  0.1340767817 -0.0113777446 -0.8559880351
#>  [811] -0.4277009143  0.3520475776  0.4664044948  3.6773466871 -6.4703589316
#>  [816] -0.0053619779  0.3084366959  0.0521819030 -0.9593229058  1.5795121516
#>  [821] -0.0133674392  4.4520509679  0.2955472203  1.0688344988  2.9630863307
#>  [826] -0.4936772254  1.0963163134  0.8719496215  2.6547105753  1.6371163969
#>  [831]  3.3270758097  0.6562705926  0.1480537071 -0.0886791077  0.0186937747
#>  [836]  0.0620844793 -0.1385898392  0.2356381849  1.0761467723  0.1055733968
#>  [841] -1.9719152856  0.5998210884  0.0485273627 -0.9740874062  1.0012482186
#>  [846]  3.0742979246  2.0685408230  0.5334445318  0.9591207965  0.6680675113
#>  [851] -0.3504927676 -0.5992398562  0.9089312645 -0.6416077636 -0.2156056217
#>  [856] -3.4499018140  1.1640884190 -0.9650992093  0.6130524551  0.5087168647
#>  [861]  1.9757671952 -1.1770056389 -0.2491968034  1.0147702118 -0.0052122662
#>  [866]  1.2420651133 -0.2483174328  0.0528994846 -1.7276239782  1.8987864041
#>  [871]  1.1687713073  1.3838618558 -0.2567101141 -1.9086540103 -0.1945686808
#>  [876] -0.6751540934 -0.0642086350  1.8523007100 -0.7520547021  0.6918329798
#>  [881]  0.0234963575 -0.5020652209  0.2696429963  1.0067731679  2.0907209301
#>  [886] -0.3342182481  2.0752734522  3.7735565395  1.1606208869 -1.0140238229
#>  [891]  0.1541599974  0.6383762969 -0.2714590977 -0.1468503102  1.0863251220
#>  [896] -0.0026418074 -2.0874906664 -0.8291435968 -0.1417931448 -0.3774320657
#>  [901]  0.4080076984  0.5598392981 -0.3871088401  0.3172174264  1.7281993688
#>  [906]  0.1859628098 -0.5570218348  1.4530564501 -0.4158255319 -2.4575855332
#>  [911]  0.7095539892  2.2026666502 -0.1014395888  0.1796834352 -1.6106347413
#>  [916] -0.2866928832 -0.6728206266  0.9609960013  2.6711595258 -1.1152300109
#>  [921]  1.3143842664  3.0108477046  6.3252363047 -0.4675977894 -0.5871109172
#>  [926]  2.9914420521 -0.5912602600 -0.0197274286  3.1215164184 -0.5996727123
#>  [931] -0.2421836962  1.1942524003 -0.5165099004  1.8046160573 -0.1519277595
#>  [936]  1.6558518123 -0.4337851553  0.1777111376 -0.9345799966  0.3415013735
#>  [941] -0.6577461498 -0.6878102745 -0.0380173447 -1.1632390375 -1.7309548234
#>  [946] -0.4283487219  0.0984949974 -1.4346914296 -0.6586996601  0.1721831917
#>  [951]  0.3345461919  0.6068298360  1.1874594917  2.0102361411  0.5047738935
#>  [956] -0.0599676533 -1.8690434199 -0.2074233803  0.9540016031 -2.5437283452
#>  [961] -1.0461253526 -2.3575810234  1.5216018860  0.6574505472  0.0303195484
#>  [966]  0.1872920115  0.3700795474 -0.5444707765  1.1547363733  2.8780726702
#>  [971] -0.0236946272  1.1023169920 -1.8308981818 -2.7831889019 -0.6807993345
#>  [976]  1.5982209898 -0.3246568033 -1.0739274307  0.7692982098  0.7565085822
#>  [981] -3.3253987156 -0.2075335481 -0.1355385436 -0.2671900645 -4.8656867837
#>  [986] -0.1444968537 -0.9370558525 -0.1143422906 -2.2331063841  1.3367630779
#>  [991]  0.2563810744  1.0995077645  2.0310667975  2.9031385587  0.3044082532
#>  [996] -0.6952434522  1.3592674530 -0.2096029570  0.4967088690 -0.8575316462
#> 
#> $summary
#>        location scale
#> input      0.00  1.00
#> sample     0.03  0.98
#> 

generate_distribution(distribution = "gamma", alpha_shape = 3)
#> $data
#>    [1]  2.9982028  4.4356420  2.9101286  1.3484473  7.2796373  2.9186778
#>    [7]  3.9319749  9.5923683  1.1018435  1.6308848  0.3062231  1.0113663
#>   [13]  7.4579877  3.5137812  2.2306594  1.7201839  3.8290648  1.6229003
#>   [19]  5.1871664  2.5534472  2.8307473  3.5916542  4.5340822  0.7044389
#>   [25]  3.0519700  3.2855990  1.9070822  2.4336163  1.5466458  3.0182525
#>   [31]  7.2521452  0.9162253  4.5011770  2.9435758  3.3143879  2.1431869
#>   [37]  0.8733413  2.6194682  2.1014132  2.3548575  3.8573719  3.0509457
#>   [43]  6.3130077  1.5844639 12.1078802  2.5975433  4.0443563  1.0878199
#>   [49]  3.1643155  6.1186067  3.7443594  1.0996616  1.9445746  8.3432892
#>   [55]  4.9812606  2.8459340  7.2946416  6.7608269  2.7717754  4.4190488
#>   [61]  3.2164771  1.0902992  1.7404303  5.4229807  3.0168200  2.2965724
#>   [67]  3.9664047  6.1592793  1.8936871  3.6651785  2.1217136  6.5188663
#>   [73]  2.3435140  3.6577508  1.9808065  2.3012444  2.1714953  2.6619022
#>   [79]  3.1297294  4.4334801  1.9986071  4.6629802  2.3032051  2.1318950
#>   [85]  4.6187395  1.4971472  2.8249837  2.4548324  4.2770270  2.9084193
#>   [91]  3.4757942  3.5087011  0.4366386  0.9559139  1.1239945  6.1674093
#>   [97]  4.1661068  1.4377552  0.9181054  3.9208652  1.2191646  4.8798310
#>  [103]  4.7590980  1.9006763  3.0992719  7.8408779  1.4563305  0.9075892
#>  [109]  3.1988826  2.9103552  2.3223106  1.7840253  1.0347861  1.6519726
#>  [115]  5.6815658  1.2772517  3.9922750  2.0998008  6.0119951  1.0995889
#>  [121]  2.8441265  1.8495763  3.6780148  3.0865405  3.9118922  6.4200413
#>  [127]  0.6114179  0.9301245  1.4858916  3.4250945  2.4608119  1.4415606
#>  [133]  1.5575741  3.0420832  4.7384355  2.1677770  1.4026021  1.1587331
#>  [139]  2.3414865  1.5207662  2.8452287  2.8836630  4.0025794  0.6406239
#>  [145]  3.6691334  4.4314137  1.7494702  2.0493317  2.1301954  2.6936776
#>  [151]  5.8674436  1.8235272  2.1057946  6.4297393  1.3537983  4.2341646
#>  [157]  1.8834599  1.2930051  7.5921676  3.9281871  4.0882203  0.9922443
#>  [163]  2.0483418  3.7309182  2.3916444  1.9530579  3.7592821  3.4984314
#>  [169]  3.3512330  4.9205170  1.5997155  2.9354718  1.0254213  4.9258357
#>  [175]  1.4053876  4.6985670  3.6975343  2.8251785  0.7486631  3.7557858
#>  [181]  1.6725750  4.0388444  3.5817276  4.1886959  2.1064598  3.5357880
#>  [187]  3.9642354  3.6796540  1.6031126  2.5673750  9.1422049  3.5022129
#>  [193]  2.8847428  0.9158932  3.4056130  0.5617862  2.4188992  1.9904288
#>  [199]  1.0613183  1.3809265  2.1718180  1.8506033  4.3515987  1.1368529
#>  [205]  2.4449131  3.6200243  1.4972197  1.6419865  3.1699864  1.3034783
#>  [211]  1.6614050  3.1716891  0.5921338  5.6285613  0.6705353  2.1106361
#>  [217]  5.1246889  2.9321586  1.5202371  1.8739973  3.0169182  4.3260041
#>  [223]  5.7744097  4.8999385  2.4896875  1.5031872  1.3674657  1.3126851
#>  [229]  3.1258147  3.0698600  1.0201202  8.5102503  2.9536985  0.5914796
#>  [235]  3.5147704  0.9980626  6.8658697  3.4134022  0.9139375  3.2018868
#>  [241]  1.8828139  5.2259497  4.4469853  3.6276099  1.0048457  2.1756503
#>  [247]  5.1754196  1.7962386  1.2184571  1.8368665  2.8921761  4.4679389
#>  [253]  2.4944835  1.8557498  3.2050882  2.3185556  2.5442110  1.5521811
#>  [259]  3.7950207  2.4007651  2.3773563  4.1612024  2.2804883  3.5493577
#>  [265]  4.2661275  1.6745448  1.4142402  2.4436726  2.1236166  4.6626663
#>  [271]  2.0412737  4.1052431  2.1873679  5.7116509  1.6975595  2.2429170
#>  [277]  3.6805489  1.5018633  3.9949787  1.3244075  3.1991186  1.2624309
#>  [283]  3.1181522  2.6451972  1.4737550  1.7367610  1.4090199  0.8203262
#>  [289]  3.1511895  1.7214365  1.2760058  2.8970000  1.7223733  3.9657055
#>  [295]  1.5547884  2.8413781  6.9252108  0.7311365  3.5926845  2.2178055
#>  [301]  0.6114648  1.3632211  0.8957224  0.8365642  1.4903679  3.3334429
#>  [307]  3.7925241  3.9853103  2.2721609  3.4014963  1.7558932  4.9948342
#>  [313]  2.1126809  4.6544789  3.2259475  2.3868247  3.6048436  5.2355742
#>  [319]  3.7310377  1.8581189  1.8882025  2.5352259  1.1759798  2.8108804
#>  [325]  2.2316259  2.5431509  6.7558405  4.4847846  4.7955157  2.0093074
#>  [331]  5.0183501  3.6326547  3.0656022  0.7374370  2.0746689  5.5178338
#>  [337]  3.6063334  4.6154757  2.4422491  6.3215745  2.0808979  3.7007097
#>  [343]  2.3409106  3.7104966  1.4475602  4.3575577  3.0622718  0.9377085
#>  [349]  2.6528331  1.2915892  7.4531832  4.5216895  2.8770856  7.9351119
#>  [355]  2.4235937  2.8064979  2.3921318  4.4134602  3.6022802  0.9048058
#>  [361]  3.3622938  3.3390930  0.9729677  1.3822990  3.3879942  2.9163764
#>  [367]  1.8956443  0.7076483  2.9640691  6.0587414  1.6803403  3.2795575
#>  [373]  6.6919172  3.6504701  2.6176017  2.9357533  3.3999399  5.1176638
#>  [379]  0.6576248  2.4477303  3.3582827  1.2221393  3.3358601  3.7546031
#>  [385]  2.1496051  1.1511894  1.7026383  1.3792340  0.6861291  4.0876700
#>  [391]  0.9123238  3.2223858  4.0969185  1.5462090  3.6532393  5.1484219
#>  [397]  4.6163672  7.1179285  1.5155123  5.5697287  1.2664166  6.9584744
#>  [403]  1.7616046  1.8762896  3.2788256  1.2216185  8.2454409  3.5158087
#>  [409]  3.0809668  4.2378773  2.0486645  1.9105662  3.1897883  4.8513179
#>  [415]  2.3614906  2.0269784  3.2522691  1.0264754  4.1569647  1.9612660
#>  [421]  1.0751638  3.1557221  1.1490329  4.3530217  3.6421754  2.4259246
#>  [427]  1.6791204  4.5878653  2.7514988  1.9385851  4.1331922  2.3190606
#>  [433]  3.6522102  1.5471839  1.7120962  5.5660668  5.5865762  1.2017371
#>  [439]  2.4470156  6.3297105  2.5192428  1.4979845  5.5025024  4.6891551
#>  [445]  1.4977117  0.8436375  1.3039281  0.9588180  3.3692248  3.1259496
#>  [451]  2.7907500  0.9624828  1.9796283  1.4534783  0.8793458  2.5290155
#>  [457]  5.2216127  4.7234665  3.1033720  6.9820555  3.2222058  4.0172049
#>  [463]  1.6475472  3.5037845  3.2090957  2.2607011  0.5573524  4.0960181
#>  [469]  4.8655872  6.3635676  5.5782803  2.8895474  4.3285821  1.9650699
#>  [475]  0.8306138  3.5534055  2.0455067  1.3825601  4.3688680  4.5565292
#>  [481]  4.8853722  9.5086369  1.2935444  2.0482223  1.0844030  4.6911371
#>  [487]  2.2214866  3.3530107  1.6142241  2.1939994  3.2448519  1.5673707
#>  [493]  4.1378860  1.7873535  2.0598232  1.6613779  0.4971407  3.4189690
#>  [499]  0.7136461  2.2712468  1.5532374  0.5105254  2.2498660  1.6268725
#>  [505]  1.8961405  1.5070496  1.7530196  2.1780950  2.3193423  1.9332164
#>  [511]  2.4576139  3.3477276  4.7187087  2.3515667  1.7429875  5.2913510
#>  [517]  3.6780223  1.3925542  2.1637944  2.3616513  1.3896252  2.6102332
#>  [523]  1.1616361  2.4392273  0.6749743  2.5258140  2.1885860  4.4033989
#>  [529]  1.7661149  2.4307449  2.6980386  6.5091404  2.7938384  2.0434719
#>  [535]  3.8611044  4.5288668  1.8099404  3.6436345  2.7057250  2.9740134
#>  [541]  3.0801344  2.0111767  3.7236997  1.7011890  1.0433581  1.6042703
#>  [547]  2.2572131  4.8920699  8.5340771  5.4684484  2.9782106  2.0556083
#>  [553]  1.4614236  3.5346096  2.4204447  1.6171723  5.3079117  2.2216556
#>  [559]  0.9078282  2.8167592  2.0406571  2.4932861  3.4481484  1.9287942
#>  [565]  1.9955140  4.4434482  2.3722318  1.5410026  3.5996507  3.7964981
#>  [571]  1.4465864  3.5888277  2.6445667  1.2013826  0.7273593  4.8069228
#>  [577]  6.4884802  3.1303331  3.1022111  3.1808168  2.7469840  2.1314439
#>  [583]  3.2511674  1.3329075  1.2268198  7.7762647  3.5375367  1.8384520
#>  [589]  1.4329188  2.9536388  3.0176377  1.9977448  1.2612572  3.2363975
#>  [595]  1.4218378  1.8080711  1.6590129  3.6937906  2.2473440  3.2840634
#>  [601]  2.6471612  4.2630246  3.6325008  1.7793598  1.4154422  3.9587369
#>  [607]  1.0557357  0.4170193  1.4624754  3.9496873  1.6171807  4.9269180
#>  [613]  6.3938632  3.7552404  5.2885010  3.1902000  1.6342452  2.8577780
#>  [619]  3.6252620  2.5634226  6.2681084  1.4335362  1.5343004  3.1812078
#>  [625]  1.5463501  1.8920463  0.7942247  4.8064204  5.8402249  2.2021955
#>  [631]  7.2858167  3.1196905  1.2341175  4.1302534  1.0586429  1.9495670
#>  [637]  1.8495571  4.2402299  3.8261583  1.3869706  2.9692535  5.3834519
#>  [643]  1.1170748  3.2732924  5.4260262  3.4858598  1.9327315  1.2398297
#>  [649]  3.6573692  0.3636196  1.7229053  1.3832557  5.5483086  7.4790223
#>  [655]  2.2146120  4.0174977  3.9973804  4.4590450  1.9167913  2.9724904
#>  [661]  1.9866060  2.2845225  4.4761410  2.7677281  4.8523739  0.7323897
#>  [667]  6.5417013  1.4755465  3.9432740  2.3023164  6.1798606  3.4500490
#>  [673]  5.1223539  2.5159732  1.8943515  3.1129281  1.7956933  7.3390186
#>  [679]  2.1969074  4.7302959  7.8409908  1.4558142  1.4380886  0.9392420
#>  [685]  2.6166245  1.1571670  1.0822498  3.8050547  3.4067521  3.4702045
#>  [691]  6.1982229  4.3199017  2.0759266  1.8469291  2.5112877  3.7777791
#>  [697]  6.1366437  2.3602515  1.0400956  1.3164519  1.4430558  1.7599103
#>  [703]  2.0493645  2.9498720  1.2978266  2.8623467  2.1362407  0.9925647
#>  [709]  1.4258341  1.8140145  6.5619191  2.5200253  8.1017791  4.5440946
#>  [715]  1.7764580  4.4007816  0.6187638  2.3620804  5.7371071  4.8506792
#>  [721]  5.0669244  6.0357109  3.7224884  5.3201594  2.9841271  1.8880435
#>  [727]  1.1568905  4.9141423  3.5199619  3.9476839  4.9862381  4.1983461
#>  [733]  1.0258210  3.9215126  1.9797885  2.1603059  1.0284473  4.3838570
#>  [739]  0.9627353  1.9112734  1.2803915  1.2694184  1.9225441  1.8407986
#>  [745]  7.1639493  1.1306215  3.2787859  6.8452226  4.9418658  2.0211329
#>  [751]  3.4869708  5.5307107  2.9349834  4.5034929  1.8960824  3.8635674
#>  [757]  2.4135900  1.8138002  3.0895570  2.1105119  1.3704221  5.8453095
#>  [763]  2.0708424  5.0779635  4.4762118  3.9844230  4.2592843  3.4502770
#>  [769]  2.7950741  3.3779946  4.2216550  2.2583346  0.8196429  2.3409331
#>  [775]  5.0428776  4.3175286  5.0967431  1.4469266  3.3164573  4.0464518
#>  [781]  0.8277260  2.9239044  2.7313850  3.6906152  2.5897951  6.3379246
#>  [787]  3.9641702  2.7201839  3.1827147  5.3662308  5.2716084  2.2121922
#>  [793]  1.0170654  3.9516147  1.0559712  2.1960319  1.0525839  2.9290761
#>  [799]  2.9700395  3.3107330  2.4759345  3.1292588  4.7571073  5.1631699
#>  [805]  9.8730064  3.0569746  1.3805112  1.9153838  3.1884786  3.5916556
#>  [811]  2.4693922  2.1792706  4.0676323  0.6362322  1.9831764  3.1005652
#>  [817]  1.5072184  2.3527921  1.4156803  4.0460543  1.4407811  5.5351255
#>  [823]  6.7544381  2.6325963  2.1221565  3.3521829  1.0133924  3.3454528
#>  [829]  6.2757897  4.7060385  3.0484653  3.4770323  1.1788433  3.6524882
#>  [835]  5.4370859  1.8009611  3.1193663  0.6269011  0.6361991  2.8634051
#>  [841]  2.9850688  1.9452124  4.0008943  2.4247773  1.2408420  2.1285927
#>  [847]  1.8895929  3.0573853  3.3121458  5.0021514  2.5225871  4.3954564
#>  [853]  3.7098815  0.6506453  1.8391066  3.2692277  5.1269951  3.4774280
#>  [859]  2.5369065  2.5007890  1.5885773  1.9604413  5.5542682  0.4473006
#>  [865]  4.3477409  1.1084007  3.5437426  1.6768580  2.1133887  0.5331009
#>  [871]  1.3774777  2.3572226  1.3428488  3.1703229  1.8071387  2.5881065
#>  [877]  1.5824014  0.4164160  1.9720681  0.7465863  6.4553800  0.6530973
#>  [883]  2.1802174  1.8230100  3.2438867  1.7551143  0.9279966  5.7149362
#>  [889]  1.0338859  2.3423735  4.8288995  3.7316456  2.8124447  1.8335661
#>  [895]  4.8348595  2.4313094  3.6827872  2.5538495  2.3422152  3.8760587
#>  [901]  3.0497846  1.6188510  3.0548672  2.1309439  3.2998828  3.5437254
#>  [907]  0.9478614  2.0194885  4.6630209  1.2592649  1.8096141  0.8199782
#>  [913]  3.0873280  2.6896098  2.6127534  1.1170918  3.8279988  2.1671165
#>  [919]  4.6817582  3.6270002  2.1652035  1.6117543  3.1331609  2.2095712
#>  [925]  3.4513007  1.6645842  1.8615085  3.0786481  3.1115725  6.1399171
#>  [931]  5.1299015  1.6659027  4.5895871  1.9919497  1.5440833  1.7592995
#>  [937]  4.2457910  1.4710044  4.5091194  1.3250747  2.5530219  0.5705311
#>  [943]  1.9601932  4.4861277  0.8886172  3.6746970  2.4407806  1.8208970
#>  [949]  5.0366813  2.0469737  2.2019134  1.3324037  2.5524973  1.9250826
#>  [955]  6.4911806  2.9040047  1.6988899  2.7285598  1.8092552  2.8084471
#>  [961]  4.4325461  2.1882024  1.7444863  1.1362189  5.0859791  6.5690630
#>  [967]  2.2461716  3.3594266  7.7704914  5.6276846  2.3525012  2.3235855
#>  [973]  0.2850906  5.2655682  1.4543694  3.5471590  0.8822131  2.3751016
#>  [979]  3.1156263  6.2327645  2.0478710  0.9736861  5.6826848  7.6267141
#>  [985]  5.5492431  2.9101278  1.2302170  1.9428579  3.9698335  2.8032315
#>  [991]  4.1591250  4.4393395  5.5244260  3.8803361  1.6605067  7.8723190
#>  [997]  2.1696219  1.3350167  0.6539002  1.2602814
#> 
#> $summary
#>        shape rate
#> input   3.00 1.00
#> sample  3.14 1.05
#> 

generate_distribution(distribution = "beta", alpha_shape = 2, beta_shape = 2)
#> $data
#>    [1] 0.17622633 0.84860544 0.83640955 0.79753648 0.54399693 0.71597214
#>    [7] 0.48574559 0.10853165 0.39003657 0.17857425 0.87650639 0.61725699
#>   [13] 0.80285973 0.08843678 0.19160656 0.45448839 0.56903592 0.46592922
#>   [19] 0.23613750 0.39698743 0.26935951 0.47168380 0.45904672 0.15266367
#>   [25] 0.78091106 0.78054919 0.45352296 0.36081072 0.86989375 0.88192281
#>   [31] 0.28495461 0.32540902 0.57615419 0.52169919 0.16226859 0.59847120
#>   [37] 0.73890191 0.67762816 0.42933491 0.39421320 0.74107009 0.62236932
#>   [43] 0.12530111 0.37295969 0.50324266 0.57700644 0.81532494 0.46103899
#>   [49] 0.47240541 0.61117350 0.70167726 0.73851576 0.54789389 0.13393735
#>   [55] 0.42780445 0.79003451 0.29499292 0.06804159 0.48513707 0.52515437
#>   [61] 0.54895262 0.32209120 0.87008706 0.31554280 0.36249198 0.45108341
#>   [67] 0.28758876 0.89874742 0.45344438 0.34722224 0.65199042 0.82035664
#>   [73] 0.81789025 0.45457007 0.62776874 0.44344044 0.73323850 0.67757320
#>   [79] 0.25265174 0.12852661 0.08886347 0.42347928 0.19709740 0.38723797
#>   [85] 0.63117195 0.33830853 0.81782191 0.66159854 0.89172163 0.77823341
#>   [91] 0.81596413 0.70995920 0.38552192 0.86396078 0.42906093 0.21028036
#>   [97] 0.37328730 0.54891745 0.61002922 0.81896763 0.73895034 0.56971373
#>  [103] 0.62567102 0.52497821 0.54792236 0.16754455 0.57084745 0.24432948
#>  [109] 0.31662586 0.34028165 0.41244520 0.67887169 0.68358409 0.33193889
#>  [115] 0.77462069 0.31848933 0.45695378 0.66034019 0.49311773 0.60019569
#>  [121] 0.41471669 0.19181079 0.64883117 0.59645727 0.68409308 0.03181726
#>  [127] 0.64842819 0.36395621 0.60816589 0.59697896 0.34234900 0.59215142
#>  [133] 0.81911370 0.54180791 0.60669565 0.83303030 0.24786225 0.72076800
#>  [139] 0.20816890 0.77887853 0.92937439 0.64842126 0.91957575 0.45136847
#>  [145] 0.60930944 0.26311474 0.14011283 0.75140411 0.76412999 0.73663463
#>  [151] 0.25056089 0.12967366 0.48004318 0.67630741 0.16529947 0.76867269
#>  [157] 0.55726346 0.87451525 0.15148852 0.63319634 0.33944867 0.83193204
#>  [163] 0.78577549 0.58152469 0.90970927 0.59545083 0.43191892 0.66907619
#>  [169] 0.52183950 0.90806102 0.79765960 0.69983704 0.32346272 0.05419194
#>  [175] 0.25768420 0.07299058 0.09446610 0.12799208 0.91420716 0.25721078
#>  [181] 0.42188036 0.70926198 0.80419395 0.42873877 0.46561355 0.87816990
#>  [187] 0.11637848 0.42568939 0.91757678 0.52514710 0.46353833 0.81388853
#>  [193] 0.63676360 0.27542893 0.61806988 0.43086562 0.83997378 0.54463215
#>  [199] 0.68339591 0.42142442 0.50267529 0.54005638 0.50916152 0.54116386
#>  [205] 0.50553183 0.21471488 0.63450468 0.22877851 0.23784555 0.43906376
#>  [211] 0.76377692 0.65906357 0.45046589 0.71940326 0.40857647 0.19833393
#>  [217] 0.63276058 0.21642607 0.07711068 0.77072066 0.85529269 0.78439321
#>  [223] 0.43483543 0.47790955 0.32135934 0.35049372 0.58451660 0.71970778
#>  [229] 0.83523560 0.38178654 0.45282365 0.26231302 0.09549654 0.34510542
#>  [235] 0.56598565 0.25253158 0.41948731 0.12530808 0.19109187 0.83233268
#>  [241] 0.42985880 0.55682374 0.54798226 0.23949060 0.56754939 0.37029151
#>  [247] 0.41822295 0.43119420 0.56374763 0.74351625 0.36676660 0.39404591
#>  [253] 0.60889602 0.74381150 0.75171150 0.09883635 0.34641879 0.80425793
#>  [259] 0.57737411 0.50268495 0.99346163 0.75021254 0.62938316 0.39449282
#>  [265] 0.87125003 0.39433145 0.34595013 0.61451005 0.87933064 0.61984366
#>  [271] 0.47374474 0.66421819 0.60978113 0.30829958 0.38253039 0.47156549
#>  [277] 0.73053556 0.02243097 0.40470795 0.21268699 0.40670676 0.53211776
#>  [283] 0.18992312 0.38540675 0.20914821 0.41379792 0.24178848 0.91428915
#>  [289] 0.34470168 0.69446892 0.30682874 0.20060549 0.24661367 0.51906349
#>  [295] 0.38945662 0.70409801 0.57875854 0.65741896 0.79487996 0.77702861
#>  [301] 0.69930925 0.83294487 0.76033568 0.20929531 0.69066912 0.39493666
#>  [307] 0.81554071 0.24889385 0.29104940 0.49226610 0.52898816 0.49664355
#>  [313] 0.52163522 0.60320429 0.44235763 0.36685024 0.82504063 0.78027117
#>  [319] 0.59972427 0.22803801 0.52505997 0.60946663 0.55747298 0.64819701
#>  [325] 0.91907402 0.45377050 0.41824565 0.21586260 0.66001902 0.09632210
#>  [331] 0.58008957 0.62805404 0.61598223 0.35667248 0.82909940 0.60793776
#>  [337] 0.75253794 0.22475572 0.69671103 0.02155416 0.39125191 0.32601836
#>  [343] 0.39207284 0.69682585 0.16378639 0.75720272 0.09641000 0.49048719
#>  [349] 0.53822659 0.30616069 0.51509455 0.77381854 0.08935177 0.54504135
#>  [355] 0.65289589 0.43817853 0.80656214 0.29408465 0.72862252 0.29747099
#>  [361] 0.38894853 0.08440257 0.42088738 0.21458296 0.38137062 0.78045739
#>  [367] 0.38601594 0.62158668 0.17804583 0.56254169 0.57277073 0.31823661
#>  [373] 0.80744719 0.64191134 0.38147343 0.31196415 0.44134774 0.28835091
#>  [379] 0.92571122 0.62476033 0.58758827 0.47953894 0.59088172 0.72795054
#>  [385] 0.52329516 0.20960832 0.17679964 0.57182047 0.36267188 0.93534544
#>  [391] 0.27395892 0.21546642 0.73774112 0.25823388 0.74728840 0.40397041
#>  [397] 0.68618569 0.57533143 0.45992553 0.50698688 0.68921845 0.63036808
#>  [403] 0.29944318 0.14276809 0.73138457 0.20599031 0.45630031 0.71861221
#>  [409] 0.73229685 0.80779308 0.51768434 0.09947333 0.38752285 0.49326479
#>  [415] 0.56333862 0.64662901 0.42995191 0.67724209 0.49586840 0.39688520
#>  [421] 0.13911158 0.65645926 0.49081100 0.05258341 0.19645105 0.81412514
#>  [427] 0.68897928 0.56383561 0.37579868 0.52935539 0.89568483 0.98278295
#>  [433] 0.55061889 0.10414666 0.28596536 0.26588228 0.18099406 0.42158464
#>  [439] 0.33701590 0.80495350 0.25459587 0.56374427 0.37873723 0.23268937
#>  [445] 0.72377604 0.48620526 0.18586103 0.24305478 0.40792875 0.54778382
#>  [451] 0.86509266 0.34829519 0.23922766 0.23196760 0.36790430 0.25540113
#>  [457] 0.60257953 0.65548576 0.33972929 0.63131556 0.44251116 0.39006054
#>  [463] 0.45608184 0.67477435 0.39751947 0.90353802 0.34714864 0.35696256
#>  [469] 0.65760625 0.20371575 0.45036176 0.37995884 0.35416107 0.79291764
#>  [475] 0.44804137 0.85922312 0.69889484 0.67540220 0.80006706 0.71342149
#>  [481] 0.58556734 0.23421685 0.52976256 0.79918159 0.62262479 0.70886321
#>  [487] 0.92050545 0.52417400 0.44673975 0.01929110 0.58195368 0.28769601
#>  [493] 0.34036099 0.45560561 0.87108491 0.95846105 0.26575011 0.38422855
#>  [499] 0.74541573 0.75872184 0.68319019 0.27622651 0.50636522 0.47774648
#>  [505] 0.40821779 0.33324942 0.65035790 0.05298173 0.79789708 0.20525475
#>  [511] 0.51110383 0.91433511 0.53226690 0.59975083 0.19098957 0.44394433
#>  [517] 0.41673961 0.79344833 0.20360241 0.78259119 0.66986192 0.48823428
#>  [523] 0.62927777 0.26009780 0.65968004 0.36989145 0.14292451 0.03503043
#>  [529] 0.21247558 0.53633766 0.18826897 0.83976181 0.27672514 0.34304499
#>  [535] 0.50242138 0.46600067 0.54909274 0.90125924 0.70414455 0.41922628
#>  [541] 0.37962491 0.93668393 0.71842510 0.62001258 0.23569566 0.76180941
#>  [547] 0.13778648 0.56843643 0.80658606 0.51774527 0.32238833 0.77006868
#>  [553] 0.50251431 0.77132794 0.34781573 0.53730885 0.37774880 0.29672786
#>  [559] 0.39652831 0.37475533 0.38374758 0.72910792 0.36974564 0.66687100
#>  [565] 0.15429164 0.60409276 0.39175628 0.72983102 0.70956072 0.55753031
#>  [571] 0.34136913 0.25388314 0.46518179 0.68394931 0.56785316 0.53557050
#>  [577] 0.66753321 0.83124385 0.26506845 0.21342131 0.65499325 0.23582921
#>  [583] 0.59733844 0.78881028 0.35171842 0.80989080 0.16157736 0.27172711
#>  [589] 0.66612537 0.21988491 0.31782349 0.14424696 0.44046244 0.81212057
#>  [595] 0.71519150 0.76488085 0.93425701 0.48830679 0.69686117 0.49522446
#>  [601] 0.19643620 0.32866358 0.33117022 0.46863165 0.34128673 0.28357551
#>  [607] 0.72141626 0.82353890 0.39651046 0.65851907 0.39079241 0.15479447
#>  [613] 0.73159989 0.56676763 0.50618982 0.33760218 0.16337087 0.84722330
#>  [619] 0.43217484 0.36001179 0.09787823 0.48754613 0.19089395 0.13507143
#>  [625] 0.62331007 0.38944006 0.09152887 0.60235348 0.55451612 0.11625961
#>  [631] 0.37805379 0.29879013 0.28558481 0.74647660 0.30366409 0.88865298
#>  [637] 0.46593152 0.64974921 0.81008528 0.73770818 0.10882143 0.81993093
#>  [643] 0.26710429 0.74402335 0.08105044 0.72828740 0.42647412 0.50137508
#>  [649] 0.26202132 0.27186599 0.59736314 0.12138793 0.82865360 0.67464354
#>  [655] 0.33182167 0.32412746 0.31942371 0.21176443 0.89869246 0.82995289
#>  [661] 0.56122085 0.25499323 0.55035320 0.57722115 0.83770783 0.04055639
#>  [667] 0.41459062 0.64645775 0.53702255 0.36643302 0.21990409 0.33835079
#>  [673] 0.63120039 0.60371734 0.43621403 0.36911022 0.27206314 0.53112904
#>  [679] 0.41082458 0.25005632 0.19064412 0.09081921 0.86892835 0.56560385
#>  [685] 0.53679089 0.45011539 0.54556359 0.57893095 0.47183088 0.69566384
#>  [691] 0.77525849 0.95439874 0.25106311 0.16532296 0.57352767 0.26950224
#>  [697] 0.48333965 0.23780913 0.49114548 0.47240870 0.21533290 0.22665947
#>  [703] 0.35710590 0.75991570 0.28815301 0.16376668 0.40834383 0.70261297
#>  [709] 0.73441292 0.79815212 0.57305145 0.34307047 0.28291385 0.57566136
#>  [715] 0.53519149 0.14352596 0.41612175 0.82670558 0.69508400 0.33941096
#>  [721] 0.04782922 0.37297898 0.51745880 0.37539397 0.54715915 0.31148836
#>  [727] 0.68191538 0.49892469 0.28763008 0.61297379 0.49899603 0.52981569
#>  [733] 0.54329485 0.59124498 0.31674179 0.62484305 0.25107694 0.75022330
#>  [739] 0.43323534 0.59767988 0.66214863 0.62486394 0.49119161 0.37506343
#>  [745] 0.46250862 0.31662398 0.19206652 0.56679223 0.41489535 0.61597267
#>  [751] 0.22427838 0.58817903 0.71899531 0.65075240 0.55847013 0.29339118
#>  [757] 0.60435802 0.37114596 0.22831527 0.63047187 0.33112976 0.72190620
#>  [763] 0.19145728 0.21664467 0.63516860 0.33693067 0.46555253 0.22737660
#>  [769] 0.33537483 0.45003614 0.53431350 0.59294579 0.52169886 0.74294704
#>  [775] 0.23955724 0.34799769 0.14550795 0.75917055 0.42554925 0.38020137
#>  [781] 0.07857246 0.83080394 0.59573458 0.99346647 0.46270997 0.31617989
#>  [787] 0.92549259 0.72527179 0.35889110 0.45715675 0.26643116 0.69645857
#>  [793] 0.68655427 0.59263072 0.55466695 0.50997996 0.90440026 0.65335787
#>  [799] 0.39769021 0.68948474 0.08312539 0.60574834 0.83407847 0.11083697
#>  [805] 0.22969476 0.22916579 0.50735118 0.77976491 0.68886028 0.52153131
#>  [811] 0.17206329 0.32718412 0.59128964 0.75890145 0.77864565 0.40072890
#>  [817] 0.73980140 0.18889024 0.13780249 0.85846067 0.46879572 0.26751654
#>  [823] 0.34114763 0.63390165 0.45977806 0.87543546 0.62635098 0.58464178
#>  [829] 0.23525474 0.69162638 0.50275427 0.15512308 0.52687977 0.35137631
#>  [835] 0.31869159 0.47884636 0.86696598 0.26582491 0.14816597 0.93773071
#>  [841] 0.49842977 0.63216073 0.60311161 0.40742205 0.92922169 0.56602421
#>  [847] 0.83451780 0.52510207 0.15547232 0.80191160 0.44418698 0.58293213
#>  [853] 0.85020036 0.73904769 0.66544483 0.20408819 0.22522199 0.52560215
#>  [859] 0.31382216 0.66096835 0.73160745 0.22018997 0.76398652 0.22836991
#>  [865] 0.54255913 0.72093915 0.26027138 0.27008347 0.32365616 0.32653515
#>  [871] 0.44654328 0.31424272 0.35648309 0.46797758 0.17892977 0.63470728
#>  [877] 0.43685459 0.68896549 0.52977463 0.46165772 0.48506004 0.48427656
#>  [883] 0.41542047 0.35819950 0.67340467 0.20136562 0.82450590 0.22536536
#>  [889] 0.44618357 0.33639029 0.39541631 0.86305816 0.53377189 0.75903002
#>  [895] 0.21780182 0.65957746 0.14632939 0.53260058 0.51894179 0.19770311
#>  [901] 0.65756100 0.70455255 0.52790644 0.67072674 0.17951366 0.38731075
#>  [907] 0.19772213 0.62662533 0.53887878 0.61555694 0.24573291 0.66799774
#>  [913] 0.41455351 0.69353428 0.14042923 0.14476617 0.89938828 0.63495065
#>  [919] 0.69999202 0.37072185 0.60336641 0.33409833 0.50548014 0.50968371
#>  [925] 0.57427119 0.70400452 0.56559813 0.59864034 0.61148911 0.49982161
#>  [931] 0.26382218 0.07772544 0.24615257 0.45967254 0.86234287 0.41624022
#>  [937] 0.42110390 0.20802519 0.22874172 0.68928754 0.75657564 0.70095905
#>  [943] 0.42431052 0.19087261 0.60118638 0.18707482 0.73590291 0.73037195
#>  [949] 0.26778852 0.25248976 0.83500580 0.82690921 0.48857545 0.49969701
#>  [955] 0.44771265 0.93451945 0.52593214 0.65654134 0.03833165 0.38665002
#>  [961] 0.64454597 0.11913035 0.63432371 0.53298509 0.51229772 0.23288664
#>  [967] 0.40561484 0.25585478 0.18853793 0.69190901 0.82427995 0.61202624
#>  [973] 0.86618705 0.20494903 0.90460123 0.20712791 0.26343813 0.84594748
#>  [979] 0.36529649 0.45987698 0.47932832 0.47530761 0.44060215 0.76575641
#>  [985] 0.25332281 0.94620690 0.39446307 0.65305219 0.62894807 0.44323548
#>  [991] 0.26622017 0.40250097 0.52978553 0.71135107 0.79024128 0.26693432
#>  [997] 0.36424564 0.79379151 0.94467927 0.37695025
#> 
#> $summary
#>        alpha shape beta shape
#> input         2.00          2
#> sample        1.98          2
#>