Generate k groups of quantitative data with optional standardized effect

generate_quantitative() is a function to generate k groups of quantitative data with specified first four moments and optional standardized effect

generate_quantitative(
  k = 1,
  mean = 0,
  var = 1,
  skew = 0,
  kurt = 0,
  n = 100,
  effect_size = 0,
  k_names = seq(1:k),
  use_effect_size = FALSE,
  summary = TRUE,
  decimals = 3,
  replication = 1
)

Arguments

k

A numeric vector for the number of groups to generate data for. The default value is 1.

mean

A numeric vector for the population mean of the first variable. The default value is 0.

var

A numeric vector for the population variance of the first variable. The default value is 1.

skew

A numeric vector for the population skew of the variables. Separate values of skew can be entered. The default value is 0.

kurt

A numeric vector for the population kurtosis of the variables. Separate values of kurtosis can be entered. The default value is 0.

n

A numeric vector for the sample size. The default value is 100.

effect_size

An optional numeric vector for the standardized effect size between groups of data.

k_names

An optional character vector for naming successes and failures. The default values are 1:k.

use_effect_size

A logical vector to indicate whether to use a standardized effect size in generating data for k when k >= 2. The default value is FALSE.

summary

A logical vector for whether the return should include summary statistics. The default value is TRUE.

decimals

A numeric vector for the number of decimals to round the sample data to. The default value is 3.

replication

A numeric vector for the number of times to replicate the sampling. The default value is 1.

Value

If summary = TRUE, a list containing a matrix ("summary") of the input parameters and descriptive statistics and a data frame ("sample") with n rows and 2 columns. If summary = FALSE, a data frame with n rows and 2 columns.

Examples

generate_quantitative()
#> $summary
#>         mean   sd  skew kurtosis
#> input 1 0.00 1.00  0.00     0.00
#> k 1     0.03 1.06 -0.08    -0.34
#> 
#> $sample
#>     Group        Value
#> 1       1  1.001470396
#> 2       1 -0.407549943
#> 3       1 -0.951635255
#> 4       1 -0.006694559
#> 5       1 -2.210258385
#> 6       1 -0.933209093
#> 7       1  0.036322008
#> 8       1  2.339142073
#> 9       1  0.095554647
#> 10      1 -0.122885614
#> 11      1  1.333565508
#> 12      1 -2.125733472
#> 13      1 -0.912269368
#> 14      1 -0.218098804
#> 15      1  0.730586985
#> 16      1  1.562776379
#> 17      1 -2.086864740
#> 18      1  1.695149041
#> 19      1 -1.005438481
#> 20      1 -0.381367052
#> 21      1  2.449720565
#> 22      1  0.260859847
#> 23      1  0.573788175
#> 24      1 -1.276270143
#> 25      1 -1.968386658
#> 26      1 -0.347501729
#> 27      1 -0.083058743
#> 28      1 -2.150146765
#> 29      1  0.546742948
#> 30      1 -0.204190304
#> 31      1  0.982585022
#> 32      1  1.323939437
#> 33      1  2.273471898
#> 34      1 -0.710162623
#> 35      1  0.597028327
#> 36      1  0.997813932
#> 37      1 -0.699127010
#> 38      1 -2.350006106
#> 39      1  0.264503288
#> 40      1  1.210198449
#> 41      1  0.606447324
#> 42      1 -0.372109423
#> 43      1  0.154246896
#> 44      1  0.930117949
#> 45      1 -1.638551043
#> 46      1  0.255329799
#> 47      1 -1.381741240
#> 48      1 -0.528453577
#> 49      1  0.228701620
#> 50      1  0.623010186
#> 51      1 -0.876258937
#> 52      1  0.748047799
#> 53      1 -0.274521029
#> 54      1  0.826576345
#> 55      1  1.612594140
#> 56      1  1.933676822
#> 57      1  1.292197379
#> 58      1 -1.586068413
#> 59      1  0.319403277
#> 60      1  1.570932833
#> 61      1 -0.704645524
#> 62      1  1.118114891
#> 63      1 -0.005402569
#> 64      1  1.444523603
#> 65      1  0.552889328
#> 66      1 -1.319615435
#> 67      1 -1.639631708
#> 68      1 -0.616630872
#> 69      1 -0.352422136
#> 70      1 -0.525166583
#> 71      1 -0.648871744
#> 72      1 -0.089331659
#> 73      1 -0.483081649
#> 74      1  0.547452293
#> 75      1  0.550986638
#> 76      1 -0.071623695
#> 77      1 -0.122256431
#> 78      1 -0.309518485
#> 79      1  0.982922851
#> 80      1  1.349499042
#> 81      1  0.969687622
#> 82      1 -1.071634761
#> 83      1 -0.576107270
#> 84      1 -1.222816579
#> 85      1  1.308803735
#> 86      1  0.236564021
#> 87      1  0.801206169
#> 88      1 -0.306947517
#> 89      1  0.653080416
#> 90      1  0.006105675
#> 91      1  0.492276563
#> 92      1 -0.368952875
#> 93      1  0.530513783
#> 94      1  0.030770227
#> 95      1 -1.062270026
#> 96      1  0.067033194
#> 97      1 -0.440133285
#> 98      1 -0.638808633
#> 99      1  0.526662933
#> 100     1 -0.509245448
#> 

generate_quantitative(k = 2, effect_size = 0.3, use_effect_size = TRUE)
#> $summary
#>          mean   sd  skew kurtosis
#> input 1  0.00 1.00  0.00     0.00
#> k 1     -0.11 0.97 -0.18    -0.51
#> input 2  0.30 1.00  0.00     0.00
#> k 2      0.29 1.07 -0.02    -0.37
#> 
#> $sample
#> # A tibble: 200 × 2
#>    Group   Value
#>    <fct>   <dbl>
#>  1 1     -1.47  
#>  2 1     -0.653 
#>  3 1     -0.683 
#>  4 1      0.0263
#>  5 1      0.474 
#>  6 1      0.713 
#>  7 1     -2.07  
#>  8 1      1.31  
#>  9 1      0.0708
#> 10 1     -0.301 
#> # ℹ 190 more rows
#> 

generate_quantitative(k = 2, skew = c(0.4, 0.25))
#> $summary
#>          mean   sd skew kurtosis
#> input 1  0.00 1.00 0.40     0.00
#> k 1      0.06 0.95 0.01    -0.42
#> input 2  0.00 1.00 0.25     0.00
#> k 2     -0.02 1.01 0.25     0.00
#> 
#> $sample
#> # A tibble: 200 × 2
#>    Group   Value
#>    <fct>   <dbl>
#>  1 1     -0.964 
#>  2 1      0.483 
#>  3 1     -0.840 
#>  4 1      0.404 
#>  5 1     -1.01  
#>  6 1      0.427 
#>  7 1     -0.0738
#>  8 1      0.226 
#>  9 1      1.24  
#> 10 1     -0.0890
#> # ℹ 190 more rows
#>