Generate samples of binary data

generate_binary() is function to generate sample binary data.

generate_binary(k = 1, n = 100, p = 0.5, group_names = c(1:0), replication = 1)

Arguments

k

A numeric vector for the number of variables. The default value is 1.

n

A numeric vector for the sample size. The default value is 100.

p

A numeric vector for the proportion of successes. The default value is 0.5. If k > 1, you can provide separate proportion values for each k.

group_names

An optional character vector for naming successes and failures. The default values are 1 for success and 0 for failure.

replication

A numeric vector for the number of times to replicate the sampling. The default value is 1.

Value

A list containing

sample, a data frame with n rows and k columns

summary, a matrix of the sample size, number of successes, proportion statistic, and proportion parameter for each variable, and

freq.table, a matrix of the frequencies of successes and failures for each variable

Examples

generate_binary()
#> $summary
#>          sample size successes sample proportion population proportion
#> sample 1         100        44              0.44                   0.5
#> 
#> $sample
#>     v1
#> 1    0
#> 2    1
#> 3    1
#> 4    0
#> 5    0
#> 6    0
#> 7    0
#> 8    0
#> 9    1
#> 10   1
#> 11   1
#> 12   0
#> 13   0
#> 14   0
#> 15   0
#> 16   0
#> 17   0
#> 18   0
#> 19   0
#> 20   1
#> 21   0
#> 22   1
#> 23   1
#> 24   0
#> 25   1
#> 26   1
#> 27   0
#> 28   1
#> 29   1
#> 30   1
#> 31   0
#> 32   0
#> 33   0
#> 34   1
#> 35   0
#> 36   0
#> 37   1
#> 38   1
#> 39   0
#> 40   0
#> 41   1
#> 42   0
#> 43   1
#> 44   1
#> 45   0
#> 46   1
#> 47   1
#> 48   1
#> 49   1
#> 50   0
#> 51   0
#> 52   0
#> 53   0
#> 54   1
#> 55   0
#> 56   1
#> 57   0
#> 58   1
#> 59   0
#> 60   0
#> 61   1
#> 62   0
#> 63   1
#> 64   0
#> 65   1
#> 66   0
#> 67   1
#> 68   1
#> 69   1
#> 70   0
#> 71   0
#> 72   0
#> 73   0
#> 74   0
#> 75   0
#> 76   0
#> 77   0
#> 78   1
#> 79   1
#> 80   0
#> 81   1
#> 82   1
#> 83   0
#> 84   0
#> 85   0
#> 86   1
#> 87   0
#> 88   0
#> 89   1
#> 90   0
#> 91   1
#> 92   0
#> 93   1
#> 94   1
#> 95   1
#> 96   0
#> 97   1
#> 98   1
#> 99   0
#> 100  0
#> 
#> $freq.table
#>          successes failures
#> sample 1        44       56
#> 

generate_binary(k = 2, p = c(0.5, 0.7))
#> $summary
#>          sample size successes sample proportion population proportion
#> sample 1         100        60              0.60                   0.5
#> sample 2         100        68              0.68                   0.7
#> 
#> $sample
#>     v1 v2
#> 1    1  1
#> 2    1  0
#> 3    0  1
#> 4    0  1
#> 5    0  1
#> 6    1  0
#> 7    1  1
#> 8    1  1
#> 9    1  1
#> 10   1  0
#> 11   1  1
#> 12   1  0
#> 13   0  1
#> 14   1  0
#> 15   1  1
#> 16   1  0
#> 17   0  1
#> 18   1  0
#> 19   1  0
#> 20   1  0
#> 21   1  1
#> 22   1  1
#> 23   0  1
#> 24   1  1
#> 25   1  1
#> 26   0  1
#> 27   0  1
#> 28   1  0
#> 29   0  0
#> 30   0  0
#> 31   0  0
#> 32   1  1
#> 33   1  1
#> 34   1  1
#> 35   1  1
#> 36   1  1
#> 37   1  1
#> 38   0  1
#> 39   0  0
#> 40   1  1
#> 41   0  0
#> 42   1  1
#> 43   1  1
#> 44   0  1
#> 45   1  0
#> 46   0  1
#> 47   1  1
#> 48   0  0
#> 49   1  1
#> 50   1  1
#> 51   1  1
#> 52   0  1
#> 53   1  0
#> 54   0  1
#> 55   0  1
#> 56   0  0
#> 57   1  1
#> 58   0  1
#> 59   0  1
#> 60   1  1
#> 61   1  1
#> 62   1  0
#> 63   1  0
#> 64   1  0
#> 65   1  1
#> 66   1  1
#> 67   1  1
#> 68   0  1
#> 69   0  1
#> 70   1  1
#> 71   1  1
#> 72   0  0
#> 73   0  1
#> 74   1  0
#> 75   0  1
#> 76   1  1
#> 77   1  0
#> 78   0  1
#> 79   0  1
#> 80   1  0
#> 81   1  1
#> 82   0  1
#> 83   1  1
#> 84   0  1
#> 85   1  0
#> 86   0  1
#> 87   0  0
#> 88   0  0
#> 89   1  0
#> 90   0  1
#> 91   1  0
#> 92   0  1
#> 93   0  1
#> 94   1  1
#> 95   1  1
#> 96   0  1
#> 97   1  1
#> 98   1  1
#> 99   0  0
#> 100  1  1
#> 
#> $freq.table
#>          successes failures
#> sample 1        60       40
#> sample 2        68       32
#> 

generate_binary(group_names = c("success", "fail"))
#> $summary
#>          sample size successes sample proportion population proportion
#> sample 1         100        58              0.58                   0.5
#> 
#> $sample
#>          v1
#> 1   success
#> 2      fail
#> 3   success
#> 4   success
#> 5      fail
#> 6      fail
#> 7      fail
#> 8      fail
#> 9   success
#> 10     fail
#> 11  success
#> 12  success
#> 13  success
#> 14  success
#> 15     fail
#> 16     fail
#> 17  success
#> 18     fail
#> 19  success
#> 20     fail
#> 21     fail
#> 22  success
#> 23  success
#> 24  success
#> 25     fail
#> 26     fail
#> 27  success
#> 28  success
#> 29     fail
#> 30  success
#> 31     fail
#> 32     fail
#> 33     fail
#> 34  success
#> 35  success
#> 36     fail
#> 37     fail
#> 38  success
#> 39  success
#> 40  success
#> 41     fail
#> 42  success
#> 43     fail
#> 44  success
#> 45  success
#> 46  success
#> 47  success
#> 48  success
#> 49  success
#> 50     fail
#> 51  success
#> 52  success
#> 53     fail
#> 54     fail
#> 55     fail
#> 56  success
#> 57     fail
#> 58     fail
#> 59  success
#> 60  success
#> 61     fail
#> 62  success
#> 63     fail
#> 64  success
#> 65     fail
#> 66     fail
#> 67  success
#> 68  success
#> 69  success
#> 70  success
#> 71  success
#> 72     fail
#> 73     fail
#> 74  success
#> 75  success
#> 76     fail
#> 77  success
#> 78  success
#> 79  success
#> 80  success
#> 81  success
#> 82  success
#> 83     fail
#> 84     fail
#> 85     fail
#> 86  success
#> 87     fail
#> 88  success
#> 89  success
#> 90     fail
#> 91  success
#> 92     fail
#> 93  success
#> 94  success
#> 95  success
#> 96     fail
#> 97  success
#> 98  success
#> 99  success
#> 100    fail
#> 
#> $freq.table
#>          successes failures
#> sample 1        58       42
#>