Generate samples of categorical data

generate_categorical() is function to generate n variables of sample categorical data with k groups.

generate_categorical(
  p = 1,
  n = 100,
  k = 1,
  k_prop = "equal",
  k_names = "default",
  replication = 1
)

Arguments

p

A numeric vector for the number of variables. The default value is 1.

n

A numeric vector for the sample size. The default value is 100.

k

A numeric vector for the number of groups within each variable. The default value is 1. If p>1, you can provide separate group numbers for each p.

k_prop

A vector for the proportion of each group in the population. The default value is "equal". For unequal proportions, provide a concatenated vector the same length as the number of groups (k). If p>1, you can provide separate sets of proportions for each p.

k_names

An optional character vector for naming successes and failures. The default values are 1:k.

replication

A numeric vector for the number of times to replicate the sampling. The default value is 1.

Value

A list containing a data frame ("sample") with n rows and p columns and a matrix ("proportion_summary") of the input and sample proportions for each group and variable.

Examples

generate_categorical()
#> $sample
#>     v1
#> 1    1
#> 2    1
#> 3    1
#> 4    1
#> 5    1
#> 6    1
#> 7    1
#> 8    1
#> 9    1
#> 10   1
#> 11   1
#> 12   1
#> 13   1
#> 14   1
#> 15   1
#> 16   1
#> 17   1
#> 18   1
#> 19   1
#> 20   1
#> 21   1
#> 22   1
#> 23   1
#> 24   1
#> 25   1
#> 26   1
#> 27   1
#> 28   1
#> 29   1
#> 30   1
#> 31   1
#> 32   1
#> 33   1
#> 34   1
#> 35   1
#> 36   1
#> 37   1
#> 38   1
#> 39   1
#> 40   1
#> 41   1
#> 42   1
#> 43   1
#> 44   1
#> 45   1
#> 46   1
#> 47   1
#> 48   1
#> 49   1
#> 50   1
#> 51   1
#> 52   1
#> 53   1
#> 54   1
#> 55   1
#> 56   1
#> 57   1
#> 58   1
#> 59   1
#> 60   1
#> 61   1
#> 62   1
#> 63   1
#> 64   1
#> 65   1
#> 66   1
#> 67   1
#> 68   1
#> 69   1
#> 70   1
#> 71   1
#> 72   1
#> 73   1
#> 74   1
#> 75   1
#> 76   1
#> 77   1
#> 78   1
#> 79   1
#> 80   1
#> 81   1
#> 82   1
#> 83   1
#> 84   1
#> 85   1
#> 86   1
#> 87   1
#> 88   1
#> 89   1
#> 90   1
#> 91   1
#> 92   1
#> 93   1
#> 94   1
#> 95   1
#> 96   1
#> 97   1
#> 98   1
#> 99   1
#> 100  1
#> 
#> $proportion_summary
#>           k   n k1
#> input p1  1 100  1
#> sample p1 1 100  1
#> 

generate_categorical(k = 4)
#> $sample
#>     v1
#> 1    2
#> 2    2
#> 3    3
#> 4    1
#> 5    2
#> 6    4
#> 7    2
#> 8    2
#> 9    1
#> 10   1
#> 11   2
#> 12   3
#> 13   4
#> 14   3
#> 15   3
#> 16   2
#> 17   3
#> 18   4
#> 19   1
#> 20   2
#> 21   4
#> 22   1
#> 23   2
#> 24   2
#> 25   3
#> 26   3
#> 27   3
#> 28   3
#> 29   2
#> 30   2
#> 31   2
#> 32   4
#> 33   1
#> 34   4
#> 35   2
#> 36   2
#> 37   3
#> 38   4
#> 39   4
#> 40   4
#> 41   4
#> 42   3
#> 43   2
#> 44   1
#> 45   3
#> 46   1
#> 47   4
#> 48   3
#> 49   4
#> 50   2
#> 51   1
#> 52   4
#> 53   1
#> 54   4
#> 55   1
#> 56   4
#> 57   1
#> 58   4
#> 59   3
#> 60   3
#> 61   1
#> 62   3
#> 63   2
#> 64   1
#> 65   4
#> 66   2
#> 67   4
#> 68   1
#> 69   4
#> 70   4
#> 71   1
#> 72   3
#> 73   2
#> 74   2
#> 75   4
#> 76   1
#> 77   1
#> 78   4
#> 79   3
#> 80   2
#> 81   3
#> 82   4
#> 83   3
#> 84   4
#> 85   2
#> 86   2
#> 87   4
#> 88   4
#> 89   1
#> 90   1
#> 91   2
#> 92   1
#> 93   2
#> 94   2
#> 95   1
#> 96   4
#> 97   4
#> 98   3
#> 99   1
#> 100  3
#> 
#> $proportion_summary
#>           k   n   k1   k2   k3   k4
#> input p1  4 100 0.25 0.25 0.25 0.25
#> sample p1 4 100 0.23 0.27 0.22 0.28
#> 

generate_categorical(k = 4, k_names = c("freshman", "sophomore", "junior", "senior"))
#> $sample
#>            v1
#> 1      junior
#> 2      senior
#> 3   sophomore
#> 4      senior
#> 5    freshman
#> 6    freshman
#> 7    freshman
#> 8   sophomore
#> 9      senior
#> 10   freshman
#> 11     junior
#> 12   freshman
#> 13     senior
#> 14   freshman
#> 15   freshman
#> 16  sophomore
#> 17   freshman
#> 18     senior
#> 19  sophomore
#> 20     junior
#> 21     junior
#> 22     senior
#> 23  sophomore
#> 24     senior
#> 25     senior
#> 26     senior
#> 27  sophomore
#> 28     senior
#> 29     senior
#> 30     junior
#> 31     senior
#> 32     senior
#> 33   freshman
#> 34     senior
#> 35     junior
#> 36     junior
#> 37  sophomore
#> 38     junior
#> 39     senior
#> 40  sophomore
#> 41     senior
#> 42     senior
#> 43     senior
#> 44  sophomore
#> 45   freshman
#> 46     senior
#> 47     senior
#> 48  sophomore
#> 49   freshman
#> 50  sophomore
#> 51  sophomore
#> 52   freshman
#> 53     senior
#> 54     junior
#> 55   freshman
#> 56  sophomore
#> 57     senior
#> 58  sophomore
#> 59     senior
#> 60   freshman
#> 61   freshman
#> 62   freshman
#> 63  sophomore
#> 64   freshman
#> 65   freshman
#> 66     senior
#> 67     junior
#> 68   freshman
#> 69     senior
#> 70     junior
#> 71   freshman
#> 72   freshman
#> 73  sophomore
#> 74     senior
#> 75  sophomore
#> 76     senior
#> 77     senior
#> 78  sophomore
#> 79   freshman
#> 80   freshman
#> 81     junior
#> 82     senior
#> 83   freshman
#> 84     junior
#> 85     junior
#> 86  sophomore
#> 87  sophomore
#> 88     junior
#> 89  sophomore
#> 90     senior
#> 91   freshman
#> 92     senior
#> 93     junior
#> 94     senior
#> 95     junior
#> 96     senior
#> 97     junior
#> 98   freshman
#> 99   freshman
#> 100 sophomore
#> 
#> $proportion_summary
#>           k   n   k1   k2   k3   k4
#> input p1  4 100 0.25 0.25 0.25 0.25
#> sample p1 4 100 0.27 0.22 0.18 0.33
#> 

generate_categorical(k = 4, k_prop = c(0.25, 0.5, 0.1, 0.15))
#> $sample
#>     v1
#> 1    2
#> 2    1
#> 3    2
#> 4    1
#> 5    1
#> 6    2
#> 7    3
#> 8    2
#> 9    3
#> 10   1
#> 11   3
#> 12   2
#> 13   2
#> 14   2
#> 15   3
#> 16   2
#> 17   2
#> 18   4
#> 19   2
#> 20   4
#> 21   2
#> 22   2
#> 23   4
#> 24   2
#> 25   4
#> 26   2
#> 27   3
#> 28   2
#> 29   2
#> 30   1
#> 31   4
#> 32   2
#> 33   2
#> 34   4
#> 35   2
#> 36   2
#> 37   4
#> 38   2
#> 39   2
#> 40   3
#> 41   1
#> 42   4
#> 43   2
#> 44   1
#> 45   2
#> 46   2
#> 47   2
#> 48   2
#> 49   2
#> 50   3
#> 51   4
#> 52   2
#> 53   1
#> 54   2
#> 55   2
#> 56   1
#> 57   2
#> 58   1
#> 59   4
#> 60   2
#> 61   1
#> 62   4
#> 63   2
#> 64   1
#> 65   2
#> 66   2
#> 67   1
#> 68   2
#> 69   1
#> 70   2
#> 71   1
#> 72   2
#> 73   1
#> 74   2
#> 75   4
#> 76   2
#> 77   2
#> 78   2
#> 79   1
#> 80   3
#> 81   2
#> 82   1
#> 83   4
#> 84   2
#> 85   3
#> 86   2
#> 87   2
#> 88   1
#> 89   1
#> 90   2
#> 91   3
#> 92   1
#> 93   2
#> 94   1
#> 95   4
#> 96   3
#> 97   2
#> 98   2
#> 99   2
#> 100  2
#> 
#> $proportion_summary
#>           k   n   k1   k2   k3   k4
#> input p1  4 100 0.25 0.50 0.10 0.15
#> sample p1 4 100 0.22 0.53 0.11 0.14
#> 

generate_categorical(p = 2, k = c(2, 3), k_prop = c(0.3, 0.7, 0.2, 0.2, 0.6))
#> $sample
#>     v1 v2
#> 1    1  1
#> 2    2  3
#> 3    2  2
#> 4    1  1
#> 5    2  3
#> 6    2  3
#> 7    1  3
#> 8    2  3
#> 9    2  3
#> 10   2  1
#> 11   2  3
#> 12   2  1
#> 13   1  3
#> 14   1  3
#> 15   1  3
#> 16   2  2
#> 17   2  1
#> 18   2  2
#> 19   2  3
#> 20   2  3
#> 21   2  1
#> 22   2  1
#> 23   2  3
#> 24   2  3
#> 25   2  1
#> 26   2  2
#> 27   2  3
#> 28   1  3
#> 29   1  3
#> 30   2  3
#> 31   1  3
#> 32   2  1
#> 33   2  3
#> 34   1  3
#> 35   2  3
#> 36   2  3
#> 37   1  1
#> 38   2  3
#> 39   2  2
#> 40   1  2
#> 41   1  3
#> 42   1  2
#> 43   2  3
#> 44   2  2
#> 45   1  3
#> 46   2  2
#> 47   2  3
#> 48   2  1
#> 49   2  3
#> 50   2  3
#> 51   1  3
#> 52   1  2
#> 53   2  3
#> 54   1  3
#> 55   2  3
#> 56   2  1
#> 57   2  2
#> 58   2  2
#> 59   1  3
#> 60   1  2
#> 61   2  3
#> 62   2  3
#> 63   2  2
#> 64   2  3
#> 65   1  2
#> 66   2  3
#> 67   1  1
#> 68   2  3
#> 69   2  1
#> 70   2  3
#> 71   2  3
#> 72   2  3
#> 73   2  2
#> 74   2  3
#> 75   2  2
#> 76   1  3
#> 77   2  1
#> 78   1  3
#> 79   2  3
#> 80   2  2
#> 81   2  3
#> 82   2  2
#> 83   1  3
#> 84   2  3
#> 85   2  3
#> 86   2  2
#> 87   2  3
#> 88   2  3
#> 89   2  3
#> 90   2  2
#> 91   2  3
#> 92   2  2
#> 93   2  3
#> 94   1  3
#> 95   2  2
#> 96   2  1
#> 97   1  3
#> 98   1  3
#> 99   2  1
#> 100  2  1
#> 
#> $proportion_summary
#>           k   n   k1   k2   k3
#> input p1  2 100 0.30 0.70   NA
#> sample p1 2 100 0.28 0.72   NA
#> input p2  3 100 0.20 0.20 0.60
#> sample p2 3 100 0.18 0.23 0.59
#> 

generate_categorical(p = 2, k = 2, k_prop = c(0.3, 0.7))
#> $sample
#>     v1 v2
#> 1    2  2
#> 2    2  2
#> 3    1  2
#> 4    1  2
#> 5    2  1
#> 6    2  1
#> 7    2  2
#> 8    2  2
#> 9    2  2
#> 10   2  1
#> 11   2  2
#> 12   2  2
#> 13   1  2
#> 14   2  2
#> 15   1  2
#> 16   2  1
#> 17   1  1
#> 18   2  2
#> 19   1  2
#> 20   1  2
#> 21   2  1
#> 22   2  1
#> 23   1  2
#> 24   1  2
#> 25   2  2
#> 26   2  1
#> 27   2  1
#> 28   2  2
#> 29   2  2
#> 30   2  2
#> 31   2  2
#> 32   2  2
#> 33   1  2
#> 34   1  2
#> 35   2  2
#> 36   1  2
#> 37   2  1
#> 38   1  2
#> 39   1  2
#> 40   2  1
#> 41   2  2
#> 42   1  2
#> 43   2  2
#> 44   2  2
#> 45   1  2
#> 46   1  1
#> 47   1  2
#> 48   2  1
#> 49   1  1
#> 50   2  1
#> 51   2  2
#> 52   1  2
#> 53   2  1
#> 54   1  2
#> 55   2  2
#> 56   2  2
#> 57   2  2
#> 58   2  2
#> 59   2  2
#> 60   2  1
#> 61   2  2
#> 62   2  2
#> 63   1  2
#> 64   1  1
#> 65   2  2
#> 66   1  2
#> 67   2  2
#> 68   1  2
#> 69   2  2
#> 70   2  2
#> 71   1  1
#> 72   2  2
#> 73   1  1
#> 74   2  1
#> 75   2  1
#> 76   1  1
#> 77   2  2
#> 78   1  2
#> 79   1  2
#> 80   2  1
#> 81   1  1
#> 82   2  2
#> 83   2  2
#> 84   2  2
#> 85   1  2
#> 86   2  2
#> 87   2  2
#> 88   2  1
#> 89   2  2
#> 90   2  2
#> 91   2  2
#> 92   2  1
#> 93   2  2
#> 94   1  1
#> 95   1  2
#> 96   2  2
#> 97   1  1
#> 98   1  2
#> 99   2  2
#> 100  2  1
#> 
#> $proportion_summary
#>           k   n   k1   k2
#> input p1  2 100 0.30 0.70
#> sample p1 2 100 0.36 0.64
#> input p2  2 100 0.30 0.70
#> sample p2 2 100 0.30 0.70
#>