Generate sample data from named distributions

generate_distribution() is a wrapper function for generating sample data from named continuous distributions.

The distributional shapes include beta, Cauchy, chi-squared, exponential, F, gamma, Laplace, logistic, log normal, normal, t, triangular, uniform, and Weibull. See additional details below for distribution parameters and default values.

Usage

generate_distribution(
  distribution = c("beta", "cauchy", "chi-squared", "exponential", "f", "gamma",
    "laplace", "logistic", "log normal", "normal", "t", "triangular", "uniform",
    "weibull"),
  n = 1000,
  seed = NULL,
  alpha_shape = NULL,
  beta_shape = NULL,
  location = NULL,
  scale = NULL,
  df = NULL,
  df2 = NULL,
  rate = NULL,
  mean_log = NULL,
  sd_log = NULL,
  mean = NULL,
  sd = NULL,
  theta = NULL,
  min = NULL,
  max = NULL,
  summary = TRUE
)

Arguments

distribution

A character vector to identify which distributional shape to sample. Options include beta, cauchy, chi-squared, exponential, f, gamma, laplace, logistic, log normal, normal, t, triangular, uniform, and weibull.

n

A numeric vector for the sample size. The default value is 1000.

seed

An optional numeric vector to use in set.seed()

alpha_shape

A non-negative numeric vector required for the shape of the beta, gamma, and Weibull distributions.

beta_shape

A non-negative numeric vector required for the shape of the beta distribution.

location

A numeric vector for the location of the Cauchy, Laplace, and logistic distributions.

scale

A numeric vector for the scale of the Cauchy, Laplace, logistic, and Weibull distributions.

df

A numeric vector required for the degrees of freedom of the chi-squared, F, and t distributions.

df2

A numeric vector required for the degrees of freedom of the F distribution.

rate

A numeric vector for the rate of the exponential and gamma distributions.

mean_log

A numeric vector for the mean (on the log scale) of the log normal distribution.

sd_log

A numeric vector for the standard deviation (on the log scale) of the log normal distribution.

mean

A numeric vector for the mean of the normal distribution.

sd

A numeric vector for the standard deviation of the normal distribution.

theta

A numeric vector required for the theta of the triangular distribution.

min

A numeric vector for the minimum value of the triangular and uniform distributions.

max

A numeric vector for the maximum value of the triangular and uniform distributions.

summary

A logical vector for whether the return should include summary statistics. The default value is TRUE.

Value

A list of

data, a vector of length n containing the generated values

summary, a descriptive matrix of input and sample values of the specified distribution's parameters

Distributional Shapes & Default Values

The beta distribution requires the alpha and beta shape parameters. There are no default values for this distribution. See stats::rbeta() for more information.

The Cauchy distribution uses the location and scale parameters. The default values are location = 0 and scale = 1. See stats::rcauchy() for more information.

The chi-squared distribution requires the df parameter. There is no default value for this distribution. See stats::rchisq() for more information.

The exponential distribution uses the rate parameter, which has a default value of 1. See stats::rexp() for more information.

The F distribution requires the df and df2 parameters. There are no default values for this distribution. See stats::rf() for more information.

The gamma distribution requires the alpha shape and rate parameter. There is no default value for alpha shape, but the default value for rate is 1. See stats::rgamma() for more information.

The Laplace distribution uses the location and scale parameters. The default values are location = 0 and scale = 1. See VGAM::rlaplace() for more information.

The logistic distribution uses the location and scale parameters. The default values are location = 0 and scale = 1. See stats::rlogis() for more information.

The log normal distribution uses the mean log and sd log parameters. The default values are mean log = 0 and sd log = 1. See stats::rlnorm() for more information.

The normal distribution uses the mean and sd parameters. The default values are mean = 0 and sd = 1. See stats::rnorm() for more information.

The t distribution requires the df parameter. There is no default value for this distribution. See stats::rt() for more information.

The triangular distribution requires the theta, min, and max parameters. There is no default value for theta, but the other default values are min = 0 and max = 1. See VGAM::rtriangle() for more information.

The uniform distribution uses the min and max parameters. The default values are min = 0 and max = 1. See stats::runif() for more information.

The Weibull distribution requires the alpha shape and scale parameter. There is no default value for alpha shape, but the default value for scale is 1. See stats::rweibull() for more information.

Examples

generate_distribution(distribution = "laplace")
#> $data
#>    [1] -2.3284022369 -0.4493223967  0.9921234668 -1.2361831337  0.9438122028
#>    [6]  0.6632238634 -1.9769620888 -1.1605278679  0.0815273966  0.6342219612
#>   [11]  1.5002533466  0.0998631814 -0.4541358992  0.5345637060  3.2206846787
#>   [16] -1.4513768212 -0.0916067736 -2.5303548226 -0.4219391893  0.8020082829
#>   [21]  0.2918765715 -0.3885768735  0.0160476620  0.0018968493  2.7094110383
#>   [26]  0.0621615262 -0.0855559861  0.1977583163  0.3130567272 -0.0003658732
#>   [31]  0.9933383196  0.3614747883  0.1231470142  0.1814954440  0.0176328582
#>   [36]  1.3885494714  1.4976568648  0.0963608545 -0.2946210399 -0.7175644197
#>   [41]  1.2660044492 -0.9172056161  0.8626963375 -0.3824947255 -0.8869771032
#>   [46] -1.3542901267  0.2553027769  1.4482322155  2.2145595524  0.8093691584
#>   [51] -0.4650025177 -1.6451955708  0.1324358008 -0.5058746059 -1.3117452110
#>   [56]  4.3440915222 -0.0984092376  0.3494105889 -0.0492736719  0.0150230560
#>   [61]  1.9092990035 -0.0561298681  0.2164822759 -0.1097245222  2.2631110089
#>   [66]  0.6118659138 -0.6212695357  0.1341953181 -3.8085597898 -1.8897133644
#>   [71] -0.6687362074  0.0041356677  0.4746771954  0.1847724079 -1.1190918125
#>   [76]  1.5667652614 -0.2268050533 -0.7567125176 -0.4436303097  0.5767845608
#>   [81] -0.0764358717  0.3509411510 -0.9685758229  1.2727484608  0.7635999912
#>   [86]  0.2417391549 -2.2498847226  0.9054189360 -0.2996201125  0.4230906065
#>   [91]  2.8573862693 -0.1139252748 -5.9926624086  0.1294505280 -1.6283745971
#>   [96]  2.2240957773  0.1213040393  1.7845175011 -0.8037772219  0.7759767305
#>  [101] -0.3916473937 -1.4421042814 -0.5251633699  0.1667105301  2.9433611850
#>  [106] -0.1248690302  1.3573696747  0.0285328484 -0.2878806495  0.7863913491
#>  [111]  0.6049228287  0.6760998986 -0.5841816560  0.0448366446 -1.9484815291
#>  [116] -0.7371735782  0.1487312908 -0.5817079777  0.9851431348  2.1078552706
#>  [121] -0.5214125163 -0.0392959211 -2.3671102131 -0.2457398510  0.4787070582
#>  [126] -4.0485541283 -2.1201193304  0.1957239948  3.8468065329  2.6028039781
#>  [131]  1.0848940284  2.4299328608  1.8508146718 -0.5635398063 -1.4217010724
#>  [136] -2.7382308444  0.4134139524 -0.7064391586 -0.7011386813 -1.7655768195
#>  [141] -0.0418607684 -1.2997845798  2.3090445523  0.1368008788  0.4190533624
#>  [146]  0.3825575779  0.4393719227 -0.1046548151 -0.2100712827 -0.8886364533
#>  [151] -2.3686884266 -0.2125452654 -0.4233022006 -0.3942173054 -1.1303703392
#>  [156]  0.8175660152 -0.3010604674 -0.3486956060  1.1184964631  0.6888692371
#>  [161]  0.2050478792  2.3252356463  0.5815620890 -0.8758416578  0.7270325812
#>  [166]  0.5292759818  0.3434412176 -0.3998563528  3.7858920162 -2.8596444406
#>  [171] -0.0503408285 -0.8005853184 -0.3730668388  0.5452221705  0.1027950954
#>  [176]  3.7521087821  0.2964479342  0.7954871651  0.5744184868  0.5089673890
#>  [181] -0.8383003966 -1.7340926363 -0.3407772758 -0.0833733927  0.7515830976
#>  [186] -3.1550846417  0.7932670202 -0.4262547287 -1.3562016853  0.8831097326
#>  [191] -0.1705550697 -0.1943374471 -0.0329447888 -0.3476590140  4.6042033112
#>  [196] -1.0517116566 -1.1543074073  0.0029469037  0.5146344460 -2.0815654266
#>  [201] -1.8503964561  0.5444550606 -0.1801176744  3.2485221183  1.0232453394
#>  [206] -0.8029148010  2.2795892666 -1.5108429371  1.5826110995  0.1311264539
#>  [211]  0.7520743928  0.1343318504 -1.0869953493 -1.6169700429 -0.4443032002
#>  [216]  2.4903568863 -3.1967831457 -1.1189110932  1.0799443714  0.9383305451
#>  [221] -0.5837997796 -0.1089401602  2.0612893964  0.8608216356 -1.5494058930
#>  [226] -0.7249142123 -0.7402527835  2.1731912473 -1.5590793856 -1.8942547391
#>  [231]  0.6753303141 -1.2542636024 -0.1286311172  0.9851367831 -1.5888840544
#>  [236]  0.5148457738  1.5208085722  0.0252547627  0.5447911560 -0.8699188658
#>  [241] -2.4901601747  0.4960938647  0.6694267029 -0.4031302687 -0.0615254334
#>  [246]  0.8207760526 -0.5482927809  1.9732781367 -0.3094729006 -1.8436178084
#>  [251] -1.0276860296  2.3519985092 -0.2765545842 -1.0736602424 -1.2010450277
#>  [256]  0.6971555626 -0.7635788272  0.7867399247  1.6996554916  1.3864798214
#>  [261] -0.6631520203  0.0312432494 -0.0348306222 -0.8325312135 -0.9290893787
#>  [266]  0.0711617214 -0.6435930740  0.1935536582  2.0337699065  1.3317112738
#>  [271]  0.9891104406  1.3783992388  1.7277481798  0.5576353915  0.7896175212
#>  [276] -0.0173141435  0.0412582557  0.2460969219 -0.3768992657 -1.2496565373
#>  [281]  0.3144199039 -1.2265472606  1.1289248426  1.1910236426 -1.3927296940
#>  [286] -1.1988597952 -1.7955177522  1.3220761466  0.0177777893  1.6332943004
#>  [291] -0.3245371054 -2.0614644494 -0.3801554715 -1.4627206041  0.1735860469
#>  [296] -0.9319588368 -2.9946616341  0.1988326695 -0.0647196875  2.1050805809
#>  [301]  0.0598125831 -0.0500072303  1.9609257070  0.8925402717 -0.3446531079
#>  [306] -1.7151982333  1.1786276078  0.0315281278  1.1690277301  0.9353424941
#>  [311] -0.0326210447 -0.5279716891 -1.2381755010 -0.2351606238  1.0233363234
#>  [316] -0.6219670211 -0.3935845842  0.0145678154 -0.2668278069 -0.0695330781
#>  [321] -3.0359799632 -0.1114332859  1.8134318296  1.2713105651 -1.2204665946
#>  [326]  0.1505597539 -1.4766190051  0.4641401730  0.4087605018 -0.0868345148
#>  [331] -0.8212853877  0.0260945818  0.7851734999  0.6258444639 -0.5422068896
#>  [336]  0.9205765819 -0.0804659296  0.6165221135 -1.4873295027 -1.0134920544
#>  [341] -0.3737494372 -0.7633740872  1.2016337956  0.4771405810 -1.9197200454
#>  [346]  0.1895862306  0.1895844103  0.7075617215  1.0199515756 -0.4134930594
#>  [351]  0.0856387873  0.9357868875  0.5335782183  0.8993938627  1.1625471240
#>  [356] -0.1212327152  0.6661889087  0.7108187007 -1.1901871904 -0.7743854465
#>  [361] -0.8397975137 -2.6244590460 -0.2698316941  2.8509153959 -0.7235651700
#>  [366]  1.6074608951  2.9540384108  1.4349384697  0.2268646914 -1.2881256801
#>  [371]  0.5166302210 -1.5623113804 -2.1573588833  0.4851757629 -0.5959148929
#>  [376] -0.3835797682 -0.1926566496  1.1637444769  1.6368150291  1.5519008707
#>  [381] -1.5453588162 -1.4392227989 -0.0042049241 -1.1457424854 -1.0974711176
#>  [386]  2.1614897888 -2.7804400923  1.3635450985  0.2180348525 -0.0412921392
#>  [391] -1.3539473740  0.2226201351  2.9271313894  0.4209627142 -0.1867510562
#>  [396]  3.9321293184 -2.0991800977 -1.3921505698  0.1005124550  0.4178843415
#>  [401] -0.1844664468  0.7835554369  1.0813821499  0.6559991320  0.0664093545
#>  [406] -1.5183286696  1.2058113011  3.2987544800 -1.3459282249  1.4012061905
#>  [411]  0.9720035010  1.0185304760  0.1610037970 -0.1148368968  0.0672197473
#>  [416] -0.5633593137 -0.3716169044 -1.0577265635 -0.1024982637  1.1986138405
#>  [421] -0.2627067699  0.3692510293 -0.5119262708 -0.1563745694 -2.8811735465
#>  [426] -0.7380262792 -0.2190060215  1.8973512721 -0.5440792626 -0.0976927444
#>  [431]  0.2742132660 -1.6309452365  1.8033233187  0.0639023233 -1.6948767238
#>  [436]  0.1292928736 -4.2312868574  0.0617153634  0.4012873535  0.9059568501
#>  [441] -1.2894125675  3.0255366006 -1.2060504032 -1.5059787928  0.9054493412
#>  [446] -0.1470613204 -0.4775153780 -0.7074968033  1.0214785426  1.1657502209
#>  [451] -1.1265901124  0.9716693435 -0.0670490973 -0.2741007612 -0.9627737592
#>  [456] -1.3368880723  2.0446442434  0.1426165100  0.1838717668  2.9234283420
#>  [461] -0.1174659413  0.5285442358 -1.5183782749 -0.1456683593  0.7436224328
#>  [466] -1.4025182643 -0.7280827079 -0.8195119470 -1.2390392323  0.4056616721
#>  [471] -0.5256450550  2.0471410192  0.8620711360  0.4363385418  2.1905877359
#>  [476]  0.0077888528  0.0625210643  2.1274355061  2.6493838650  1.1291551375
#>  [481]  2.6227566700 -0.2359820347  0.2333334731  0.4327075966 -0.4417027640
#>  [486] -0.2777422753  0.2684389843  0.7054709225  0.8411089119 -1.6847229314
#>  [491]  1.4295250905  1.1235532451  1.9063955246  2.3591222413  0.4633183373
#>  [496]  1.2243202897 -0.1030040056  2.4655063640  0.6364677294 -2.9771468223
#>  [501]  0.0380181950 -1.3244914819 -0.6977229693  0.3489365078  0.0354741809
#>  [506] -0.1129423621  0.8253506463  0.2559283605 -0.1034606980 -1.9245205749
#>  [511]  2.3550297704 -1.1635068589  0.5914710358 -1.1155816170  0.5743820381
#>  [516] -1.2346267245 -0.1869203120  1.5147856036  0.6541025736  0.3275177336
#>  [521]  3.2877693765  0.3630533170  0.2681479326  0.5113012250 -2.7153441234
#>  [526]  0.3857158872 -0.2369325311  1.9897888220 -0.0458858353  2.0621424609
#>  [531] -1.1524648674  0.7434525473 -0.7182378035 -0.4602255720  0.8162645064
#>  [536]  0.0811492145 -0.0975963731  1.8741664495  0.1320260332 -3.4758522027
#>  [541] -0.3337148305 -0.0280630127  0.0283353820 -0.9952331849 -0.3071891960
#>  [546] -2.0980857340  1.2802504352 -0.9789614280  1.1986566277  0.0583884624
#>  [551]  0.1839120078  2.0933592264 -0.3225932472  1.6454033020 -2.0069905007
#>  [556]  1.4435279511  0.2991992358 -0.2268125605 -1.8558183178  1.2828776923
#>  [561] -0.7179107028  2.3955977425 -0.5389197284  4.3462636399  0.3304800218
#>  [566]  0.7180930267 -0.7505043081 -1.5509478954  0.7174885342  0.8216363717
#>  [571] -0.3725459204 -1.8384341978 -1.5168694842 -0.0111569709 -0.0969618111
#>  [576] -0.9065379875 -0.4641098884  0.4990330074  0.9562727787 -0.8570884096
#>  [581] -0.4336646247 -0.4361375363  0.2506424869 -0.7428842571  0.4839101719
#>  [586] -0.5063492702  0.6287533269  1.4932937942  0.6529402462 -1.8929047321
#>  [591] -0.0184258789  0.1766321769 -1.3023854214 -1.0661612374  0.3731395289
#>  [596]  0.0652530357  0.6867081301 -0.3571314079  0.1121048817 -0.1241526198
#>  [601]  0.1064139881  2.7399899693  0.3522681208  1.8958541170 -1.5105948312
#>  [606]  0.7742279963 -0.1360829160  0.8617551403 -0.3300336067 -0.0509399927
#>  [611] -1.3979291470 -1.3124577142  1.8152007814  1.8043294637  2.1286407239
#>  [616] -1.1622138332  1.5434128811 -0.6420197321  0.8790603164 -1.5329633620
#>  [621] -0.0418597385 -1.5568993793 -0.5932006143  0.5295618959  1.1693261215
#>  [626] -0.0271798051 -0.7519300583  0.6656508014 -0.0768678113 -0.0304812424
#>  [631]  0.1682496179 -0.3102963244 -0.8630317821 -0.1814799398  0.6588221665
#>  [636]  0.1340767817 -0.0113777446 -0.8559880351 -0.4277009143  0.3520475776
#>  [641]  0.4664044948  3.6773466871 -6.4703589316 -0.0053619779  0.3084366959
#>  [646]  0.0521819030 -0.9593229058  1.5795121516 -0.0133674392  4.4520509679
#>  [651]  0.2955472203  1.0688344988  2.9630863307 -0.4936772254  1.0963163134
#>  [656]  0.8719496215  2.6547105753  1.6371163969  3.3270758097  0.6562705926
#>  [661]  0.1480537071 -0.0886791077  0.0186937747  0.0620844793 -0.1385898392
#>  [666]  0.2356381849  1.0761467723  0.1055733968 -1.9719152856  0.5998210884
#>  [671]  0.0485273627 -0.9740874062  1.0012482186  3.0742979246  2.0685408230
#>  [676]  0.5334445318  0.9591207965  0.6680675113 -0.3504927676 -0.5992398562
#>  [681]  0.9089312645 -0.6416077636 -0.2156056217 -3.4499018140  1.1640884190
#>  [686] -0.9650992093  0.6130524551  0.5087168647  1.9757671952 -1.1770056389
#>  [691] -0.2491968034  1.0147702118 -0.0052122662  1.2420651133 -0.2483174328
#>  [696]  0.0528994846 -1.7276239782  1.8987864041  1.1687713073  1.3838618558
#>  [701] -0.2567101141 -1.9086540103 -0.1945686808 -0.6751540934 -0.0642086350
#>  [706]  1.8523007100 -0.7520547021  0.6918329798  0.0234963575 -0.5020652209
#>  [711]  0.2696429963  1.0067731679  2.0907209301 -0.3342182481  2.0752734522
#>  [716]  3.7735565395  1.1606208869 -1.0140238229  0.1541599974  0.6383762969
#>  [721] -0.2714590977 -0.1468503102  1.0863251220 -0.0026418074 -2.0874906664
#>  [726] -0.8291435968 -0.1417931448 -0.3774320657  0.4080076984  0.5598392981
#>  [731] -0.3871088401  0.3172174264  1.7281993688  0.1859628098 -0.5570218348
#>  [736]  1.4530564501 -0.4158255319 -2.4575855332  0.7095539892  2.2026666502
#>  [741] -0.1014395888  0.1796834352 -1.6106347413 -0.2866928832 -0.6728206266
#>  [746]  0.9609960013  2.6711595258 -1.1152300109  1.3143842664  3.0108477046
#>  [751]  6.3252363047 -0.4675977894 -0.5871109172  2.9914420521 -0.5912602600
#>  [756] -0.0197274286  3.1215164184 -0.5996727123 -0.2421836962  1.1942524003
#>  [761] -0.5165099004  1.8046160573 -0.1519277595  1.6558518123 -0.4337851553
#>  [766]  0.1777111376 -0.9345799966  0.3415013735 -0.6577461498 -0.6878102745
#>  [771] -0.0380173447 -1.1632390375 -1.7309548234 -0.4283487219  0.0984949974
#>  [776] -1.4346914296 -0.6586996601  0.1721831917  0.3345461919  0.6068298360
#>  [781]  1.1874594917  2.0102361411  0.5047738935 -0.0599676533 -1.8690434199
#>  [786] -0.2074233803  0.9540016031 -2.5437283452 -1.0461253526 -2.3575810234
#>  [791]  1.5216018860  0.6574505472  0.0303195484  0.1872920115  0.3700795474
#>  [796] -0.5444707765  1.1547363733  2.8780726702 -0.0236946272  1.1023169920
#>  [801] -1.8308981818 -2.7831889019 -0.6807993345  1.5982209898 -0.3246568033
#>  [806] -1.0739274307  0.7692982098  0.7565085822 -3.3253987156 -0.2075335481
#>  [811] -0.1355385436 -0.2671900645 -4.8656867837 -0.1444968537 -0.9370558525
#>  [816] -0.1143422906 -2.2331063841  1.3367630779  0.2563810744  1.0995077645
#>  [821]  2.0310667975  2.9031385587  0.3044082532 -0.6952434522  1.3592674530
#>  [826] -0.2096029570  0.4967088690 -0.8575316462  0.2697405973  1.5926007934
#>  [831]  1.2250018942 -0.4089331505  0.2194735091 -2.4606993727 -0.9137768033
#>  [836]  0.4257205387  0.8223247346  3.6694387390 -4.0561941593  0.2243058503
#>  [841] -0.7237083301  0.8635824210 -1.4760195666  6.0202211431  0.1757886203
#>  [846] -1.2453992589 -2.4608252221  0.5144176802 -0.6104908217  0.9701170754
#>  [851] -0.4462913247 -3.2231869991  0.3733361755  2.5446552683 -1.3872729091
#>  [856]  1.1621563256  2.0193593605  3.8410195040 -0.6542806019  0.5845096551
#>  [861]  0.0480859118 -0.1497501065  0.3452489667  0.6480297480 -0.5280140914
#>  [866]  0.3568534223 -3.1900228762  0.7930578569 -0.1576394748 -0.6181496605
#>  [871]  0.1075893020  0.8159540347  1.8098266697 -2.2843530001  0.0271895118
#>  [876]  0.7205016958  0.1750924403 -1.7762307571  0.6349380233 -1.3405572571
#>  [881]  1.2986347904  0.4373564471 -1.9812336600  0.0198361955  0.1552091169
#>  [886]  0.3009486180 -1.2301245733  0.4410068458  0.3069136965 -0.3726446176
#>  [891] -0.3216088073 -1.1132541784 -0.0342957598  1.0184216426  2.3368380718
#>  [896] -0.6937874434  0.3017160719 -2.1077316329  0.2813322259 -2.2236411163
#>  [901]  3.6431522043 -0.1798109228 -1.5508229645 -0.0139022769  2.0565614303
#>  [906]  1.2739172973  0.0732176017  0.2384370928  0.2654866501  0.4587540922
#>  [911] -1.6407888308 -0.2049211413 -0.5475112410 -0.3116351287 -1.6299777916
#>  [916]  0.0369322063  0.5376801408  0.0613733518  0.7173112122 -0.2325254799
#>  [921] -0.9908554379 -0.0540036667 -0.0771299306  1.8724309064 -0.2463728101
#>  [926]  0.8123395141  2.6099644262  0.3003504321 -0.0622380002  2.7728823033
#>  [931] -2.6540623751 -0.6557017874 -0.9921922933 -1.6491313164 -6.1995912206
#>  [936]  1.6922019017 -0.8176946363  2.7540131002 -0.8988440026 -0.3807364366
#>  [941]  0.1471836941 -0.7106729533 -0.3813714934  0.0251649093  0.0499485132
#>  [946] -0.8243386078  0.9419224493 -1.1454784304 -1.2665909902  1.4537128382
#>  [951]  1.0228010521  0.3674068935 -1.7242233430  2.5999158268  1.7371601983
#>  [956]  0.7358990426  1.9095540425 -1.2486781294  0.1957612019 -1.3300087258
#>  [961] -0.3440236900  0.4458395638 -0.3226073039 -2.4755086777 -0.3566776831
#>  [966] -0.3341078714  0.1071921292  0.0087489591  1.6445592470  2.9087924207
#>  [971]  0.1835144695  1.3658229688  3.6838032445  0.1505619895  3.1809565721
#>  [976] -1.3861210030  0.1426999686  0.6049765727  1.2126833373  0.0277963495
#>  [981]  0.3988278240 -0.9031235696 -1.2628238834  0.9101610053  0.9399794545
#>  [986] -0.5100953655  1.9369317763 -0.7369581764  2.0034649013  0.3769742966
#>  [991]  0.2805022367  0.7309967391 -0.1104087171 -0.2976518848  0.1994157075
#>  [996]  0.8874376645 -0.4331845237  2.6358869223  0.0288801725 -0.3830656717
#> 
#> $summary
#>        location scale
#> input      0.00  1.00
#> sample     0.04  0.99
#> 

generate_distribution(distribution = "gamma",
                      alpha_shape = 3)
#> $data
#>    [1] 2.2767158 1.9304243 0.6447609 2.3435140 3.6577508 1.9808065 2.3012444
#>    [8] 2.1714953 2.6619022 3.1297294 4.4334801 1.9986071 4.6629802 2.3032051
#>   [15] 2.1318950 4.6187395 1.4971472 2.8249837 2.4548324 4.2770270 2.9084193
#>   [22] 3.4757942 3.5087011 0.4366386 0.9559139 1.1239945 6.1674093 4.1661068
#>   [29] 1.4377552 0.9181054 3.9208652 1.2191646 4.8798310 4.7590980 1.9006763
#>   [36] 3.0992719 7.8408779 1.4563305 0.9075892 3.1988826 2.9103552 2.3223106
#>   [43] 1.7840253 1.0347861 1.6519726 5.6815658 1.2772517 3.9922750 2.0998008
#>   [50] 6.0119951 1.0995889 2.8441265 1.8495763 3.6780148 3.0865405 3.9118922
#>   [57] 6.4200413 0.6114179 0.9301245 1.4858916 3.4250945 2.4608119 1.4415606
#>   [64] 1.5575741 3.0420832 4.7384355 2.1677770 1.4026021 1.1587331 2.3414865
#>   [71] 1.5207662 2.8452287 2.8836630 4.0025794 0.6406239 3.6691334 4.4314137
#>   [78] 1.7494702 2.0493317 2.1301954 2.6936776 5.8674436 1.8235272 2.1057946
#>   [85] 6.4297393 1.3537983 4.2341646 1.8834599 1.2930051 7.5921676 3.9281871
#>   [92] 4.0882203 0.9922443 2.0483418 3.7309182 2.3916444 1.9530579 3.7592821
#>   [99] 3.4984314 3.3512330 4.9205170 1.5997155 2.9354718 1.0254213 4.9258357
#>  [106] 1.4053876 4.6985670 3.6975343 2.8251785 0.7486631 3.7557858 1.6725750
#>  [113] 4.0388444 3.5817276 4.1886959 2.1064598 3.5357880 3.9642354 3.6796540
#>  [120] 1.6031126 2.5673750 9.1422049 3.5022129 2.8847428 0.9158932 3.4056130
#>  [127] 0.5617862 2.4188992 1.9904288 1.0613183 1.3809265 2.1718180 1.8506033
#>  [134] 4.3515987 1.1368529 2.4449131 3.6200243 1.4972197 1.6419865 3.1699864
#>  [141] 1.3034783 1.6614050 3.1716891 0.5921338 5.6285613 0.6705353 2.1106361
#>  [148] 5.1246889 2.9321586 1.5202371 1.8739973 3.0169182 4.3260041 5.7744097
#>  [155] 4.8999385 2.4896875 1.5031872 1.3674657 1.3126851 3.1258147 3.0698600
#>  [162] 1.0201202 8.5102503 2.9536985 0.5914796 3.5147704 0.9980626 6.8658697
#>  [169] 3.4134022 0.9139375 3.2018868 1.8828139 5.2259497 4.4469853 3.6276099
#>  [176] 1.0048457 2.1756503 5.1754196 1.7962386 1.2184571 1.8368665 2.8921761
#>  [183] 4.4679389 2.4944835 1.8557498 3.2050882 2.3185556 2.5442110 1.5521811
#>  [190] 3.7950207 2.4007651 2.3773563 4.1612024 2.2804883 3.5493577 4.2661275
#>  [197] 1.6745448 1.4142402 2.4436726 2.1236166 4.6626663 2.0412737 4.1052431
#>  [204] 2.1873679 5.7116509 1.6975595 2.2429170 3.6805489 1.5018633 3.9949787
#>  [211] 1.3244075 3.1991186 1.2624309 3.1181522 2.6451972 1.4737550 1.7367610
#>  [218] 1.4090199 0.8203262 3.1511895 1.7214365 1.2760058 2.8970000 1.7223733
#>  [225] 3.9657055 1.5547884 2.8413781 6.9252108 0.7311365 3.5926845 2.2178055
#>  [232] 0.6114648 1.3632211 0.8957224 0.8365642 1.4903679 3.3334429 3.7925241
#>  [239] 3.9853103 2.2721609 3.4014963 1.7558932 4.9948342 2.1126809 4.6544789
#>  [246] 3.2259475 2.3868247 3.6048436 5.2355742 3.7310377 1.8581189 1.8882025
#>  [253] 2.5352259 1.1759798 2.8108804 2.2316259 2.5431509 6.7558405 4.4847846
#>  [260] 4.7955157 2.0093074 5.0183501 3.6326547 3.0656022 0.7374370 2.0746689
#>  [267] 5.5178338 3.6063334 4.6154757 2.4422491 6.3215745 2.0808979 3.7007097
#>  [274] 2.3409106 3.7104966 1.4475602 4.3575577 3.0622718 0.9377085 2.6528331
#>  [281] 1.2915892 7.4531832 4.5216895 2.8770856 7.9351119 2.4235937 2.8064979
#>  [288] 2.3921318 4.4134602 3.6022802 0.9048058 3.3622938 3.3390930 0.9729677
#>  [295] 1.3822990 3.3879942 2.9163764 1.8956443 0.7076483 2.9640691 6.0587414
#>  [302] 1.6803403 3.2795575 6.6919172 3.6504701 2.6176017 2.9357533 3.3999399
#>  [309] 5.1176638 0.6576248 2.4477303 3.3582827 1.2221393 3.3358601 3.7546031
#>  [316] 2.1496051 1.1511894 1.7026383 1.3792340 0.6861291 4.0876700 0.9123238
#>  [323] 3.2223858 4.0969185 1.5462090 3.6532393 5.1484219 4.6163672 7.1179285
#>  [330] 1.5155123 5.5697287 1.2664166 6.9584744 1.7616046 1.8762896 3.2788256
#>  [337] 1.2216185 8.2454409 3.5158087 3.0809668 4.2378773 2.0486645 1.9105662
#>  [344] 3.1897883 4.8513179 2.3614906 2.0269784 3.2522691 1.0264754 4.1569647
#>  [351] 1.9612660 1.0751638 3.1557221 1.1490329 4.3530217 3.6421754 2.4259246
#>  [358] 1.6791204 4.5878653 2.7514988 1.9385851 4.1331922 2.3190606 3.6522102
#>  [365] 1.5471839 1.7120962 5.5660668 5.5865762 1.2017371 2.4470156 6.3297105
#>  [372] 2.5192428 1.4979845 5.5025024 4.6891551 1.4977117 0.8436375 1.3039281
#>  [379] 0.9588180 3.3692248 3.1259496 2.7907500 0.9624828 1.9796283 1.4534783
#>  [386] 0.8793458 2.5290155 5.2216127 4.7234665 3.1033720 6.9820555 3.2222058
#>  [393] 4.0172049 1.6475472 3.5037845 3.2090957 2.2607011 0.5573524 4.0960181
#>  [400] 4.8655872 6.3635676 5.5782803 2.8895474 4.3285821 1.9650699 0.8306138
#>  [407] 3.5534055 2.0455067 1.3825601 4.3688680 4.5565292 4.8853722 9.5086369
#>  [414] 1.2935444 2.0482223 1.0844030 4.6911371 2.2214866 3.3530107 1.6142241
#>  [421] 2.1939994 3.2448519 1.5673707 4.1378860 1.7873535 2.0598232 1.6613779
#>  [428] 0.4971407 3.4189690 0.7136461 2.2712468 1.5532374 0.5105254 2.2498660
#>  [435] 1.6268725 1.8961405 1.5070496 1.7530196 2.1780950 2.3193423 1.9332164
#>  [442] 2.4576139 3.3477276 4.7187087 2.3515667 1.7429875 5.2913510 3.6780223
#>  [449] 1.3925542 2.1637944 2.3616513 1.3896252 2.6102332 1.1616361 2.4392273
#>  [456] 0.6749743 2.5258140 2.1885860 4.4033989 1.7661149 2.4307449 2.6980386
#>  [463] 6.5091404 2.7938384 2.0434719 3.8611044 4.5288668 1.8099404 3.6436345
#>  [470] 2.7057250 2.9740134 3.0801344 2.0111767 3.7236997 1.7011890 1.0433581
#>  [477] 1.6042703 2.2572131 4.8920699 8.5340771 5.4684484 2.9782106 2.0556083
#>  [484] 1.4614236 3.5346096 2.4204447 1.6171723 5.3079117 2.2216556 0.9078282
#>  [491] 2.8167592 2.0406571 2.4932861 3.4481484 1.9287942 1.9955140 4.4434482
#>  [498] 2.3722318 1.5410026 3.5996507 3.7964981 1.4465864 3.5888277 2.6445667
#>  [505] 1.2013826 0.7273593 4.8069228 6.4884802 3.1303331 3.1022111 3.1808168
#>  [512] 2.7469840 2.1314439 3.2511674 1.3329075 1.2268198 7.7762647 3.5375367
#>  [519] 1.8384520 1.4329188 2.9536388 3.0176377 1.9977448 1.2612572 3.2363975
#>  [526] 1.4218378 1.8080711 1.6590129 3.6937906 2.2473440 3.2840634 2.6471612
#>  [533] 4.2630246 3.6325008 1.7793598 1.4154422 3.9587369 1.0557357 0.4170193
#>  [540] 1.4624754 3.9496873 1.6171807 4.9269180 6.3938632 3.7552404 5.2885010
#>  [547] 3.1902000 1.6342452 2.8577780 3.6252620 2.5634226 6.2681084 1.4335362
#>  [554] 1.5343004 3.1812078 1.5463501 1.8920463 0.7942247 4.8064204 5.8402249
#>  [561] 2.2021955 7.2858167 3.1196905 1.2341175 4.1302534 1.0586429 1.9495670
#>  [568] 1.8495571 4.2402299 3.8261583 1.3869706 2.9692535 5.3834519 1.1170748
#>  [575] 3.2732924 5.4260262 3.4858598 1.9327315 1.2398297 3.6573692 0.3636196
#>  [582] 1.7229053 1.3832557 5.5483086 7.4790223 2.2146120 4.0174977 3.9973804
#>  [589] 4.4590450 1.9167913 2.9724904 1.9866060 2.2845225 4.4761410 2.7677281
#>  [596] 4.8523739 0.7323897 6.5417013 1.4755465 3.9432740 2.3023164 6.1798606
#>  [603] 3.4500490 5.1223539 2.5159732 1.8943515 3.1129281 1.7956933 7.3390186
#>  [610] 2.1969074 4.7302959 7.8409908 1.4558142 1.4380886 0.9392420 2.6166245
#>  [617] 1.1571670 1.0822498 3.8050547 3.4067521 3.4702045 6.1982229 4.3199017
#>  [624] 2.0759266 1.8469291 2.5112877 3.7777791 6.1366437 2.3602515 1.0400956
#>  [631] 1.3164519 1.4430558 1.7599103 2.0493645 2.9498720 1.2978266 2.8623467
#>  [638] 2.1362407 0.9925647 1.4258341 1.8140145 6.5619191 2.5200253 8.1017791
#>  [645] 4.5440946 1.7764580 4.4007816 0.6187638 2.3620804 5.7371071 4.8506792
#>  [652] 5.0669244 6.0357109 3.7224884 5.3201594 2.9841271 1.8880435 1.1568905
#>  [659] 4.9141423 3.5199619 3.9476839 4.9862381 4.1983461 1.0258210 3.9215126
#>  [666] 1.9797885 2.1603059 1.0284473 4.3838570 0.9627353 1.9112734 1.2803915
#>  [673] 1.2694184 1.9225441 1.8407986 7.1639493 1.1306215 3.2787859 6.8452226
#>  [680] 4.9418658 2.0211329 3.4869708 5.5307107 2.9349834 4.5034929 1.8960824
#>  [687] 3.8635674 2.4135900 1.8138002 3.0895570 2.1105119 1.3704221 5.8453095
#>  [694] 2.0708424 5.0779635 4.4762118 3.9844230 4.2592843 3.4502770 2.7950741
#>  [701] 3.3779946 4.2216550 2.2583346 0.8196429 2.3409331 5.0428776 4.3175286
#>  [708] 5.0967431 1.4469266 3.3164573 4.0464518 0.8277260 2.9239044 2.7313850
#>  [715] 3.6906152 2.5897951 6.3379246 3.9641702 2.7201839 3.1827147 5.3662308
#>  [722] 5.2716084 2.2121922 1.0170654 3.9516147 1.0559712 2.1960319 1.0525839
#>  [729] 2.9290761 2.9700395 3.3107330 2.4759345 3.1292588 4.7571073 5.1631699
#>  [736] 9.8730064 3.0569746 1.3805112 1.9153838 3.1884786 3.5916556 2.4693922
#>  [743] 2.1792706 4.0676323 0.6362322 1.9831764 3.1005652 1.5072184 2.3527921
#>  [750] 1.4156803 4.0460543 1.4407811 5.5351255 6.7544381 2.6325963 2.1221565
#>  [757] 3.3521829 1.0133924 3.3454528 6.2757897 4.7060385 3.0484653 3.4770323
#>  [764] 1.1788433 3.6524882 5.4370859 1.8009611 3.1193663 0.6269011 0.6361991
#>  [771] 2.8634051 2.9850688 1.9452124 4.0008943 2.4247773 1.2408420 2.1285927
#>  [778] 1.8895929 3.0573853 3.3121458 5.0021514 2.5225871 4.3954564 3.7098815
#>  [785] 0.6506453 1.8391066 3.2692277 5.1269951 3.4774280 2.5369065 2.5007890
#>  [792] 1.5885773 1.9604413 5.5542682 0.4473006 4.3477409 1.1084007 3.5437426
#>  [799] 1.6768580 2.1133887 0.5331009 1.3774777 2.3572226 1.3428488 3.1703229
#>  [806] 1.8071387 2.5881065 1.5824014 0.4164160 1.9720681 0.7465863 6.4553800
#>  [813] 0.6530973 2.1802174 1.8230100 3.2438867 1.7551143 0.9279966 5.7149362
#>  [820] 1.0338859 2.3423735 4.8288995 3.7316456 2.8124447 1.8335661 4.8348595
#>  [827] 2.4313094 3.6827872 2.5538495 2.3422152 3.8760587 3.0497846 1.6188510
#>  [834] 3.0548672 2.1309439 3.2998828 3.5437254 0.9478614 2.0194885 4.6630209
#>  [841] 1.2592649 1.8096141 0.8199782 3.0873280 2.6896098 2.6127534 1.1170918
#>  [848] 3.8279988 2.1671165 4.6817582 3.6270002 2.1652035 1.6117543 3.1331609
#>  [855] 2.2095712 3.4513007 1.6645842 1.8615085 3.0786481 3.1115725 6.1399171
#>  [862] 5.1299015 1.6659027 4.5895871 1.9919497 1.5440833 1.7592995 4.2457910
#>  [869] 1.4710044 4.5091194 1.3250747 2.5530219 0.5705311 1.9601932 4.4861277
#>  [876] 0.8886172 3.6746970 2.4407806 1.8208970 5.0366813 2.0469737 2.2019134
#>  [883] 1.3324037 2.5524973 1.9250826 6.4911806 2.9040047 1.6988899 2.7285598
#>  [890] 1.8092552 2.8084471 4.4325461 2.1882024 1.7444863 1.1362189 5.0859791
#>  [897] 6.5690630 2.2461716 3.3594266 7.7704914 5.6276846 2.3525012 2.3235855
#>  [904] 0.2850906 5.2655682 1.4543694 3.5471590 0.8822131 2.3751016 3.1156263
#>  [911] 6.2327645 2.0478710 0.9736861 5.6826848 7.6267141 5.5492431 2.9101278
#>  [918] 1.2302170 1.9428579 3.9698335 2.8032315 4.1591250 4.4393395 5.5244260
#>  [925] 3.8803361 1.6605067 7.8723190 2.1696219 1.3350167 0.6539002 1.2602814
#>  [932] 0.8921835 1.2726066 5.6230577 4.6415888 2.7530668 3.9177866 2.4207303
#>  [939] 0.4756703 7.2641549 4.2012456 5.9711483 1.4476289 1.1736510 6.8993672
#>  [946] 2.2458747 2.0110439 5.6652990 1.4037399 2.2971358 2.6231477 0.8259467
#>  [953] 5.1628494 2.6023437 4.4424539 2.8904850 1.8852836 1.9376463 6.0991458
#>  [960] 5.9014911 1.9689585 0.6473473 0.5233565 2.9523983 4.8260858 2.2861505
#>  [967] 2.2131908 3.8061292 4.1017468 2.7761478 0.6896127 2.8945121 2.6970918
#>  [974] 1.3383361 2.4173736 1.8777249 1.5814355 1.9205557 1.7788368 5.3097854
#>  [981] 1.7533857 2.2454552 2.3745951 5.7731954 2.9010823 4.7226186 5.4089400
#>  [988] 2.1167774 4.2911773 2.6423631 4.3385606 3.6350305 3.6697230 1.6291111
#>  [995] 2.2288476 2.2642223 3.8951847 5.4274376 3.2379631 3.1850318
#> 
#> $summary
#>        shape rate
#> input   3.00 1.00
#> sample  3.19 1.08
#> 

generate_distribution(distribution = "beta",
                      alpha_shape = 2,
                      beta_shape = 2)
#> $data
#>    [1] 0.34676768 0.55773876 0.65253134 0.93546206 0.20050519 0.81241450
#>    [7] 0.48232761 0.91636219 0.90612064 0.17412134 0.54320923 0.17916779
#>   [13] 0.49584945 0.54811195 0.49432164 0.24582931 0.64270638 0.86738291
#>   [19] 0.39739937 0.20654316 0.42297482 0.60195421 0.21063596 0.49379776
#>   [25] 0.45824446 0.85220042 0.38489352 0.44535294 0.36467528 0.31282751
#>   [31] 0.76936615 0.50346694 0.43354423 0.67656871 0.49427589 0.47058870
#>   [37] 0.68669132 0.67236010 0.68580079 0.15775015 0.67654015 0.36685900
#>   [43] 0.31531884 0.68229995 0.20499308 0.76541774 0.29977505 0.27033807
#>   [49] 0.68930566 0.13053580 0.35697729 0.24988746 0.52762690 0.24310210
#>   [55] 0.60397918 0.55130820 0.19665283 0.67155131 0.82033841 0.52673455
#>   [61] 0.27491776 0.73859403 0.21793774 0.49991788 0.48412941 0.59512201
#>   [67] 0.35880831 0.31229722 0.84857075 0.29828767 0.63381513 0.71940887
#>   [73] 0.29641759 0.12653773 0.29691451 0.29695827 0.86644666 0.21499830
#>   [79] 0.27654997 0.40350905 0.67169518 0.96952183 0.19844371 0.69591567
#>   [85] 0.82749170 0.35705258 0.96924600 0.61703588 0.14437430 0.89534457
#>   [91] 0.51510276 0.69340940 0.59739832 0.58453765 0.17891111 0.68657313
#>   [97] 0.40423114 0.26562242 0.21903359 0.43197902 0.45545094 0.27624119
#>  [103] 0.54251722 0.61461715 0.20233034 0.37067898 0.32238706 0.20275025
#>  [109] 0.62995273 0.09510210 0.38478501 0.61255364 0.40978764 0.64715086
#>  [115] 0.60782627 0.38834187 0.55193075 0.50454980 0.29135395 0.35948216
#>  [121] 0.22423738 0.09089519 0.75232674 0.49448478 0.61752153 0.18006841
#>  [127] 0.81035313 0.82718916 0.97477217 0.60456098 0.14367917 0.56015530
#>  [133] 0.78657966 0.34666161 0.49726163 0.24544179 0.80202556 0.63444686
#>  [139] 0.53237917 0.36713359 0.60595968 0.20728160 0.28357624 0.51647707
#>  [145] 0.16443940 0.15965335 0.77054650 0.46569583 0.81541918 0.75684006
#>  [151] 0.29681941 0.42425506 0.90675620 0.30809099 0.38461694 0.39039744
#>  [157] 0.87298641 0.73232190 0.46293208 0.72631421 0.59388137 0.86961375
#>  [163] 0.73844842 0.32460152 0.17431580 0.81694149 0.75808471 0.88183601
#>  [169] 0.27629642 0.16870091 0.63380992 0.48059335 0.59637150 0.23276440
#>  [175] 0.62076007 0.38621072 0.47684805 0.58640522 0.57591241 0.51709763
#>  [181] 0.54820418 0.61538920 0.32024927 0.53246153 0.81671613 0.55300136
#>  [187] 0.38008596 0.16132387 0.19788269 0.38622439 0.04233332 0.62892814
#>  [193] 0.64598127 0.49109592 0.91605430 0.24279605 0.38672888 0.69520656
#>  [199] 0.95215807 0.51127117 0.73335308 0.31811951 0.24416509 0.33629199
#>  [205] 0.06711404 0.48888797 0.65254015 0.58464389 0.32278757 0.91817917
#>  [211] 0.19897957 0.80135520 0.12235901 0.24250712 0.38783305 0.75083352
#>  [217] 0.73476378 0.14850306 0.50949966 0.73655983 0.48587054 0.54755646
#>  [223] 0.33551265 0.78576800 0.59878011 0.82977122 0.25950450 0.83867376
#>  [229] 0.38182808 0.49070759 0.57704528 0.70650093 0.23646725 0.60810199
#>  [235] 0.76965049 0.07182849 0.44678921 0.51087399 0.38786832 0.76091511
#>  [241] 0.38105947 0.43268321 0.54799679 0.20236633 0.37637155 0.44209302
#>  [247] 0.65443843 0.44520956 0.33990797 0.56890993 0.71340244 0.62736211
#>  [253] 0.56015614 0.71758115 0.35137924 0.40716536 0.19963820 0.77626349
#>  [259] 0.53063600 0.76588024 0.86989331 0.73812467 0.60074454 0.78215275
#>  [265] 0.31924225 0.31851047 0.16964514 0.42021040 0.12789257 0.94064168
#>  [271] 0.67330185 0.23340975 0.10653308 0.72805906 0.35370269 0.68875135
#>  [277] 0.89352732 0.34681729 0.86535597 0.69471840 0.98910346 0.42065962
#>  [283] 0.61726330 0.91943526 0.76705601 0.78964088 0.53650939 0.63996312
#>  [289] 0.70310803 0.43590673 0.80090870 0.10211663 0.13913523 0.25735592
#>  [295] 0.64166867 0.57743933 0.52340204 0.39480152 0.20824918 0.48073177
#>  [301] 0.80437656 0.66905133 0.57910480 0.37710501 0.93798009 0.30229738
#>  [307] 0.69560090 0.34350581 0.43188506 0.56329617 0.66604969 0.44148536
#>  [313] 0.53786625 0.74110925 0.27055966 0.77498796 0.96699490 0.83458341
#>  [319] 0.59839774 0.53600626 0.53044713 0.54904402 0.12781823 0.06272549
#>  [325] 0.38249576 0.33665857 0.17798508 0.83683452 0.44027094 0.64634629
#>  [331] 0.46485783 0.82399224 0.78050407 0.05090139 0.70493709 0.72663242
#>  [337] 0.56504546 0.62565166 0.39641376 0.34636902 0.26928976 0.26112164
#>  [343] 0.54083563 0.52734866 0.32471826 0.70338708 0.76912328 0.56772397
#>  [349] 0.55907624 0.93203965 0.46013076 0.14186974 0.52162308 0.32769305
#>  [355] 0.19935726 0.30002638 0.46488284 0.20704867 0.63632029 0.89256921
#>  [361] 0.32363796 0.67708087 0.82610455 0.41474130 0.30398663 0.37577169
#>  [367] 0.20995955 0.44966032 0.71616479 0.68077741 0.58875192 0.22169476
#>  [373] 0.35294154 0.77564259 0.12393953 0.23495350 0.89898217 0.08860960
#>  [379] 0.45247746 0.53476841 0.75618066 0.43895521 0.23901179 0.70728369
#>  [385] 0.27572722 0.71951452 0.87612327 0.34302465 0.13035489 0.30591915
#>  [391] 0.41881643 0.34649907 0.63116646 0.42853363 0.61452831 0.21765029
#>  [397] 0.74347472 0.44109913 0.61448242 0.15570209 0.43357511 0.97327186
#>  [403] 0.21751575 0.18479952 0.28254972 0.10012128 0.43252805 0.33136843
#>  [409] 0.70149393 0.55462196 0.61299127 0.39525112 0.25558875 0.23358776
#>  [415] 0.07980521 0.07016046 0.51757604 0.37805935 0.65147767 0.41055107
#>  [421] 0.88808995 0.57108330 0.29132473 0.22907693 0.64461913 0.29203035
#>  [427] 0.17633941 0.71100561 0.25542563 0.77750236 0.55890797 0.78457763
#>  [433] 0.30218670 0.65064311 0.82641320 0.57419317 0.39517389 0.46219444
#>  [439] 0.63422324 0.92540236 0.10287516 0.10127448 0.63801869 0.63811190
#>  [445] 0.24756843 0.64939806 0.63545478 0.83390276 0.32843242 0.87404051
#>  [451] 0.51257713 0.69687368 0.67798564 0.58646946 0.63393638 0.66788744
#>  [457] 0.32363730 0.12259366 0.63659196 0.48981161 0.43246556 0.28598148
#>  [463] 0.37070351 0.49730124 0.83804520 0.73584978 0.53675079 0.52135389
#>  [469] 0.80722347 0.91334516 0.53759557 0.20613696 0.16857455 0.66602614
#>  [475] 0.70634300 0.29138098 0.34160523 0.84074777 0.38738173 0.75328837
#>  [481] 0.39929981 0.95551824 0.18436908 0.35312577 0.83663859 0.17721362
#>  [487] 0.64047106 0.12702609 0.81604098 0.14458665 0.66815686 0.57519830
#>  [493] 0.36474172 0.95689936 0.65174177 0.52143697 0.52398509 0.18818675
#>  [499] 0.44026058 0.13175684 0.77020086 0.63037965 0.71567018 0.66807587
#>  [505] 0.43755422 0.70400896 0.41235742 0.20215177 0.36845149 0.61822541
#>  [511] 0.44598893 0.32976324 0.57643185 0.15945795 0.61808141 0.38279790
#>  [517] 0.62590952 0.40554528 0.14779991 0.85901057 0.09346964 0.62831981
#>  [523] 0.50088058 0.52582730 0.32732053 0.48528063 0.82534106 0.48528141
#>  [529] 0.61121197 0.32892660 0.47283953 0.32016065 0.22346830 0.75651366
#>  [535] 0.66221301 0.46661850 0.29025996 0.60508792 0.50058662 0.65058320
#>  [541] 0.78308711 0.51665629 0.95143290 0.36285982 0.33867713 0.33536155
#>  [547] 0.63838782 0.55232315 0.92621043 0.86410876 0.06317120 0.50583685
#>  [553] 0.07804367 0.29265154 0.40982555 0.53529093 0.39482299 0.39814355
#>  [559] 0.87458589 0.74365776 0.92384311 0.53083531 0.49534948 0.45467783
#>  [565] 0.79794230 0.49704539 0.72514709 0.15941833 0.07694927 0.19388252
#>  [571] 0.35115042 0.66282950 0.55639244 0.40427436 0.46474343 0.10959186
#>  [577] 0.27175768 0.75294041 0.45628142 0.18062374 0.80762755 0.51828697
#>  [583] 0.20004940 0.58631840 0.56194942 0.89499142 0.65400822 0.59228860
#>  [589] 0.69285638 0.13712949 0.61270733 0.49476383 0.26911066 0.67013332
#>  [595] 0.72566488 0.37076694 0.18244430 0.37660493 0.16418903 0.44749962
#>  [601] 0.76305086 0.24750040 0.47411762 0.13138538 0.43126756 0.63166488
#>  [607] 0.32737069 0.67257015 0.34230186 0.34199412 0.74392278 0.15231791
#>  [613] 0.43662513 0.30989526 0.38058372 0.56849963 0.83573004 0.39982113
#>  [619] 0.76932508 0.45052753 0.46601797 0.20622982 0.10972713 0.85900141
#>  [625] 0.75897467 0.41697861 0.86955799 0.37325482 0.39981672 0.74593607
#>  [631] 0.05304303 0.86529639 0.12836525 0.92472695 0.49708371 0.30207120
#>  [637] 0.32280565 0.82797554 0.66537787 0.56451945 0.79999003 0.59750021
#>  [643] 0.90976154 0.17615095 0.58376325 0.60632357 0.59367954 0.49788602
#>  [649] 0.40546903 0.28860393 0.70892088 0.34368248 0.44122459 0.42814870
#>  [655] 0.33078596 0.78480922 0.58357416 0.25272912 0.87829094 0.59468863
#>  [661] 0.42346182 0.43261882 0.84778874 0.66140926 0.24802366 0.71933822
#>  [667] 0.42301027 0.28902937 0.36289251 0.59932741 0.30320372 0.40178714
#>  [673] 0.49064659 0.48249480 0.33923311 0.44516946 0.23695295 0.66525497
#>  [679] 0.09965049 0.73576890 0.59977154 0.44206998 0.54283788 0.30679743
#>  [685] 0.58474753 0.51103893 0.23800297 0.90290243 0.13556366 0.61279416
#>  [691] 0.15587997 0.44333925 0.34793873 0.37382262 0.58919849 0.60718143
#>  [697] 0.70510961 0.35682030 0.17752117 0.71806844 0.27129500 0.58422370
#>  [703] 0.08721574 0.35995625 0.11958000 0.14569825 0.51573568 0.76088710
#>  [709] 0.51915433 0.23327793 0.33013124 0.07293955 0.72524265 0.35016390
#>  [715] 0.45946034 0.18859599 0.23039877 0.66908869 0.36545330 0.13562065
#>  [721] 0.53784085 0.13515239 0.92423296 0.44837675 0.39111432 0.94342377
#>  [727] 0.42512571 0.34289097 0.20710953 0.91498922 0.35812608 0.06880618
#>  [733] 0.96934995 0.36997753 0.43131456 0.40681750 0.25193514 0.40041858
#>  [739] 0.89477300 0.70048102 0.05260197 0.41719633 0.57094776 0.43232952
#>  [745] 0.21979627 0.81775934 0.59943671 0.35112280 0.48837109 0.74415368
#>  [751] 0.45871659 0.79736886 0.47370084 0.89857380 0.10673219 0.66163370
#>  [757] 0.84682111 0.32994347 0.81799448 0.57662736 0.51098023 0.29865327
#>  [763] 0.63194633 0.65855981 0.81285521 0.68140441 0.47722849 0.29165676
#>  [769] 0.37962739 0.45784759 0.61207521 0.47378135 0.83186127 0.60206244
#>  [775] 0.34865044 0.87340897 0.71055317 0.61308244 0.77898338 0.28622931
#>  [781] 0.72193112 0.07591445 0.36369376 0.73884099 0.22944643 0.48148833
#>  [787] 0.10898510 0.80688855 0.39378410 0.34877710 0.61426565 0.69837690
#>  [793] 0.85400509 0.26762836 0.46909705 0.30459947 0.56284985 0.15630522
#>  [799] 0.81052872 0.23269973 0.43411446 0.79605526 0.52601214 0.48540754
#>  [805] 0.66754717 0.70011927 0.85671238 0.40968863 0.04888457 0.46417805
#>  [811] 0.61871200 0.30073826 0.41120567 0.60199691 0.53345177 0.89203465
#>  [817] 0.09825772 0.67258527 0.08921818 0.74369945 0.59624281 0.62432845
#>  [823] 0.60964044 0.93677829 0.59723535 0.31199268 0.73766633 0.89734893
#>  [829] 0.39385414 0.58490111 0.52073146 0.19402454 0.50768688 0.31650696
#>  [835] 0.81380730 0.66359911 0.83108197 0.66348050 0.79558837 0.80992684
#>  [841] 0.52259320 0.12338042 0.65555377 0.52029024 0.77315634 0.65543052
#>  [847] 0.30575856 0.37861349 0.70814282 0.72844120 0.07369017 0.82001383
#>  [853] 0.43346366 0.28161406 0.76922247 0.10953811 0.56005331 0.37862862
#>  [859] 0.73671697 0.19526504 0.46954159 0.38976949 0.27956917 0.70371351
#>  [865] 0.83765415 0.56517595 0.54229662 0.79177922 0.19043624 0.55475780
#>  [871] 0.66512431 0.83735497 0.36829960 0.27655219 0.52298215 0.88351821
#>  [877] 0.71238892 0.56287834 0.32021299 0.80078299 0.33145931 0.24967164
#>  [883] 0.90070280 0.64148717 0.79215750 0.43068898 0.60854959 0.45753481
#>  [889] 0.44746021 0.17454757 0.34346298 0.55479560 0.49720000 0.50079916
#>  [895] 0.71244301 0.86047125 0.67226163 0.60551354 0.56090477 0.75161774
#>  [901] 0.17375823 0.27470049 0.76805913 0.52658694 0.05491532 0.43459279
#>  [907] 0.66997439 0.23439051 0.56308305 0.47368825 0.23549977 0.41393842
#>  [913] 0.60236297 0.61306862 0.71323894 0.29368837 0.57443490 0.54269118
#>  [919] 0.61579048 0.72718516 0.63207225 0.43209285 0.58150180 0.35230725
#>  [925] 0.79163432 0.54853593 0.90018769 0.65131711 0.46993319 0.49144317
#>  [931] 0.31732205 0.79080126 0.14161303 0.41698331 0.24805720 0.31637317
#>  [937] 0.85265923 0.94303942 0.78949348 0.64299435 0.16919417 0.81625416
#>  [943] 0.52579418 0.14390899 0.69120420 0.49212931 0.54223613 0.54658048
#>  [949] 0.76288155 0.22685151 0.30667484 0.54667589 0.30728955 0.26104599
#>  [955] 0.69489769 0.46296539 0.39261835 0.25055345 0.37525919 0.59505593
#>  [961] 0.50913339 0.73150189 0.72625187 0.94140396 0.33897957 0.22455063
#>  [967] 0.41612950 0.70550356 0.60531617 0.31447868 0.67370023 0.49598304
#>  [973] 0.49150394 0.66139444 0.34218230 0.29740437 0.26185992 0.64042157
#>  [979] 0.51782997 0.49470577 0.33386890 0.62167556 0.54282490 0.05937977
#>  [985] 0.16056850 0.19804084 0.35965069 0.70193302 0.25392384 0.20120883
#>  [991] 0.85475671 0.68469739 0.24719620 0.28960070 0.08375970 0.39287588
#>  [997] 0.13011098 0.24170923 0.48570018 0.49012943
#> 
#> $summary
#>        alpha shape beta shape
#> input         2.00        2.0
#> sample        1.92        1.9
#>