Generate samples of quantitative matched pair data with correlation and optional standardized effect

generate_matched() is a function to generate one-group samples of quantitative matched pair data with specified first four moments, correlation, and optional standardized effect size

generate_matched(
  n = 100,
  mean = 0,
  var = 1,
  skew = 0,
  kurt = 0,
  effect_size = 0.5,
  r = 0.5,
  use_effect_size = TRUE,
  summary = TRUE,
  decimals = 3,
  replication = 1
)

Arguments

n

A numeric vector for the sample size. The default value is 100.

mean

A numeric vector for the population mean of the first variable. The default value is 0.

var

A numeric vector for the population variance of the first variable. The default value is 1.

skew

A numeric vector for the population skew of the variables. Separate values of skew can be entered. The default value is 0.

kurt

A numeric vector for the population kurtosis of the variables. Separate values of kurtosis can be entered. The default value is 0.

effect_size

An optional numeric vector for the standardized effect size between groups of data.

r

A numeric vector for the correlation between the variables. The default value is 0.5.

use_effect_size

A logical vector to indicate whether to use a standardized effect size in generating data for k when k >= 2. The default value is FALSE.

summary

A logical vector for whether the return should include summary statistics. The default value is TRUE.

decimals

A numeric vector for the number of decimals to round the sample data to. The default value is 3.

replication

A numeric vector for the number of times to replicate the sampling. The default value is 1.

Value

If summary = TRUE, a list containing a vector ("correlation") with the correlation between the pairs, a matrix ("summary") of the input parameters and descriptive statistics, and a data frame ("sample") with n rows and 3 columns (values from each set of the pair and the difference of X2 and X1).

If summary = FALSE, a data frame with n rows and 3 columns.

Details

This function is an extension of generate_quantitative() that generates matched pair data based on specified mean, variance, skew, and kurtosis parameters. The user can also specify a standardized effect size difference between the two variables as well as the correlation between them. The data is sampled from the specified proportions as a parameter, so the output statistics will not be an exact match to the parameters.

Because the output is a sample, the summary argument allows for summary statistics to be generated with the sample data. This allows the user to easily compare the simulation statistics to the specified parameters.

This is useful for teaching about paired t tests, confounding variables, and conditions for valid inference.

Examples

generate_matched()
#> $correlation
#> [1] 0.47
#> 
#> $summary
#>            mean   sd  skew kurtosis
#> x input    0.00 1.00  0.00     0.00
#> sample x   0.06 1.01  0.28    -0.32
#> y input    0.50 1.00  0.00     0.00
#> sample y   0.48 0.99 -0.04    -0.50
#> difference 0.42 1.03 -0.18    -0.23
#> 
#> $sample
#>         X1     X2 Difference
#> 1    0.782  1.667      0.885
#> 2    0.272  0.745      0.473
#> 3    0.110  1.134      1.024
#> 4   -1.365 -0.220      1.145
#> 5   -0.628 -0.248      0.380
#> 6   -0.424  1.688      2.112
#> 7    0.418  1.823      1.405
#> 8   -0.726  0.741      1.467
#> 9    1.096  0.504     -0.592
#> 10   0.153 -0.376     -0.529
#> 11   2.119  0.067     -2.052
#> 12  -1.423  1.452      2.875
#> 13  -0.813 -0.142      0.671
#> 14  -1.375 -0.448      0.927
#> 15   1.181 -0.178     -1.359
#> 16   0.651 -0.194     -0.845
#> 17   0.281  1.461      1.180
#> 18  -0.763 -0.722      0.041
#> 19   0.050  0.345      0.295
#> 20   0.980  1.201      0.221
#> 21  -0.689  0.365      1.054
#> 22   0.779  1.965      1.186
#> 23   0.349  1.644      1.295
#> 24  -0.420  0.856      1.276
#> 25   0.822  1.418      0.596
#> 26  -1.500 -1.812     -0.312
#> 27   2.719  1.330     -1.389
#> 28  -1.353  1.170      2.523
#> 29   2.671  0.388     -2.283
#> 30  -0.779 -0.990     -0.211
#> 31   0.587  0.096     -0.491
#> 32  -0.336  1.176      1.512
#> 33   0.726  1.823      1.097
#> 34   0.539 -1.212     -1.751
#> 35  -0.398  1.103      1.501
#> 36   0.144 -0.563     -0.707
#> 37  -1.496  0.445      1.941
#> 38   1.702  0.274     -1.428
#> 39   0.456  0.156     -0.300
#> 40  -0.193 -0.188      0.005
#> 41   0.431  0.089     -0.342
#> 42  -1.684 -1.515      0.169
#> 43   0.098  0.580      0.482
#> 44  -0.758  0.050      0.808
#> 45  -0.095  0.432      0.527
#> 46   0.917  0.223     -0.694
#> 47  -0.886  0.504      1.390
#> 48  -0.142  1.755      1.897
#> 49  -0.617 -0.103      0.514
#> 50   2.237  1.851     -0.386
#> 51  -0.652  0.726      1.378
#> 52  -1.133 -0.373      0.760
#> 53   0.434  1.054      0.620
#> 54  -1.337 -0.594      0.743
#> 55   0.750  2.382      1.632
#> 56   0.494  1.248      0.754
#> 57   0.453  0.642      0.189
#> 58  -0.001 -0.926     -0.925
#> 59   1.181  0.284     -0.897
#> 60  -0.357  0.088      0.445
#> 61  -1.017 -0.115      0.902
#> 62   0.649  0.423     -0.226
#> 63  -1.351 -1.181      0.170
#> 64  -1.195  0.261      1.456
#> 65   0.573  1.575      1.002
#> 66   1.717  0.636     -1.081
#> 67   1.115  1.173      0.058
#> 68  -1.245 -1.030      0.215
#> 69   1.308  1.572      0.264
#> 70   0.310  0.357      0.047
#> 71   1.424  0.187     -1.237
#> 72  -0.317 -0.215      0.102
#> 73   0.451  1.508      1.057
#> 74  -0.625  0.353      0.978
#> 75   0.924  1.266      0.342
#> 76   1.982  2.713      0.731
#> 77   0.040  0.640      0.600
#> 78  -1.392  1.060      2.452
#> 79   0.776  1.643      0.867
#> 80  -0.332  1.430      1.762
#> 81  -0.222  2.101      2.323
#> 82  -1.901 -0.254      1.647
#> 83  -0.734  0.315      1.049
#> 84  -0.437 -1.112     -0.675
#> 85  -0.742 -0.745     -0.003
#> 86  -0.231  1.136      1.367
#> 87  -0.658 -0.170      0.488
#> 88   0.453  1.465      1.012
#> 89   1.118  2.822      1.704
#> 90   0.701  0.174     -0.527
#> 91   1.175  1.493      0.318
#> 92   0.102  0.569      0.467
#> 93  -0.536 -0.851     -0.315
#> 94  -1.885 -0.273      1.612
#> 95  -0.571 -1.130     -0.559
#> 96   0.111 -0.659     -0.770
#> 97   0.810  1.000      0.190
#> 98   0.648  0.143     -0.505
#> 99   1.473  1.200     -0.273
#> 100 -0.862 -1.648     -0.786
#> 

generate_matched(use_effect_size = FALSE)
#> $correlation
#> [1] 0.496
#> 
#> $summary
#>             mean   sd  skew kurtosis
#> x input     0.00 1.00  0.00     0.00
#> sample x   -0.19 0.96 -0.29    -0.07
#> y input     0.00 1.00  0.00     0.00
#> sample y    0.14 1.07  0.25    -0.09
#> difference  0.32 1.03  0.27    -0.49
#> 
#> $sample
#>         X1     X2 Difference
#> 1   -0.364  0.333      0.697
#> 2   -0.537  0.372      0.909
#> 3    0.117 -0.166     -0.283
#> 4   -1.513 -2.745     -1.232
#> 5   -0.760 -1.000     -0.240
#> 6    0.642  1.135      0.493
#> 7   -1.720  0.006      1.726
#> 8    2.320  2.772      0.452
#> 9    0.351 -0.388     -0.739
#> 10  -1.638 -0.865      0.773
#> 11  -0.373 -0.151      0.222
#> 12  -1.084  0.123      1.207
#> 13  -0.538 -0.955     -0.417
#> 14  -0.598  0.430      1.028
#> 15   0.370  1.104      0.734
#> 16   0.346  1.475      1.129
#> 17   1.122  0.442     -0.680
#> 18   0.560 -0.629     -1.189
#> 19   1.072  0.513     -0.559
#> 20   0.983  0.956     -0.027
#> 21  -2.014 -0.828      1.186
#> 22   0.143 -0.757     -0.900
#> 23  -1.169 -1.566     -0.397
#> 24  -0.512  1.640      2.152
#> 25   0.401  1.324      0.923
#> 26  -0.251 -0.967     -0.716
#> 27  -1.004 -0.515      0.489
#> 28   0.395  0.157     -0.238
#> 29  -0.341  0.413      0.754
#> 30  -0.112  0.104      0.216
#> 31  -1.265 -0.237      1.028
#> 32   0.882  0.885      0.003
#> 33   0.155  1.376      1.221
#> 34   0.036  0.754      0.718
#> 35   0.031  0.512      0.481
#> 36  -0.523  1.697      2.220
#> 37  -0.729 -0.764     -0.035
#> 38  -1.572  0.308      1.880
#> 39   0.643  1.377      0.734
#> 40  -0.494 -0.965     -0.471
#> 41   0.817  0.844      0.027
#> 42   0.597 -0.798     -1.395
#> 43  -1.472 -0.794      0.678
#> 44   1.969  0.978     -0.991
#> 45  -0.388 -1.179     -0.791
#> 46  -1.296  0.552      1.848
#> 47  -0.431 -0.312      0.119
#> 48   0.390  2.314      1.924
#> 49   1.158 -0.200     -1.358
#> 50  -0.615  2.662      3.277
#> 51   0.210  0.658      0.448
#> 52  -0.948 -0.673      0.275
#> 53   0.199 -1.173     -1.372
#> 54  -0.168  1.837      2.005
#> 55   0.096 -0.994     -1.090
#> 56  -0.221 -0.020      0.201
#> 57   0.154 -1.128     -1.282
#> 58  -0.167 -1.641     -1.474
#> 59  -0.523  1.086      1.609
#> 60   1.037  1.161      0.124
#> 61  -0.584 -0.808     -0.224
#> 62   0.616  3.011      2.395
#> 63  -0.456  0.161      0.617
#> 64  -1.630  0.336      1.966
#> 65   0.593  0.541     -0.052
#> 66  -0.564 -0.406      0.158
#> 67  -1.609 -0.866      0.743
#> 68   0.212  1.175      0.963
#> 69  -0.065 -0.453     -0.388
#> 70   0.651  1.233      0.582
#> 71   1.120  1.892      0.772
#> 72   0.308 -0.856     -1.164
#> 73   0.656 -0.524     -1.180
#> 74   0.416 -0.818     -1.234
#> 75  -2.262 -0.437      1.825
#> 76   0.933  0.593     -0.340
#> 77   0.420 -0.240     -0.660
#> 78   0.259  1.527      1.268
#> 79  -2.559 -0.944      1.615
#> 80   0.989 -0.237     -1.226
#> 81  -0.620 -0.058      0.562
#> 82   0.166  0.041     -0.125
#> 83  -0.686 -0.233      0.453
#> 84  -0.782 -1.322     -0.540
#> 85  -0.405  0.871      1.276
#> 86   0.550  0.682      0.132
#> 87   1.349  1.083     -0.266
#> 88   0.816  0.143     -0.673
#> 89   0.092  1.607      1.515
#> 90  -0.179  1.103      1.282
#> 91  -0.333 -0.376     -0.043
#> 92  -2.687 -0.716      1.971
#> 93   0.153  0.058     -0.095
#> 94  -1.926 -1.628      0.298
#> 95  -0.616  0.568      1.184
#> 96  -1.640 -1.740     -0.100
#> 97   1.056  0.946     -0.110
#> 98  -1.550 -1.547      0.003
#> 99  -1.279  0.144      1.423
#> 100  0.556  0.186     -0.370
#> 

generate_matched(skew = c(0.4, 0.25))
#> $correlation
#> [1] 0.49
#> 
#> $summary
#>             mean   sd  skew kurtosis
#> x input     0.00 1.00  0.40     0.00
#> sample x   -0.05 0.93 -0.12    -0.41
#> y input     0.50 1.00  0.25     0.00
#> sample y    0.51 1.09  0.45    -0.22
#> difference  0.56 1.03  0.23    -0.51
#> 
#> $sample
#>         X1     X2 Difference
#> 1    0.061  1.461      1.400
#> 2    2.099  1.432     -0.667
#> 3    1.175  0.087     -1.088
#> 4   -0.340  0.562      0.902
#> 5   -2.023 -1.285      0.738
#> 6   -0.226  1.147      1.373
#> 7    1.089  1.423      0.334
#> 8   -0.792 -0.556      0.236
#> 9    0.682  1.409      0.727
#> 10  -0.075  0.334      0.409
#> 11  -0.912  1.503      2.415
#> 12   0.299 -0.586     -0.885
#> 13   1.283  2.261      0.978
#> 14  -0.488  0.317      0.805
#> 15   0.812  2.077      1.265
#> 16  -0.072 -1.240     -1.168
#> 17   0.797  1.740      0.943
#> 18   0.446 -0.044     -0.490
#> 19   0.186  0.105     -0.081
#> 20  -1.615 -0.733      0.882
#> 21  -1.904 -0.131      1.773
#> 22   0.423  0.566      0.143
#> 23   0.037  3.137      3.100
#> 24  -0.719 -1.151     -0.432
#> 25  -0.144 -0.953     -0.809
#> 26   0.335  0.735      0.400
#> 27  -0.057  2.083      2.140
#> 28  -1.161 -0.342      0.819
#> 29  -1.057 -1.454     -0.397
#> 30  -0.274  1.573      1.847
#> 31   1.107  3.635      2.528
#> 32   1.063  3.008      1.945
#> 33   1.297  1.877      0.580
#> 34  -0.492  0.112      0.604
#> 35   0.417 -0.218     -0.635
#> 36   0.921  0.790     -0.131
#> 37   0.154  0.370      0.216
#> 38   1.426  0.299     -1.127
#> 39   0.897  2.664      1.767
#> 40  -0.526  0.344      0.870
#> 41  -0.686 -0.088      0.598
#> 42  -0.768 -0.349      0.419
#> 43   1.171  1.564      0.393
#> 44  -1.624  0.944      2.568
#> 45   1.231  0.707     -0.524
#> 46  -0.454  0.071      0.525
#> 47  -0.646 -0.932     -0.286
#> 48  -0.290  2.663      2.953
#> 49   0.035  0.466      0.431
#> 50   1.394  0.951     -0.443
#> 51   0.545  2.309      1.764
#> 52  -0.710 -0.628      0.082
#> 53  -0.306  0.230      0.536
#> 54  -0.609 -1.326     -0.717
#> 55  -0.220 -0.420     -0.200
#> 56  -0.697  1.221      1.918
#> 57   1.102 -0.232     -1.334
#> 58   2.116  1.806     -0.310
#> 59   0.790  1.493      0.703
#> 60   1.624  0.599     -1.025
#> 61  -1.058  0.701      1.759
#> 62   0.479  0.842      0.363
#> 63  -1.970 -0.729      1.241
#> 64   0.318  0.210     -0.108
#> 65  -0.469  0.119      0.588
#> 66  -1.336 -0.056      1.280
#> 67  -1.028 -0.331      0.697
#> 68  -1.101 -0.288      0.813
#> 69  -0.698  0.426      1.124
#> 70  -0.761 -0.418      0.343
#> 71  -1.866 -0.157      1.709
#> 72   0.645  0.569     -0.076
#> 73   0.562  0.266     -0.296
#> 74   0.066  1.658      1.592
#> 75   0.313  2.018      1.705
#> 76   0.422  1.486      1.064
#> 77  -0.309  0.066      0.375
#> 78   0.618 -0.949     -1.567
#> 79  -0.093 -1.353     -1.260
#> 80   0.015  1.065      1.050
#> 81  -0.017  0.094      0.111
#> 82   0.416  1.793      1.377
#> 83  -0.393  0.288      0.681
#> 84  -0.223  1.218      1.441
#> 85  -0.085 -0.813     -0.728
#> 86  -2.087  0.351      2.438
#> 87   0.577  0.853      0.276
#> 88   0.716  0.052     -0.664
#> 89  -0.081 -0.255     -0.174
#> 90   0.826  0.941      0.115
#> 91   0.490  0.159     -0.331
#> 92  -1.854 -0.568      1.286
#> 93   0.715 -0.047     -0.762
#> 94  -0.817 -0.915     -0.098
#> 95  -1.169  1.091      2.260
#> 96  -0.304  1.702      2.006
#> 97  -1.361 -0.310      1.051
#> 98   0.547  0.912      0.365
#> 99   0.072  0.675      0.603
#> 100 -1.086 -1.099     -0.013
#>