Generate samples of quantitative matched pair data with correlation and optional standardized effect

generate_matched() is a function to generate samples of quantitative matched pair data with specified first four moments, correlation, and optional standardized effect size

generate_matched(
  n = 100,
  mean = 0,
  var = 1,
  skew = 0,
  kurt = 0,
  effect_size = 0.5,
  r = 0.5,
  use_effect_size = TRUE,
  summary = TRUE,
  decimals = 3,
  replication = 1
)

Arguments

n

A numeric vector for the sample size. The default value is 100.

mean

A numeric vector for the population mean of the first variable. The default value is 0.

var

A numeric vector for the population variance of the first variable. The default value is 1.

skew

A numeric vector for the population skew of the variables. Separate values of skew can be entered. The default value is 0.

kurt

A numeric vector for the population kurtosis of the variables. Separate values of kurtosis can be entered. The default value is 0.

effect_size

An optional numeric vector for the standardized effect size between groups of data.

r

A numeric vector for the correlation between the variables. The default value is 0.5.

use_effect_size

A logical vector to indicate whether to use a standardized effect size in generating data for k when k >= 2. The default value is FALSE.

summary

A logical vector for whether the return should include summary statistics. The default value is TRUE.

decimals

A numeric vector for the number of decimals to round the sample data to. The default value is 3.

replication

A numeric vector for the number of times to replicate the sampling. The default value is 1.

@return If summary = TRUE, a list containing a vector ("correlation") with the correlation between the pairs, a matrix ("summary") of the input parameters and descriptive statistics, and a data frame ("sample") with n rows and 3 columns (values from each set of the pair and the difference of X2 and X1). If summary = FALSE, a data frame with n rows and 3 columns.

Examples

generate_matched()
#> $correlation
#> [1] 0.371
#> 
#> $summary
#>            mean   sd  skew kurtosis
#> x input    0.00 1.00  0.00     0.00
#> sample x   0.12 0.98 -0.02    -0.13
#> y input    0.50 1.00  0.00     0.00
#> sample y   0.49 0.92  0.01     0.30
#> difference 0.36 1.06  0.15    -0.16
#> 
#> $sample
#>         X1     X2 Difference
#> 1    0.952  0.866     -0.086
#> 2   -1.566  0.373      1.939
#> 3   -0.724  2.255      2.979
#> 4    0.808 -0.800     -1.608
#> 5   -2.179  0.666      2.845
#> 6    0.588  1.542      0.954
#> 7   -1.217 -0.176      1.041
#> 8    1.361  1.126     -0.235
#> 9    0.445  0.513      0.068
#> 10   1.943  0.648     -1.295
#> 11   0.670  1.082      0.412
#> 12   0.504  0.274     -0.230
#> 13   0.455  1.291      0.836
#> 14   0.892  1.062      0.170
#> 15   1.495  1.621      0.126
#> 16   0.208  0.065     -0.143
#> 17   1.358  0.332     -1.026
#> 18   1.242  0.599     -0.643
#> 19  -0.466 -1.195     -0.729
#> 20   0.122  0.250      0.128
#> 21  -1.064 -1.394     -0.330
#> 22   0.363  0.460      0.097
#> 23   0.496  0.148     -0.348
#> 24   0.733  0.063     -0.670
#> 25  -0.281 -0.337     -0.056
#> 26  -0.318  0.166      0.484
#> 27   0.463  0.233     -0.230
#> 28   2.343  0.591     -1.752
#> 29  -1.716  0.092      1.808
#> 30   0.526  2.870      2.344
#> 31   1.041  0.717     -0.324
#> 32   2.238  0.103     -2.135
#> 33   1.181  0.378     -0.803
#> 34  -0.225 -0.040      0.185
#> 35  -0.585  0.160      0.745
#> 36  -0.187 -0.817     -0.630
#> 37   1.451 -0.593     -2.044
#> 38  -0.456  0.598      1.054
#> 39  -0.899 -0.321      0.578
#> 40  -0.779  0.986      1.765
#> 41   1.656  1.641     -0.015
#> 42  -0.582 -0.944     -0.362
#> 43  -0.286  1.231      1.517
#> 44  -0.303  0.772      1.075
#> 45   0.671  0.159     -0.512
#> 46  -0.051  0.234      0.285
#> 47   1.393  0.905     -0.488
#> 48  -0.350 -0.736     -0.386
#> 49   0.125  1.352      1.227
#> 50   0.308  0.501      0.193
#> 51   0.654  0.710      0.056
#> 52   2.029  2.195      0.166
#> 53  -0.797  0.951      1.748
#> 54  -2.604 -0.684      1.920
#> 55   0.869  2.434      1.565
#> 56  -0.919 -0.470      0.449
#> 57   0.436  0.134     -0.302
#> 58  -0.268 -1.941     -1.673
#> 59   0.397  0.634      0.237
#> 60  -1.151 -2.012     -0.861
#> 61  -0.561 -0.446      0.115
#> 62  -0.299  1.596      1.895
#> 63  -0.652  0.583      1.235
#> 64   0.137  0.028     -0.109
#> 65   1.308  0.794     -0.514
#> 66  -0.194  1.382      1.576
#> 67  -0.044  0.058      0.102
#> 68   1.149  1.433      0.284
#> 69  -0.118 -0.764     -0.646
#> 70   2.308  1.034     -1.274
#> 71  -0.853 -0.096      0.757
#> 72   0.210  1.353      1.143
#> 73  -0.850  0.021      0.871
#> 74   0.034  0.365      0.331
#> 75  -0.797  0.509      1.306
#> 76  -0.797  0.482      1.279
#> 77   0.802  1.883      1.081
#> 78   1.107  0.737     -0.370
#> 79   0.667  0.156     -0.511
#> 80   1.300  1.754      0.454
#> 81  -0.166  0.665      0.831
#> 82   1.108  1.737      0.629
#> 83  -0.707 -0.481      0.226
#> 84  -0.737  0.321      1.058
#> 85  -1.549 -0.010      1.539
#> 86  -0.978 -0.553      0.425
#> 87   0.223  0.668      0.445
#> 88   0.231  2.686      2.455
#> 89  -0.349  1.590      1.939
#> 90   0.500  0.920      0.420
#> 91   0.431 -0.438     -0.869
#> 92   0.773  1.949      1.176
#> 93  -0.812  0.133      0.945
#> 94   0.311 -0.362     -0.673
#> 95  -0.342  1.001      1.343
#> 96  -1.499  1.228      2.727
#> 97  -0.640  1.012      1.652
#> 98   0.915  0.743     -0.172
#> 99   0.069 -0.026     -0.095
#> 100 -0.663 -0.368      0.295
#> 

generate_matched(use_effect_size = FALSE)
#> $correlation
#> [1] 0.509
#> 
#> $summary
#>             mean   sd  skew kurtosis
#> x input     0.00 1.00  0.00     0.00
#> sample x    0.07 1.13 -0.22    -0.21
#> y input     0.00 1.00  0.00     0.00
#> sample y    0.05 0.98  0.02     0.28
#> difference -0.02 1.05  0.52     0.52
#> 
#> $sample
#>         X1     X2 Difference
#> 1    1.232  0.974     -0.258
#> 2   -2.369  0.383      2.752
#> 3   -1.918 -0.216      1.702
#> 4   -1.320 -0.455      0.865
#> 5   -0.326  2.745      3.071
#> 6    0.297  0.192     -0.105
#> 7    0.421 -0.698     -1.119
#> 8    0.867 -0.716     -1.583
#> 9   -0.314 -1.881     -1.567
#> 10   1.047  0.364     -0.683
#> 11   0.194  0.401      0.207
#> 12   0.077 -1.021     -1.098
#> 13  -0.637 -1.129     -0.492
#> 14   1.332  1.765      0.433
#> 15   0.240 -0.803     -1.043
#> 16   1.431  1.446      0.015
#> 17  -0.685 -0.847     -0.162
#> 18   0.881  1.884      1.003
#> 19   0.856  0.774     -0.082
#> 20   0.969  0.274     -0.695
#> 21   1.190  1.270      0.080
#> 22   0.052  0.740      0.688
#> 23  -1.631 -0.364      1.267
#> 24   1.857  1.007     -0.850
#> 25  -0.667 -0.162      0.505
#> 26  -1.204  0.649      1.853
#> 27   0.030 -2.741     -2.771
#> 28   0.632 -0.604     -1.236
#> 29  -1.848 -1.505      0.343
#> 30   0.972 -0.023     -0.995
#> 31   1.067  0.872     -0.195
#> 32   0.420  0.396     -0.024
#> 33   0.465  1.622      1.157
#> 34  -1.421 -0.077      1.344
#> 35  -1.497  0.185      1.682
#> 36   0.744  1.184      0.440
#> 37   1.297  0.674     -0.623
#> 38   0.421  0.626      0.205
#> 39   0.304  0.086     -0.218
#> 40  -0.314  0.090      0.404
#> 41  -0.289 -1.947     -1.658
#> 42  -1.236 -0.376      0.860
#> 43   0.688  0.172     -0.516
#> 44   0.228 -0.391     -0.619
#> 45  -2.735 -0.316      2.419
#> 46   1.025 -0.783     -1.808
#> 47  -0.171  0.892      1.063
#> 48   2.709  2.468     -0.241
#> 49  -0.627 -0.231      0.396
#> 50   0.392 -0.760     -1.152
#> 51   0.061 -0.971     -1.032
#> 52  -1.149  0.603      1.752
#> 53  -1.198 -1.601     -0.403
#> 54  -0.764 -1.141     -0.377
#> 55   0.501 -1.173     -1.674
#> 56   0.368  0.628      0.260
#> 57  -0.523 -0.096      0.427
#> 58   1.297 -0.854     -2.151
#> 59   0.208  0.107     -0.101
#> 60  -0.201 -0.071      0.130
#> 61   0.663  0.755      0.092
#> 62  -1.525  0.539      2.064
#> 63   0.820  0.708     -0.112
#> 64   2.280  1.480     -0.800
#> 65   0.404 -0.269     -0.673
#> 66   0.278  0.376      0.098
#> 67   0.324  0.375      0.051
#> 68   0.253 -0.488     -0.741
#> 69   1.237  0.649     -0.588
#> 70   0.422 -0.306     -0.728
#> 71   0.863  0.069     -0.794
#> 72   0.492 -0.140     -0.632
#> 73  -2.252 -2.283     -0.031
#> 74  -1.038 -0.721      0.317
#> 75  -0.248 -0.216      0.032
#> 76  -0.079 -0.821     -0.742
#> 77   0.714  0.222     -0.492
#> 78  -1.750 -1.181      0.569
#> 79   0.238  0.110     -0.128
#> 80  -0.554 -0.509      0.045
#> 81   2.840  1.575     -1.265
#> 82  -1.273  0.534      1.807
#> 83  -0.400 -0.339      0.061
#> 84  -1.071 -0.344      0.727
#> 85  -1.324 -1.226      0.098
#> 86  -1.196 -1.135      0.061
#> 87   0.319  1.057      0.738
#> 88   1.115  0.673     -0.442
#> 89   1.402  1.778      0.376
#> 90   0.777  0.162     -0.615
#> 91   2.018  1.554     -0.464
#> 92  -1.392  0.667      2.059
#> 93   1.719  1.426     -0.293
#> 94  -0.467 -0.737     -0.270
#> 95  -2.281 -0.356      1.925
#> 96   0.183 -0.250     -0.433
#> 97   0.426  0.011     -0.415
#> 98   1.408  0.575     -0.833
#> 99   0.493 -0.008     -0.501
#> 100  0.649 -0.400     -1.049
#> 

generate_matched(skew = c(0.4, 0.25))#'
#> $correlation
#> [1] 0.495
#> 
#> $summary
#>            mean   sd  skew kurtosis
#> x input    0.00 1.00  0.40     0.00
#> sample x   0.01 0.89  0.35     0.44
#> y input    0.50 1.00  0.25     0.00
#> sample y   0.36 0.93  0.64     0.46
#> difference 0.35 0.92 -0.09     0.83
#> 
#> $sample
#>         X1     X2 Difference
#> 1   -0.219 -0.462     -0.243
#> 2    0.296  0.865      0.569
#> 3   -1.132 -1.024      0.108
#> 4   -0.149 -0.733     -0.584
#> 5    0.568 -0.056     -0.624
#> 6    1.414  2.927      1.513
#> 7   -0.495 -0.146      0.349
#> 8    0.012  0.043      0.031
#> 9   -0.065  1.492      1.557
#> 10   3.013  0.939     -2.074
#> 11  -0.201  0.657      0.858
#> 12   0.161  0.432      0.271
#> 13  -0.025  0.462      0.487
#> 14  -0.851 -0.016      0.835
#> 15   0.134  1.673      1.539
#> 16  -0.410 -1.098     -0.688
#> 17  -1.126  0.436      1.562
#> 18   1.408  2.648      1.240
#> 19   0.089  1.370      1.281
#> 20   0.569  0.404     -0.165
#> 21  -0.006  0.685      0.691
#> 22  -0.021 -0.412     -0.391
#> 23   1.538  1.015     -0.523
#> 24   0.871  0.446     -0.425
#> 25   0.571  0.886      0.315
#> 26  -0.595  0.384      0.979
#> 27   0.607 -0.267     -0.874
#> 28  -0.906  0.552      1.458
#> 29   0.192  0.307      0.115
#> 30   0.049  1.496      1.447
#> 31   1.719  2.819      1.100
#> 32  -1.058  0.016      1.074
#> 33  -0.013 -0.503     -0.490
#> 34  -1.331  0.117      1.448
#> 35  -0.314 -0.444     -0.130
#> 36   0.353 -0.047     -0.400
#> 37   0.329  0.525      0.196
#> 38  -0.209  0.018      0.227
#> 39  -0.581 -0.356      0.225
#> 40   0.506  0.204     -0.302
#> 41  -0.116  0.166      0.282
#> 42   1.193  0.116     -1.077
#> 43  -0.635 -0.349      0.286
#> 44   0.402  0.087     -0.315
#> 45  -1.506 -1.073      0.433
#> 46  -0.395 -0.792     -0.397
#> 47  -0.283  1.063      1.346
#> 48   0.492  0.398     -0.094
#> 49  -1.526 -0.382      1.144
#> 50   0.886  2.014      1.128
#> 51   0.074 -1.333     -1.407
#> 52  -0.755  0.857      1.612
#> 53  -0.295 -0.386     -0.091
#> 54  -0.905 -0.597      0.308
#> 55  -0.308  0.894      1.202
#> 56  -0.027  0.295      0.322
#> 57   1.141 -0.325     -1.466
#> 58  -0.165 -0.347     -0.182
#> 59  -0.687 -0.232      0.455
#> 60  -0.740 -1.371     -0.631
#> 61   1.191 -0.864     -2.055
#> 62   0.190 -0.188     -0.378
#> 63  -0.236  3.134      3.370
#> 64   1.554  0.697     -0.857
#> 65   0.766  1.115      0.349
#> 66  -0.468  1.818      2.286
#> 67  -0.145  0.276      0.421
#> 68  -0.139  0.499      0.638
#> 69  -0.936  0.156      1.092
#> 70   0.967  1.182      0.215
#> 71   0.557  0.818      0.261
#> 72  -0.111  1.694      1.805
#> 73  -0.547  0.985      1.532
#> 74   0.686  0.833      0.147
#> 75   0.785  0.162     -0.623
#> 76  -1.393 -0.729      0.664
#> 77  -0.126  1.391      1.517
#> 78  -0.049  0.006      0.055
#> 79   1.848  1.563     -0.285
#> 80  -0.634  0.032      0.666
#> 81   1.581  1.442     -0.139
#> 82  -0.035  0.753      0.788
#> 83   0.217  0.302      0.085
#> 84   1.681  1.794      0.113
#> 85   0.979  1.212      0.233
#> 86   0.069 -0.179     -0.248
#> 87  -0.255  0.199      0.454
#> 88  -1.504 -1.447      0.057
#> 89   0.973 -1.107     -2.080
#> 90  -0.835 -0.297      0.538
#> 91   0.134  0.254      0.120
#> 92  -0.084  0.733      0.817
#> 93  -0.229  0.170      0.399
#> 94  -1.897 -0.550      1.347
#> 95  -0.336  0.126      0.462
#> 96  -1.430 -0.313      1.117
#> 97   0.321  0.381      0.060
#> 98  -1.686 -0.442      1.244
#> 99   0.717  1.770      1.053
#> 100 -1.745 -0.305      1.440
#>