Generate data using Headrick's fifth-order polynomial method

headrick_method() uses specified values of mean, variance, skew, kurtosis, and the fifth and sixth moments to determine the constants for the expected values of Y and generate a sample of the specified distribution.

Usage

headrick_method(
  k,
  n,
  mean,
  var,
  skew,
  kurtosis,
  corr = diag(1, k),
  gam3 = NaN,
  gam4 = NaN
)

Arguments

k

A numeric vector indicating the number of variables to sample.

n

A numeric vector for the sample size.

mean

A numeric vector of length k for the means of each variable in the distribution. The values can be identical.

var

A numeric vector of length k for the variances of each variable in the distribution. The values can be identical.

skew

A numeric vector of length k for the skews of each variable in the distribution. The values can be identical.

kurtosis

A numeric vector of length k for the kurtoses of each variable in the distribution. The values can be identical.

corr

A kxk correlation matrix. The default value is diag(1,k).

gam3

An optional numeric value for the 5th standardized cumulant. The default value is computed internally.

gam4

An optional numeric value for the 6th standardized cumulant. The default value is computed internally.

Value

A data frame of n rows and k columns containing the generated values.

References

Headrick, T. C. (2002). Fast fifth-order polynomial transforms for generating univariate and multivariate nonnormal distributions. Computational Statistics & Data Analysis, 40, 685-711.

Olvera Astivia, O. L., & Zumbo, B. D. (2015). A Cautionary Note on the Use of the Vale and Maurelli Method to Generate Multivariate, Nonnormal Data for Simulation Purposes. Educational and Psychological Measurement, 75, 541-567.

Author

Oscar L. Olvera Astivia and Keah Tandon

Examples

headrick_method(1, 100, 0, 1, 0, 0)
#>              X1
#> 1   -0.59667051
#> 2   -1.72255181
#> 3    0.88174948
#> 4    0.37994596
#> 5   -0.99611001
#> 6    0.45912089
#> 7   -1.46026574
#> 8   -1.04324618
#> 9    1.65904097
#> 10  -0.42395777
#> 11   1.41916293
#> 12   1.46469246
#> 13  -0.07042097
#> 14  -0.05669822
#> 15   0.03270818
#> 16  -1.26428122
#> 17  -0.29498788
#> 18   0.67308682
#> 19  -0.11958829
#> 20   0.51439651
#> 21   0.06811158
#> 22   1.14801191
#> 23   0.11524448
#> 24  -0.10293639
#> 25   0.70174349
#> 26  -0.04123062
#> 27   0.64876835
#> 28   0.64400786
#> 29   0.20443883
#> 30  -0.49084092
#> 31   0.20533699
#> 32  -0.19069132
#> 33  -0.97002419
#> 34   0.28526108
#> 35   0.10361927
#> 36  -0.51125454
#> 37   1.25816245
#> 38  -0.54382597
#> 39   0.53496359
#> 40   1.13804603
#> 41   1.76800377
#> 42   0.71497820
#> 43   1.17020227
#> 44   1.17844815
#> 45  -1.62729134
#> 46   1.21214667
#> 47  -0.99750992
#> 48  -0.54524564
#> 49   0.88478912
#> 50   0.61123070
#> 51   0.41019951
#> 52  -0.19040937
#> 53   0.34832493
#> 54  -0.82684753
#> 55   0.29268690
#> 56   2.16534621
#> 57   0.09621678
#> 58   1.85195685
#> 59  -0.93781113
#> 60   0.74768648
#> 61   1.11524144
#> 62   0.59150642
#> 63  -0.58523068
#> 64   1.91855820
#> 65  -0.25381679
#> 66   0.97224177
#> 67   0.00851345
#> 68  -0.67904191
#> 69  -0.56963137
#> 70  -1.50296166
#> 71  -0.12908540
#> 72   0.70118295
#> 73  -2.15370721
#> 74   0.30324614
#> 75  -0.80757978
#> 76  -0.93099146
#> 77  -0.45490210
#> 78  -1.42542069
#> 79   0.20899117
#> 80   1.23140278
#> 81   0.36060358
#> 82   0.54721330
#> 83  -0.32799601
#> 84   0.93329983
#> 85  -0.68081082
#> 86   1.16388240
#> 87  -1.75544951
#> 88  -1.44963736
#> 89   0.44879494
#> 90  -0.14006729
#> 91   0.37365033
#> 92   1.20643412
#> 93   2.66134598
#> 94   0.02509449
#> 95   0.03437326
#> 96  -1.64235904
#> 97   1.99546215
#> 98   0.66217076
#> 99  -0.88003799
#> 100 -0.34238742

headrick_method(2, 100, c(0,0), c(1,1), c(0,0), c(0,0))
#>              X1           X2
#> 1    0.40554631 -2.417662723
#> 2   -0.38655768  0.906034938
#> 3    0.25495494  1.439136631
#> 4    1.38918987 -0.040003266
#> 5   -0.03459414 -1.815736058
#> 6    1.35888329 -0.049885145
#> 7   -0.87461792 -0.103806989
#> 8    1.88385977 -0.839949452
#> 9    0.13426958  0.093625286
#> 10  -0.43683058 -0.098097392
#> 11   1.54162879 -1.202131490
#> 12  -1.40897273 -0.088676754
#> 13   0.86263289  0.727301449
#> 14   0.48384681 -0.375865317
#> 15  -0.15461715  1.077472643
#> 16   0.32031672  1.474021047
#> 17   1.01893945  0.116107570
#> 18  -0.38976750  1.087428386
#> 19   0.50131138 -0.556897598
#> 20  -0.48401264  0.139076866
#> 21   0.48227963  0.741548351
#> 22   0.93683311 -1.655869823
#> 23   0.08951769  0.104010130
#> 24  -0.62459650 -1.433809901
#> 25  -0.22970172  0.107306429
#> 26   0.05573467  0.312778525
#> 27   2.01284879 -0.355198175
#> 28  -1.46648609 -1.012580453
#> 29   0.03776841 -2.702104352
#> 30   0.84076289  1.017884388
#> 31   0.19104142  0.173693651
#> 32  -0.24774825 -1.062947518
#> 33   1.54023073 -0.324610371
#> 34  -0.69226706  0.177410538
#> 35  -0.23399935  1.304043130
#> 36  -1.19599355  0.264354878
#> 37  -0.96675883  0.174951020
#> 38   0.51472454  0.077348519
#> 39   0.46821378 -1.644071506
#> 40   0.97663747  0.440670345
#> 41   0.34389307  1.604012555
#> 42  -0.92346925 -0.474201923
#> 43  -1.84302317  0.105916780
#> 44  -0.90697972  0.193566123
#> 45   1.88449392 -0.610842125
#> 46  -0.79780098  2.801667459
#> 47  -1.01927230  0.530120712
#> 48  -0.59288470  0.145244835
#> 49   0.83149179 -0.580905998
#> 50  -1.04896469  0.247510817
#> 51  -0.70447144  0.986681579
#> 52  -0.63772615 -0.354337366
#> 53   0.36284496 -0.018388058
#> 54  -0.82226236 -0.104567244
#> 55  -0.10557403 -1.146565198
#> 56   0.06409420  0.518914002
#> 57  -1.50574871  1.245963220
#> 58  -0.73236370  1.279271347
#> 59   2.78922714  1.006340153
#> 60   0.40638673  1.396483918
#> 61   1.50608846  0.046546913
#> 62  -0.56464217 -1.863653343
#> 63   0.42192314  0.094133564
#> 64  -0.87903890 -0.752554476
#> 65  -0.44638582  1.089193786
#> 66  -0.02084114  0.210017004
#> 67   0.08050039 -0.542323744
#> 68   0.34347429  0.337414465
#> 69  -0.18896600 -0.760667263
#> 70   1.09518822 -0.669003038
#> 71  -0.64781226  0.608030856
#> 72   0.37508406  0.747223677
#> 73   0.71902239 -2.428612469
#> 74   0.30974321  0.182355443
#> 75  -1.93345219 -0.006259069
#> 76   1.80531733  0.051583731
#> 77   0.33155886  0.544397822
#> 78  -0.10083648  0.905499247
#> 79   0.51444875  1.124369273
#> 80   0.39163163 -0.892073058
#> 81  -0.36375099  0.966783373
#> 82  -0.21084412 -0.642528784
#> 83   0.64876855 -1.417236168
#> 84  -0.54363766 -1.147854518
#> 85  -0.16994663  0.131308451
#> 86  -0.66131296  0.516076386
#> 87   1.33057062  1.064132005
#> 88   0.74314414  0.017552100
#> 89   0.18250096  0.220458483
#> 90  -0.93870139 -0.147940073
#> 91   0.50601029  0.579642262
#> 92   1.01572482 -0.604372791
#> 93  -0.74700849 -0.689123236
#> 94  -0.44398633 -1.060958080
#> 95   1.22060643 -2.325846705
#> 96  -2.11848384 -0.187430399
#> 97  -1.51851446  0.551609950
#> 98   1.39790523  0.073131050
#> 99  -2.22937628  1.661430930
#> 100  1.77342103 -0.122939822

corr <- matrix(c(1, .7, .7, 1), byrow = TRUE, nrow = 2)
headrick_method(2, 100, c(0,0), c(1,1), c(0,0), c(0,0), corr = corr)
#>               X1           X2
#> 1   -0.712337133  0.658454587
#> 2   -0.681263552 -0.387238351
#> 3    1.008574134  0.784961495
#> 4    0.342744280  0.027729557
#> 5   -0.336621413 -0.266076331
#> 6    0.377605865 -0.219737891
#> 7   -2.126480745 -0.978022069
#> 8    1.984449033  1.740469097
#> 9   -0.470996734 -0.879380570
#> 10   0.307774229 -0.333281747
#> 11  -1.229472370 -0.488005702
#> 12   0.560619288 -0.383111370
#> 13   2.703523398  2.417503669
#> 14  -0.485885983 -0.253123900
#> 15   0.230609032  0.151035334
#> 16   0.791408901  0.991149214
#> 17   1.101374299  0.091624226
#> 18   0.108663430 -0.206105220
#> 19  -1.633916671 -2.806323157
#> 20   0.143744314 -0.476684751
#> 21   0.438668365  1.352409381
#> 22  -1.910246664 -1.694777589
#> 23   0.353309924  0.416901754
#> 24   1.114074794  0.385092273
#> 25   1.372281279  0.913971237
#> 26   0.996175748  0.216743340
#> 27  -1.857678217 -1.812113060
#> 28  -1.120193091 -0.517200150
#> 29  -0.770452638 -0.539968008
#> 30   0.097843949 -0.355620135
#> 31  -0.453099485 -0.044933282
#> 32  -0.668356866 -0.374591463
#> 33  -0.501994700  0.119941103
#> 34   0.025779371 -0.297452470
#> 35  -0.403921848 -1.339509029
#> 36   0.175801129  0.113812137
#> 37  -0.334761544 -0.111826380
#> 38   0.213474312  1.443174495
#> 39  -1.447318400 -1.506573172
#> 40  -1.464445667 -1.107167729
#> 41   1.745269672  1.674038514
#> 42  -0.609068293  1.015643137
#> 43  -1.006590430  0.009347863
#> 44  -0.337902981 -0.511853194
#> 45  -0.627647746 -0.726075013
#> 46   0.399502130  0.383323427
#> 47   1.052576126  1.145388182
#> 48  -0.056440043 -0.474186783
#> 49   1.397302685  1.465743449
#> 50  -0.301156773 -0.831440911
#> 51  -0.212964555 -1.056805705
#> 52   1.461570680  2.141659796
#> 53  -1.107362216 -0.587766306
#> 54   0.406400403 -0.011772205
#> 55   0.122665577  1.100553735
#> 56  -0.319942838 -0.463621125
#> 57  -1.056354431 -1.290910617
#> 58  -1.060048612 -1.864878860
#> 59   1.104141416  1.084197490
#> 60   0.435519415  0.333381341
#> 61  -0.152500904 -0.103043152
#> 62  -0.220744737  0.105141338
#> 63   0.056865292  0.735969194
#> 64   2.104710193  2.361685902
#> 65  -0.119977149  0.615542625
#> 66   0.653295531  1.521853774
#> 67   1.088423875 -1.153398646
#> 68   0.890904734  0.607760218
#> 69  -0.105736203  0.312234301
#> 70   1.104017131  0.238220107
#> 71   0.722926132  0.005109678
#> 72   0.230010334  0.433168033
#> 73  -1.038531512 -1.405359553
#> 74   0.007860462 -1.333417556
#> 75  -0.128512306  0.540554334
#> 76   1.471385430  0.781723197
#> 77  -1.102829777 -0.792709255
#> 78   1.156044817  1.502555180
#> 79   0.480359186  1.717405605
#> 80   0.615185139  0.767350352
#> 81   0.614810570  1.143817573
#> 82   0.373725722  1.272997952
#> 83  -1.012882236 -1.349073811
#> 84   0.712717838 -0.518518960
#> 85   0.580648400  0.690449412
#> 86  -1.120304502 -0.101387893
#> 87   0.343356197 -0.700684884
#> 88   0.709210377  1.538115124
#> 89   0.517352276  0.471981267
#> 90  -0.507655809 -0.432804164
#> 91  -1.701231254 -1.997686415
#> 92  -1.917126144 -1.042517018
#> 93  -1.325977138 -0.210710109
#> 94  -0.570724713  0.594520360
#> 95   2.031998343  1.212252177
#> 96   0.362136133  0.947462712
#> 97   0.326942345  0.050414581
#> 98  -0.068200135 -0.808248187
#> 99  -0.194611628  0.090703352
#> 100 -0.044300101 -0.057423206